% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_AIC_nsamax.R, R/rfa_AIC_nsamax_mle.R,
%   R/rfa_logLik_nsamax.R
\name{AIC.nsamax}
\alias{AIC.nsamax}
\alias{AIC.nsamax.mle}
\alias{logLik.nsamax}
\title{AIC for nonstationary AMAX model}
\usage{
\method{AIC}{nsamax}(object, ..., k = 2)

\method{AIC}{nsamax.mle}(object, ..., k = 2)

\method{logLik}{nsamax}(object, ...)
}
\arguments{
\item{object}{Output}

\item{...}{Other parameters}

\item{k}{Penalty per parameter.}
}
\description{
Return the AIC of the nonstationary AMAX model.
}
\seealso{
\link{AIC}.
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
