% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_amax.R
\name{Amax}
\alias{Amax}
\alias{dAmax}
\alias{pAmax}
\alias{qAmax}
\alias{rAmax}
\alias{lAmax}
\title{Basic probability functions for distribution of annual maximums}
\usage{
dAmax(x, para, distr, log = FALSE)

pAmax(q, para, distr)

qAmax(p, para, distr)

rAmax(n, para, distr)

lAmax(l, distr, ...)
}
\arguments{
\item{x, q}{Vector of quantiles.}

\item{para}{Vector of parameters for the given distribution.}

\item{distr}{Distribution family. See \link{lmom2par}.}

\item{log}{Logical. If TRUE, probabilities \code{p} are given as
\code{log(p)}.}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations.}

\item{l}{L-moments.}

\item{...}{Other parameters.}
}
\description{
Density, distribution function, quantile function and random generation
for various distribution used in the modeling annual maximums.
}
\examples{

u <- runif(5)
u
x <- qAmax(u, c(100,3,.001), 'gno')
pAmax(x, c(100,3,.001), 'gno')

x <- rAmax(100, c(100,30,0), 'gev')
sum(dAmax(x, c(100,30, 0), 'gev', log = TRUE))

lAmax(c(100, .3, .2), 'gev', lscale = FALSE)


}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
