% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_data-manipulations-functions.R
\name{Basic_data_manipulation-functions}
\alias{Basic_data_manipulation-functions}
\title{Basic data manipulation functions}
\description{
These functions read in or convert values among formats
\describe{
 \item{read_ECDE_flows}{Reads a file of WSC daily flows from ECDataExplorer}
 \item{get_wscstation}{Reads station information from a data file produced by ECDE}
 \item{get_AHCCD_monthly}{Downloads monthly Adjusted and Homogenized Canadian Climate Data (AHCCD) values}
 \item{read_AHCCD_daily}{Reads file of daily AHCCD values}
 \item{read_AHCCD_monthly}{Reads file of monthly AHCCD values}
}
}
