% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_data_manipulations_functions.R
\name{Basic_data_manipulation_functions}
\alias{Basic_data_manipulation_functions}
\title{Basic data manipulation functions}
\description{
These functions read in or convert values among formats
\describe{
 \item{ch_read_ECDE_flows}{Reads a file of WSC daily flows from ECDataExplorer}
 \item{ch_get_ECDE_metadata}{Reads station meta data from ECDataExplorer}
 \item{ch_get_wscstation}{Reads station information from a data file produced by ECDE}
 \item{ch_get_AHCCD_monthly}{Downloads monthly Adjusted and Homogenized Canadian Climate Data (AHCCD) values}
 \item{ch_read_AHCCD_daily}{Reads file of daily AHCCD values}
 \item{ch_read_AHCCD_monthly}{Reads file of monthly AHCCD values}
 \item{ch_tidyhydat_ECDE}{Reads flows using \pkg{tidyhydat} and converts to ECDE format}
 \item{ch_tidyhydat_ECDE_meta}{Reads station meta data using \pkg{tidyhydat} and converts to ECDE-like format}
}
}
