% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booth_plot.R
\name{booth_plot}
\alias{booth_plot}
\title{Booth Plot}
\usage{
booth_plot(events, threshold, title, type = "mag")
}
\arguments{
\item{events}{: a dataframe from get_peaks}

\item{threshold}{: the threshold used in get_peaks}

\item{title}{: uses format of "05BB001 - Bow River at Banff AB*" obtained from get_WSCstation}

\item{type}{either 'mag' [magnitude default] or 'vol' [volume]}
}
\description{
A Booth plot is a plot of peaks over threhold flood events with duration on the horizontal and 
either magnitude (default) or volume on the vertical axis
}
\examples{
\dontrun{test <-booth_plot(events, threshold, title, type='mag')}
\dontrun{test <-booth_plot(events, threshold, title, type='vol')}

}
\references{
Booth, E.G., Mount, J.F., Viers, J.H. 2006. Hydrologic Variability of the Cosumnes River Floodplain. #' San Francisco Estuary & Watershed Science 4:21.

Whitfield, P.H., and J.W. Pomeroy. 2016. Changes to flood peaks of a mountain river: implications #' for analysis of the 2013 flood in the Upper Bow River, Canada. Hydrological Processes 30:4657-73. doi: #' 10.1002/hyp.10957.
}
