% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_BootNsAmax.R
\name{BootNsAmax}
\alias{BootNsAmax}
\title{Bootstrap sample of the nonstationary AMAX model}
\usage{
BootNsAmax(
  object,
  p = c(0.5, 0.8, 0.9, 0.95, 0.98, 0.99),
  newdata = NULL,
  reliability = FALSE,
  nsim = 1000,
  alpha = 0.95,
  verbose = TRUE
)
}
\arguments{
\item{object}{Output form \link{FitNsAmax}.}

\item{p}{Exceeding probability associated to a quantile or reliability level.}

\item{newdata}{Covariates. If \code{NULL} the input data that served to fit 
the model is used.}

\item{reliability}{Logical. Should the flood quantile associated with 
reliability level be returned. In that case, \code{newdata} must represent
the period of interest. See Durocher et al. (2019) for more details.}

\item{nsim}{Number of simulations.}

\item{alpha}{Confidence level.}

\item{verbose}{Logical. Should a progress bar be displayed.}

\item{...}{Other parameters.}
}
\value{
\item{para}{Parameter of the detrended distribution.}
\item{beta}{Parameter of the trend.}
\item{qua}{flood quantiles.}
}
\description{
Return bootstrap samples of the parameters and flood quantiles for the 
nonstationary AMAX model.
}
\seealso{
\link{FitNsAmax}, \link{predict.nsamax}.
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
