% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_BootNsPot.R
\name{BootNsPot}
\alias{BootNsPot}
\title{Bootstrap sample of the nonstationary POT model}
\usage{
BootNsPot(
  object,
  x,
  rt = c(2, 5, 10, 20, 50, 100),
  newdata = NULL,
  reliability = FALSE,
  nsim = 1000,
  verbose = TRUE
)
}
\arguments{
\item{object}{Output form \link{FitNsPot}.}

\item{x}{Dataset used to fit the nonstationary POT model.}

\item{rt}{Vector of return period to estimate.}

\item{newdata}{Covariates. If \code{NULL} the input data that served to fit 
the model is used}

\item{reliability}{Logical. Should the flood quantile associated with 
reliability level be returned. In that case, \code{newdata} must represent
the period of interest. See Durocher et al. (2019) for more details.}

\item{nsim}{Number of simulations.}

\item{verbose}{Logical. Should a progress bar be displayed.}
}
\description{
Return bootstrap samples of the parameters and flood quantiles for the 
nonstationary POT model.
}
\seealso{
\link{FitNsPot}, \link{predict.nspot}.
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
