% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floodfrequencyanalysis.R
\name{BootstrapCI}
\alias{BootstrapCI}
\title{BootstrapCI: estimate confidence interval for each given 
non-exceedance probability.}
\usage{
BootstrapCI(series, distribution, n.resamples = 1000, nep = nonexceeds(),
  ci = 0.9)
}
\arguments{
\item{series}{A vector representing an extreme value series (e.g., annual maximum flood)}

\item{distribution}{A three-character name of the extreme value distribution (see ?dist.list())}

\item{n.resamples}{An integer representing number of re-samples to conduct}

\item{nep}{A vector of non-exceedance probabilities}

\item{ci}{The confidence interval}

\item{fitted.model}{Fitted distribution (see ?frequencyAnalysis)}
}
\value{
A list containing a data frame of confidence bounds for quantile estimates for each 
non-exceedance probability, a matrix containing estimated distribution parameters for each resample,
and a matrix of quantile estimates for each resample

\url{http://headwateranalytics.weebly.com/blog/flood-frequency-analysis-in-r} contains a demo
}
\description{
BootstrapCI: estimate confidence interval for each given 
non-exceedance probability.
}
