% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_rfa_data.R
\docType{data}
\name{CAN01AD002}
\alias{CAN01AD002}
\title{Streamflow data}
\format{
An object of class \code{data.frame} with 32234 rows and 2 columns.
}
\source{
\url{https://wateroffice.ec.gc.ca/}
}
\usage{
CAN01AD002
}
\description{
Daily river discharge for the station 01AD002 on
St. John River at Fort Kent, New Brunswick.
Data ranges from 1926 to 2014, for basin area of 14700 sq km.
}
\author{
Martin Durocher
}
\keyword{datasets}
