% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment.R
\name{catchment}
\alias{catchment}
\title{Delineate a watershed}
\usage{
catchment(dem, lat, long, buffsize, crs, outname, fillsinks = T,
  saga.env = rsaga.env())
}
\arguments{
\item{dem}{Raster object of your dem.}

\item{lat}{A number.  Latitude of outlet in decimal degrees.}

\item{long}{A number. Longitude of outlet in decimal degrees.}

\item{buffsize}{A number, the buffer (m) around catchment outlet to find location on digital stream network.}

\item{outname}{Character.  The name for the output shapefile.}

\item{fillsinks}{Boolean, default is TRUE, using the Fill Sinks XXL method.}

\item{saga.env}{Saga environment object.  Default is to let saga find it on its own.}
}
\value{
A SpatialPolygonsDataFrame and a shapefile in yjr working directory.
}
\description{
Delineate a watershed
}
