% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_ams_plot.R
\name{ch_ams_plot}
\alias{ch_ams_plot}
\title{Plot an annual maximum series}
\usage{
ch_ams_plot(date, flow, data = NULL)
}
\arguments{
\item{date}{Vector of dates or years, or name of data column containing the 
dates (unquoted).}

\item{flow}{Vector of flows, or name of data column containing the flows
(unquoted).}

\item{data}{Optional; data frame containing the specified data. If
specified, this data frame will be searched first when looking for data
vectors.}
}
\value{
A ggplot2 object. This means that you are able to add more layers
downstream.
}
\description{
Uses ggplot2 to produce a time series plot of values (e.g., annual maximum
flow) over time, where time is binned by year. Subsequent years are connected
with a line, which is broken where a year is missing.
}
\examples{
library(ggplot2)

# Annual maximum series for a Water Survey of Canada gauge.
ams <- ch_rfa_extractamax(Flow ~ Date, CAN05AA008, tol = 350)
ch_ams_plot(Date, Flow, data = ams)

# The returned object is a ggplot, so it can be added to.
ch_ams_plot(Date, 35.31467 * Flow, data = ams) +
  ylab("Annual maximum daily discharge (cfs)") +
  theme_bw()

# Vectors work too, from any source.
ch_ams_plot(ams$Date, ams$Flow)
}
