% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_axis_doy.R
\name{ch_axis_doy}
\alias{ch_axis_doy}
\title{Generates the x-axis for day of year or water year}
\usage{
ch_axis_doy(wyear = 1)
}
\arguments{
\item{wyear}{wyear = 1 for calendar year, = 10 for October 1.}
}
\value{
No value is returned; the axis of the axis of the \R graphic created by \code{ch_regime_plot} is modified.
}
\description{
Generates the x-axis for day of year or day of water year to be 
 used by \code{ch_regime_plot}. Day of water year needs to be generated separately.
}
\examples{
\dontrun{
ch_axis_doy(wyear = 1)  # starts in January
ch_axis_doy(wyear = 10) # starts in October
}
}
\seealso{
\code{\link{ch_regime_plot}}
}
\author{
Paul Whitfield
}
