% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_color_gradient.R
\name{ch_color_gradient}
\alias{ch_color_gradient}
\title{ch_color_gradient}
\usage{
ch_color_gradient(
  x,
  colors = c("darkred", "red", "white", "green", "darkgreen"),
  colsteps = 100,
  climits = NULL
)
}
\arguments{
\item{x}{array of variable}

\item{colors}{an array of colours to form a gradient}

\item{climits}{provide specific limits for common scaling}
}
\description{
set colour gradient
}
