% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_create_wd.R
\name{ch_create_wd}
\alias{ch_create_wd}
\title{Create SAGA Working Directory}
\usage{
ch_create_wd(wd)
}
\arguments{
\item{wd}{working directory file path}
}
\description{
This function creates a SAGA working directory
}
