% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ch_cut_block}
\alias{ch_cut_block}
\title{Cuts a block in time from a time series}
\usage{
ch_cut_block(dataframe, st_date, end_date)
}
\arguments{
\item{dataframe}{A time series dataframe with a \code{Date} variable}

\item{st_date}{Starting date as a string with the format \option{Y/m/d}}

\item{end_date}{Ending date as a string with the format \option{Y/m/d}}
}
\value{
Returns a data frame with the same columns as the original data frame
}
\description{
Allows the user to select a time period from a longer record. 
Could be used to get the same period of time from several stations for comparison.
}
\examples{
subset <- ch_cut_block(W05AA008, "2000/01/01", "2010/12/31")
}
\author{
Paul Whitfield
}
