% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_doys.R
\name{ch_doys}
\alias{ch_doys}
\title{Days of year and water year}
\usage{
ch_doys(Date, water_yr = 10)
}
\arguments{
\item{Date}{an array of \R dates, as produced by \code{as.Date()}}

\item{water_yr}{the month starting the water year, default is 10 (October). If
a value of \code{1} is specified, the \code{10} will be used.}
}
\value{
Returns a dataframe with date information:
\item{Date}{in Date format}
\item{year}{numeric calendar year}
\item{month}{number calendar month}
\item{day}{numeric day of month}
\item{doy}{numeric day of year}
\item{wyear}{numeric water year, labelled by the calendar year in which the
water year \emph{starts}. With the default \code{water_yr = 10}, October 2011
to September 2012 is water year 2011. Note that \code{\link{ch_wtr_yr}} uses the
opposite convention and would label the same period 2012.}
\item{dwy}{numeric day of water year}
}
\description{
Converts an array of dates into a dataframe with date, year, month, doy,
wyear, dowy. 

The day of water year is computed from the first of the specified water year month.
}
\details{
Converts a date array into a data frame with years, wateryears, and days of year and of water year.
}
\examples{
dd <- seq.Date(as.Date("2010-01-01"), as.Date("2018-01-01"),by = 1)
output <- ch_doys(dd, water_yr=10)
head(output)

}
\seealso{
\code{\link{ch_wtr_yr}}
}
\author{
Paul Whitfield, Kevin Shook
}
