% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_fdcurve.R
\name{ch_fdcurve}
\alias{ch_fdcurve}
\title{Plots Flow Duration Curve}
\usage{
ch_fdcurve(DF, normal = FALSE, gust = TRUE, metadata = NULL)
}
\arguments{
\item{DF}{- dataframe of daily flows from \code{ch_read_ECDE_flows}}

\item{normal}{If \code{normal = TRUE} then exceedance probability is normalized. Default is FALSE}

\item{gust}{If \code{TRUE} (the default), adds the curves from Gustard et al. 1992 are added}

\item{metadata}{dataframe of metadata, defaults to HYDAT_list}
}
\value{
flow duration curve plot and a data frame containing the exceedance probability and flow
}
\description{
Creates a Flow Duration Curve based on Observations. 
The plot can include the Gustard Curves. The flows can be 
plotted against probability (default) or the normalized exceedance probability.
}
\examples{
data(HYDAT_list)
data(W05AA008)
# plot with Gustard 1992 curves
test <- ch_fdcurve(W05AA008, normal = FALSE, gust = TRUE)
# plot with normalized exceedance probability
test <- ch_fdcurve(W05AA008, normal = TRUE, gust = FALSE)

}
\references{
Gustard, A., A. Bullock, and J.M. Dixon. 1992. Low flow estimation in the United Kingdom. 
Institute of Hydrology, 292. Wallingford: Institute of Hydrology.

Vogel, R.M., and N.M. Fennessy. 1994. Flow-duration curves. I: New Interpretation and 
confidence intervals. Journal of Water Resources Planning and Management ASCE 120:485-504.

Vogel, R.M., and N.M. Fennessy. 1995. Flow duration curves II: A review of applications 
in water resources planning. Water Resources Bulletin 31:1030-9.
}
\author{
Paul Whitfield
}
