% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_kml_checkcatchment.R
\name{ch_kml_checkcatchment}
\alias{ch_kml_checkcatchment}
\title{Generate kml file to check delineation in Google Earth}
\usage{
ch_kml_checkcatchment(
  file,
  folder_name = "",
  contours,
  catchments,
  channels,
  pp,
  pp_labels = NULL
)
}
\arguments{
\item{file}{}

\item{folder_name}{}

\item{contours}{}

\item{catchments}{}

\item{channels}{}

\item{pp}{}

\item{pp_labels}{}
}
\value{

}
\description{
Generate kml file to check delineation in Google Earth
}
\author{
Dan Moore
}
