% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_read_ECDE_flows.R
\name{ch_read_ECDE_flows}
\alias{ch_read_ECDE_flows}
\title{Reads a file of WSC daily flows from ECCCDataExplorer}
\usage{
ch_read_ECDE_flows(filename)
}
\arguments{
\item{filename}{Datafile retrieved from ECDataExplorer}
}
\value{
Returns a dataframe with the last three rows removed and the Date as Date
}
\description{
Reads in a WSC daily flows file as returned from the program ECCCDataExplorer.
}
\examples{
\dontrun{
# Using a dummy file name as an example
mfile <- "04JD005_Daily_Flow_ts.csv"
mdata <- ch_read_ECDE_flows(mfile)}

}
\author{
Paul Whitfield <paul.h.whitfield@gmail.com>
}
