% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_rfa_boot_nsamax.R
\name{ch_rfa_boot_nsamax}
\alias{ch_rfa_boot_nsamax}
\title{Bootstrap sample of the nonstationary AMAX model}
\usage{
ch_rfa_boot_nsamax(
  object,
  p = c(0.5, 0.8, 0.9, 0.95, 0.98, 0.99),
  newdata = NULL,
  reliability = FALSE,
  nsim = 1000,
  alpha = 0.95,
  verbose = TRUE
)
}
\arguments{
\item{object}{Output from \link{FitNsAmax}.}

\item{p}{Exceeding probability associated to a quantile or reliability level. Default 
value is \code{c(0.5, 0.8, 0.9, 0.95, 0.98, 0.99)}.}

\item{newdata}{Covariates. If \code{NULL} (the default) the input data that served to fit 
the model is used.}

\item{reliability}{Logical. Should the flood quantile associated with 
reliability level be returned. In that case, \code{newdata} must represent
the period of interest. See Durocher et al. (2019) for more details. Default is \code{FALSE}}

\item{nsim}{Number of simulations. Default is \code{1000}.}

\item{alpha}{Confidence level. Default is \code{0.95}.}

\item{verbose}{Logical. Should a progress bar be displayed? Default is \code{TRUE}.}
}
\value{
\item{para}{Parameter of the detrended distribution.}
\item{beta}{Parameter of the trend.}
\item{qua}{flood quantiles.}
}
\description{
Calculates bootstrap samples of the parameters and flood quantiles for the 
nonstationary AMAX model.
}
\examples{
\dontrun{
# this example is explained in the vignette "At-site flood frequency analysis using annual maxima"
hat <- ch_rfa_boot_nsamax(fit.lin, p = prob, reliability = TRUE, nsim = 500, verbose = FALSE)}

}
\seealso{
\link{FitNsAmax}, \link{predict.nsamax}.
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
