% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_rfa_boot_nspot.R
\name{ch_rfa_boot_nspot}
\alias{ch_rfa_boot_nspot}
\title{Bootstrap sample of the nonstationary POT model}
\usage{
ch_rfa_boot_nspot(
  object,
  x,
  rt = c(2, 5, 10, 20, 50, 100),
  newdata = NULL,
  reliability = FALSE,
  nsim = 1000,
  verbose = TRUE
)
}
\arguments{
\item{object}{Output from \link{FitNsPot}.}

\item{x}{Dataset used to fit the nonstationary POT model.}

\item{rt}{Vector of return period to estimate. Default is \code{rt = c(2,5,10,20,50,100}}

\item{newdata}{Covariates. If \code{NULL} (the default), the input data that served to fit 
the model is used.}

\item{reliability}{Logical. Should the flood quantile associated with 
reliability level be returned? In that case, \code{newdata} must represent
the period of interest. See Durocher et al. (2019) for more details. Default is \code{FALSE}.}

\item{nsim}{Number of simulations. Default is \code{1000}.}

\item{verbose}{Logical. Should a progress bar be displayed? Default is \code{TRUE}.}
}
\value{
Returns a list of the bootstrap samples of the parameters and flood quantiles for the 
nonstationary POT model.
}
\description{
Calculates bootstrap samples of the parameters and flood quantiles for the nonstationary POT model.
}
\examples{
\dontrun{
# this example is explained in the vignette "At-site flood frequency analysis using peaks over threshold"
hat <- ch_rfa_boot_nspot(fit, x = flowStJohn, newdata = flowStJohn[yr,], nsim = 50, 
reliability = TRUE, verbose = FALSE)}

}
\seealso{
\link{FitNsPot}, \link{predict.nspot}.
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
