% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ch_safe_GET}
\alias{ch_safe_GET}
\title{Executes GET function with error trapping}
\usage{
ch_safe_GET(url = NULL, filename = NULL, quiet = FALSE)
}
\arguments{
\item{url}{Required. URL to be accessed.}

\item{filename}{Required. File to be used to save data}

\item{quiet}{Optional. if TRUE, GET error/warning messages will be returned}
}
\value{
Returns a character string: "OK", "warning" or "error" depending on the success
of GET
}
\description{
Executes GET function with error trapping
}
\examples{
\donttest{
# Not tested automatically as can be very slow
test_url <- "https://zenodo.org/record/4781469/files/sm_data.csv"
dir_name <- tempdir(check = FALSE)
test_file <- file.path(dir_name, "gs_dem25.tif")
result <- ch_safe_GET(test_url, test_file, quiet = TRUE)
}

}
\author{
Kevin Shook
}
\keyword{internal}
