% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_saga_carea.R
\name{ch_saga_carea}
\alias{ch_saga_carea}
\title{Contributing Area Grid}
\usage{
ch_saga_carea(
  dem,
  saga_wd,
  method = 4,
  linear_threshold = Inf,
  saga_env = RSAGA::rsaga.env()
)
}
\arguments{
\item{dem}{raster object of your dem in the desired projection - should have had sinks removed}

\item{saga_wd}{working directory to write and read saga files}

\item{method}{character or numeric: choice of processing algorithm (default \option{mfd}, or \code{4})}

\item{linear_threshold}{numeric (number of grid cells): threshold above which 
linear flow (i.e. the Deterministic 8 algorithm) will be used; linear flow is 
disabled for \code{linear_threshold = Inf} (the default)}

\item{saga_env}{SAGA environment object. Default is to let saga find it on its own.}
}
\value{
\item{carea}{raster of contributing areas for each pixel.}
}
\description{
Generates a grid of contributing area for each grid cell.
}
\details{
Generates a raster of contributing areas based on a DEM, 
which should have had sinks removed in a pre-processing step. For more information, see 
\command{?rsaga.topdown.processing}.
}
\examples{
# create saga wd using tempdir()
saga_wd <- tempdir()
# use volcano DEM
dem <- ch_volcano_raster()

# fill sinks
filled_dem <-  ch_saga_fillsinks(dem_raw=dem, saga_wd=saga_wd)

# determine contributing area raster using filled_dem
carea <- ch_saga_carea(filled_dem, saga_wd)

# plot contributing area raster
library(raster)
plot(carea)

}
\seealso{
\code{\link{ch_saga_fillsinks}} to fill sinks instead of removing

\code{\link[RSAGA]{rsaga.topdown.processing}} for more information
}
\author{
Dan Moore
}
