% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_saga_catchment.R
\name{ch_saga_catchment}
\alias{ch_saga_catchment}
\title{Create Catchment Polygons}
\usage{
ch_saga_catchment(
  dem,
  saga_wd,
  outlet,
  carea = NULL,
  carea_flag = 0,
  outlet_label = NULL,
  buffsize = 100,
  saga.env = RSAGA::rsaga.env()
)
}
\arguments{
\item{dem}{Raster object of your dem in the desired projection - should have had sinks removed}

\item{outlet}{sf point object with coordinates of catchment outlets in the same projection as the dem}

\item{carea}{raster object containing contributing areas (default none provided)}

\item{carea_flag}{if carea = NULL, 0 = create carea from dem; 1 = read in carea.sdat}

\item{outlet_label}{character vector of labels; if "NULL", numbers are assigned}

\item{saga.env}{SAGA environment object.  Default is to let saga find it on its own.}

\item{sagawd}{name of working directory}

\item{buff_size}{numeric; the buffer radius (m) around catchment outlet to find location on digital stream network.}
}
\value{
Returns an sf object containing catchment polygons
}
\description{
generates a grid of contributing area for each dem grid cell
}
\examples{
ch_saga_catchment()
}
