% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_saga_fillsinks.R
\name{ch_saga_fillsinks}
\alias{ch_saga_fillsinks}
\title{Fill DEM}
\usage{
ch_saga_fillsinks(
  dem_raw,
  saga_wd,
  sinkmethod = "planchon.darboux.2001",
  minslope = 0.1,
  saga.env = RSAGA::rsaga.env()
)
}
\arguments{
\item{dem_raw}{raster object of your raw dem in the desired projection}

\item{saga_wd}{working directory to write and read saga files}

\item{sinkmethod}{options are \option{planchon.darboux.2001} (default), 
\option{xxl.wang.liu.2006} and \option{wang.liu.2006}}

\item{minslope}{minimum slope to impose on filled cells to force drainage (default = 0.1)}

\item{saga.env}{SAGA environment object. Default is to let saga find it on its own.}
}
\value{
\item{dem_ns}{filled dem as a raster object.}
}
\description{
Removes sinks in a DEM by filling. This function is a wrapper to the
\pkg{RSAGA} function \code{rsaga.fill.sinks}.
}
\details{
Removes sinks by filling them. To ensure continuous flow through the DEM, 
a minimum slope can be imposed on cells. Although there are other options, the most important 
are method and minslope; these are implemented in the function. For a full list of 
available options type \command{?rsaga.fill.sinks} at the console. 
The main input is a DEM as a raster object. The main output will be a sink-filled DEM with the 
same topology and coordinate reference system as the input raster.
}
\examples{
# create saga wd using base::tempdir()
saga_wd <- tempdir()
# use volcano DEM
dem <- ch_volcano_raster()

# fill sinks
filled_dem <-  ch_saga_fillsinks(dem_raw=dem, saga_wd=saga_wd)

# plot the difference in raw and filled dem (positive -> filled)
library(raster)
plot(filled_dem-dem)
}
\seealso{
\code{\link{ch_saga_removesinks}} to remove sinks instead of filling
}
\author{
Dan Moore
}
