% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_saga_removesinks.R
\name{ch_saga_removesinks}
\alias{ch_saga_removesinks}
\title{Remove Sinks by Deepening the Drainage Network}
\usage{
ch_saga_removesinks(dem_raw, saga_wd, saga_env = RSAGA::rsaga.env())
}
\arguments{
\item{dem_raw}{raster object of your raw dem in the desired projection}

\item{saga_wd}{working directory to write and read saga files}

\item{saga_env}{SAGA environment object. Default is to let saga find it on its own.}
}
\value{
\item{dem_ns}{processed dem as a raster object.}
}
\description{
Removes sinks by deepening drainage network; alternative to \code{ch_saga_fillsinks}. This 
function acts as a wrapper to 
the \code{rsaga.sink.removal} function.
}
\examples{
# create saga wd using base::tempdir()
saga_wd <- tempdir()

# use volcano DEM
dem <- ch_volcano_raster()

# remove sinks
removed_dem <-  ch_saga_removesinks(dem_raw=dem, saga_wd=saga_wd)

# plot the difference in raw and sink-removed dem 
library(raster)
plot(removed_dem-dem)

}
\seealso{
\code{\link{ch_saga_fillsinks}} to fill sinks in a DEM (instead of removing)
}
\author{
Dan Moore
}
