% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_set_up_workspace.R
\name{ch_set_up_workspace}
\alias{ch_set_up_workspace}
\title{Create directories to hold a project}
\usage{
ch_set_up_workspace(
  super_dir,
  root_dir,
  make_project = FALSE,
  dir_names = NULL,
  rproj_source = NULL
)
}
\arguments{
\item{super_dir}{Name of directory within which to create the root directory}

\item{root_dir}{Name of the root directory for the project}

\item{make_project}{If TRUE, create a .Rproj file in the root directory}

\item{dir_names}{names of directories within the root directory; 
default names are "data", "output", "code", "document", "figures"}

\item{rproj_source}{Path to an existing .Rproj file to use for content}
}
\value{
Does not return a value, simply creates the specified directories, and
(optionally) .Rproj file.
}
\description{
This function creates sub-directories within the root directory 
for a project. It first checks whether the specified root directory exists;
if not, the root directory is created. 

The default names for the sub-directories are "data", "output", "code", 
"document", and "figures"; these can be over-ridden by specifying
alternatives via the `dir_names` argument.

If desired, the function will create a .Rproj file in the root directory.
To do so, the path to an existing .Rproj file must be specified via the
`rproj_source` argument.
}
\examples{
\donttest{
ch_set_up_workspace("~", "new_project")
}

}
\author{
R.D. (Dan) Moore <dan.moore@ubc.ca>
}
