% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_slice.R
\name{ch_slice}
\alias{ch_slice}
\title{Converts doy or dwy into a factor that is used to bin data}
\usage{
ch_slice(doy, step)
}
\arguments{
\item{doy}{A vector of the day of calendar year for the dataset}

\item{step}{Width of bin in days}
}
\value{
Returns a vector of bin numbers that is used as a factor for each day 
in the dataset and prints a message indicating the handling of partial bins
}
\description{
convert doy into a factor which is a number of bins per year.
}
\examples{
doy <- c(1:365)
# first 30 days are 1, 31-60 are 2 etc
dice <- ch_slice(doy, 30)
plot(doy, dice)
}
\seealso{
\code{\link{ch_binned_MannWhitney}} \code{\link{ch_flow_raster_trend}}
}
\author{
Paul Whitfield <paul.h.whitfield@gmail.com>
}
\keyword{internal}
