% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ch_stack_EC}
\alias{ch_stack_EC}
\title{ch_axis_doy moved to own file 2021/04/03
ch_slice move to own file 2021/04/03
ch_sub_set_years moved to own file 2021/04/03
Stacks EC values}
\usage{
ch_stack_EC(data_values = NULL, data_codes = NULL)
}
\arguments{
\item{data_values}{Required. Data frame of year x month or month x day values.}

\item{data_codes}{Required. Data frame of year x month or month x day data codes.}
}
\value{
Returns a data frame with two colums: the data values, and the data codes.
}
\description{
Converts data frames of Environment Canada year x month or 
month x day data to vectors.
}
\examples{
\dontrun{df <- ch_stack_EC(data_values, data_codes)}

}
\author{
Kevin Shook
}
\keyword{internal}
