% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ch_test_url_file}
\alias{ch_test_url_file}
\title{Tests url to see if it will work}
\usage{
ch_test_url_file(url, quiet = FALSE)
}
\arguments{
\item{url}{Required. URL to be checked}

\item{quiet}{Optional. If \code{FALSE} (the default) messages are printed.}
}
\value{
Returns \option{error} if there was an error, \option{warning} if there was a
warning. Otherwise, returns \option{OK}. Strings are returned instead of logical values
to simplify checking result in calling function.
}
\description{
Tests url to see if it will work
}
\examples{
\donttest{
# Not tested automatically as can be very slow
test_url <- "https://zenodo.org/record/4781469/files/sm_data.csv"
ch_test_url_file(test_url, quiet = TRUE)
}

}
\seealso{
See original code on post in Stack Overflow
\href{https://stackoverflow.com/questions/12193779/how-to-use-the-trycatch-function}{
How to write trycatch in R}
}
\author{
Kevin Shook
}
\keyword{internal}
