% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_tidyhydat_ECDE.R
\name{ch_tidyhydat_ECDE}
\alias{ch_tidyhydat_ECDE}
\title{Converts tidyhydat daily flow tibble data to ECDE format}
\usage{
ch_tidyhydat_ECDE(data)
}
\arguments{
\item{filename}{Tibble of daily flows retrieved using \pkg{tidyhydat} function \code{hy_daily_flows}.}
}
\value{
A dataframe or a list of with formats consistent with datafiles from ECDE
}
\description{
Accessing daily flow data using \pkg{tidyhydat} is quick and efficient. However, it sometimes
conflicts with other functions as \pkg{tidyhydat} changes variable names and some default entries. This function converts a tibble obtained 
from a \pkg{tidyhydat} tibble to a dataframe with standard Environment and Climate Change Canada Data Explorer (ECDE) headings.
}
\examples{
mdata <- hy_daily_flowsd(station_number=c("05CK004"))
m_data <- ch_tidyhydat_ECDE(mdata)

mdata <- hy_daily_flows(station_number=c("05CK004","08MF005","05BB001"))
mnew <- ch_tidyhydat_ECDE(mdata)
str(mnew[[1]])
str(mnew[[2]])
str(mnew[[3]]) 
#note the order is in increasing alphabetical order


}
\author{
Paul Whitfield <paul.h.whitfield@gmail.com>
}
