% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_tidyhydat_ECDE_meta.R
\name{ch_tidyhydat_ECDE_meta}
\alias{ch_tidyhydat_ECDE_meta}
\title{Creates an ECDE-like dataframe of metadata}
\usage{
ch_tidyhydat_ECDE_meta(stations, all_ECDE = FALSE)
}
\arguments{
\item{stations}{A vector of WSC station IDs, i.e. c("05BB001", "05BB003", "05BB004", "05BB005"). If 
\code{stations = "all"} then values are returned for all stations.}

\item{all_ECDE}{Should all ECDE values be returned? If \code{FALSE} the default, then
values of \code{Flow}, \code{Sed}, \code{OperSched}, \code{Region}, \code{Datum}, and
\code{Operator} are omitted or will differ from the ECDE values. If \code{all_ECDE = TRUE},
then the function will return values identical to ECDE.}
}
\value{

}
\description{
Extracts tombstone (meta) data for stations from \pkg{tidyhydat} in a 
format similar to that used by the Environment Canada Data Explorer (ECDE). The 
default does not capture all the fields in ECDE, which includes the most recent status 
of many fields such as operating schedule. Returning these values slows the function,
particularly when all WSC stations are selected.
}
\examples{
\dontrun{
stations <- c("05BB001", "05BB003", "05BB004", "05BB005")
result <- ch_tidyhydat_ECDE_meta(stations)
metadata <- result[[1]]
version <- result[[2]]}

}
\seealso{
\code{\link{ch_get_ECDE_metadata}} \code{\link{ch_tidyhydat_ECDE}}
}
\author{
Paul Whitfield <paul.h.whitfield@gmail.com>
}
