% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_tr_sign.R
\name{ch_tr_sign}
\alias{ch_tr_sign}
\title{ch_tr_sign}
\usage{
ch_tr_sign(x, offset = 2)
}
\arguments{
\item{x}{an array of slopes}

\item{offset}{the amount of shift to make values positive integers, default is 2.}
}
\value{
{x} {an array of indices 1, 2, 3}
}
\description{
Converts MK (or other) slopes to integers 1-2-3 (negative, 
 none, positive). These indices can be used to indicate trend direction.
}
\examples{
 mkin <- c( -0.23, 0.34, 0.0, .033, -0.55)
 mkout <- ch_tr_sign(mkin)
 # 1 3 2 3 1
}
\author{
Paul Whitfield
}
