% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_tr_signif.R
\name{ch_tr_signif}
\alias{ch_tr_signif}
\title{ch_tr_signif()}
\usage{
ch_tr_signif(x, pvalue = 0.05)
}
\arguments{
\item{x}{an array of pvalues from statistical test}

\item{pvalue}{critical value, default is 0.05}
}
\value{
Returns an array of indices 1 and 2, where 1 is NS and 2 is significant
}
\description{
Convert pvalues to integers 1 for NS and 2 for significant 
 using a pvalue that can be set (default is 0.05)
}
\examples{
 sin <- c( -0.052, 0.34, 0.012, -.033, -0.55)
 sout <- ch_tr_signif(sin)
 # 1 1 2 2 1
}
\author{
Paul Whitfield
}
