% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_volcano_raster.R
\name{ch_volcano_raster}
\alias{ch_volcano_raster}
\title{Create Test SpatRaster}
\usage{
ch_volcano_raster()
}
\value{
Returns a SpatRaster object of land surface elevations.
}
\description{
Creates a \pkg{terra} object of land surface elevations, as
used to test/demonstrate many functions requiring a digital elevation model 
(DEM).
}
\details{
No arguments are required as the DEM is created from the \pkg{base} 
\code{volcano} matrix of elevations.
}
\examples{
\dontrun{
test_raster <- ch_volcano_raster()
}

}
\author{
Dan Moore Kevin Shook and Joel Trubilowicz
}
