% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_wbt_check_whitebox.R
\name{ch_wbt_check_whitebox}
\alias{ch_wbt_check_whitebox}
\title{Checks for WhiteboxTools executable}
\usage{
ch_wbt_check_whitebox()
}
\value{
If \code{WhiteboxTools} cannot be found, a message explaining what to do is displayed.
}
\description{
Checks for WhiteboxTools executable
}
\examples{
\dontrun{
# Only proceed if Whitebox executable is installed
library(whitebox)
if (check_whitebox_binary()){
  ch_wbt_check_whitebox()
} else {
  message("Example not run as Whitebox executable not found")
}
}
}
\author{
Kevin Shook
}
\keyword{internal}
