% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_wtr_yr.R
\name{ch_wtr_yr}
\alias{ch_wtr_yr}
\title{Designation of the water year}
\source{
http://stackoverflow.com/questions/27626533/r-create-function-to-add-water-year-column
}
\usage{
ch_wtr_yr(dates, start_month = 10)
}
\arguments{
\item{dates}{A vector of dates with actual year}

\item{start_month}{Month in which the year starts (defaults to October)}
}
\value{
A numeric vector of water years, labelled by the calendar year in
which the water year \emph{ends}. With the default \code{start_month = 10},
October 2011 to September 2012 is water year 2012. Note that the \code{wyear}
column returned by \code{\link{ch_doys}} uses the opposite convention and would
label the same period 2011.
}
\description{
Display water year
}
\examples{
date <- seq(as.Date("1910/1/1"), as.Date("1912/1/1"), "days")
wtr_yr_date <- ch_wtr_yr(dates=date, start_month=10)
df <- data.frame(wtr_yr_date, date)
}
\seealso{
\code{\link{ch_doys}}
}
