% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydRology-package.R
\docType{package}
\name{CSHS-hydRology-package}
\alias{CSHS-hydRology-package}
\title{Functions for Canadian hydrological analyses}
\description{
An R package collecting useful functions for hydrologic analysis
}
\references{
To cite \pkg{CSHS-hydRology} in publications, use the command \code{citation("CSHS-hydRology")} to get the current version of the citation.\cr
}
