% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_CvRoi.R, R/rfa_head_roicv.R,
%   R/rfa_plot_roicv.R
\name{CvRoi}
\alias{CvRoi}
\alias{head.roicv}
\alias{plot.roicv}
\title{Cross-validation using region of influence and kriging}
\usage{
CvRoi(x, nk, phy, similarity, kriging = NULL, ker = TRUE,
  model = "Exp", fold = 5, verbose = TRUE)

\method{head}{roicv}(x, crit = "mad", verbose = TRUE, ...)

\method{plot}{roicv}(x, crit = "mad", best.col = "red",
  best.pch = 16, best.cex = 1, ...)
}
\arguments{
\item{x}{Data.}

\item{nk}{List of neighborhood sizes to try.}

\item{phy}{Formula defining the physical descriptors.}

\item{similarity}{Formula defining the covariates used
to evaluate the similarity between site, i.e. Euclidean distance to 
the target.}

\item{kriging}{Formula defining the spatial covariates.
Necessary for predicting residuals using spatial correlation.}

\item{ker}{Should a (Epanechnikov) kernel be used in to weight
local regression model. Otherwise uniform weight are used.}

\item{model}{Variogram model. See \link[gstat]{vgm}.}

\item{fold}{Number of group used in the cross-validation scheme. 
Can also be a vector defining the group for each site.}

\item{verbose}{Logical. Should a progress bar be displayed.}

\item{crit}{Cross-validation criteria used to evaluate the best choice.}

\item{...}{More arguments to pass to the plot function (\link{par}).}

\item{best.col, best.pch, best.cex}{Argument for the point indication the best
choice.}
}
\description{
Return a matrix of criteria evaluated by using region of
influence (ROI) and kriging. It includes: Root mean square error 
(\code{rmse}), relative RMSE (\code{rrmse}), Nash-Sutcliffe (\code{nsh}),
Mean absolute deviation (\code{mad}), relative MAD (\code{rmad}) and
the skill score based on MAD (\code{smad}). 
The latter \code{smad} has the same form as \code{nsh} except that 
absolute error are taken instead of square error.
}
\examples{

attach(flowUngauged)

## Using multidimensional scaling for projecting coordinates
coord <- cbind(lon,lat)
coord <- cmdscale(GeoDist(coord))
colnames(coord) <- c('lon','lat')

## Transform data if necessary
xdf <- data.frame(y      = log(l1),
                  area   = scale(log(area)),
                  wb     = scale(log(wb)),
                  stream = scale(log(stream)),
                  map    = scale(log(map)),
                  coord)

 ## select a validation and training set
 set.seed(9382)
 vid <- runif(nrow(xdf)) > .8
 tid <- !vid

 # formula of the relationship between flood quantile and descriptors
 fphy <- y ~ area + map + wb + stream
 fsimilarity <- ~ area + map

 ## Perform cross-validation.
 system.time(out <- CvRoi(x = xdf, nk = seq(20,150, 10), fold = 5,
                 phy = fphy,  similarity = fsimilarity, model = 'Exp'))

 head(out, 'nsh')
 plot(out, 'mad')

}
\references{
Martin Durocher, Donald H. Burn, Shabnam Mostofi Zadeh & Fahim Ashkar (2019) 
  Estimating flood quantiles at ungauged sites using nonparametric regression 
  methods with spatial components, Hydrological Sciences Journal, 64:9, 
  1056-1070, https://doi.org/10.1080/02626667.2019.1620952
}
\seealso{
\link{FitRoi}
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
