% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_DataWide.R
\name{DataWide}
\alias{DataWide}
\alias{DataWide.data.frame}
\alias{DataWide.formula}
\alias{DataWide.matrix}
\title{Transpose dataset to a wide format}
\usage{
DataWide(x, ...)

\method{DataWide}{data.frame}(x, order.time = TRUE, order.site = FALSE)

\method{DataWide}{formula}(form, x, ...)

\method{DataWide}{matrix}(x, ...)
}
\arguments{
\item{x}{Dataset in long format. First column should be the value 
to transpose.
The second and third columns are respectively site and time variable.
If the time variable is omitted, the pivot will ignore it and all columns
will start at the first position.}

\item{form}{Formula that specifies the site and time variable.
Must have the form: `value ~ site + time`.}
}
\description{
Return a matrix where columns correspond to data.
 Each row represent a specific time and missing values are added where need.
 The pivoted variable should be a numerical value.
}
\examples{

data(flowAtlantic)
xd <- flowAtlantic$ams 

## transpose by year
xd$year <- format(xd$date, '\%Y')
DataWide(ams ~ id + year, xd)[60:70,1:5]

## Without time
DataWide(ams ~ id, xd)[1:6,1:5]

}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
