% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.subset.R
\name{date.subset}
\alias{date.subset}
\title{Subset Date by String}
\usage{
date.subset(df, prd)
}
\arguments{
\item{df}{data frame of time series data; includes a 'Date' attribute}

\item{prd}{date range as string, "YYYY-MM-DD/YYYY-MM-DD"}
}
\value{
\item{df}{subsetted data frame}
}
\description{
date.subset subsets a data frame with Date attribute using date string
}
\examples{

dd <- seq.Date(as.Date("2010-10-01"),as.Date("2013-09-30"),by=1)
x <- rnorm(length(dd))
y <- abs(rnorm(length(dd)))*2
df <- data.frame("Date"=dd,x,y)
prd <- "2011-10-01/2012-09-30"
summary(date.subset(df,prd))

}
\author{
Robert Chlumsky <rchlumsk@gmail.com>

This function subsets a data frame by an inclusive date range, provided as 
a string by the prd argument. This function is meant to emulate the subsetting 
capability of the xts package.
}
\keyword{data}
\keyword{date}
\keyword{subset}
