% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_DistSeason.R
\name{DistSeason}
\alias{DistSeason}
\alias{DistSeason.data.frame}
\alias{DistSeason.formula}
\title{Distance in seasonal space}
\usage{
DistSeason(x, ...)

\method{DistSeason}{data.frame}(x, w = 1/pi)

\method{DistSeason}{formula}(form, x, w = 1/pi)
}
\arguments{
\item{w}{Weight to favor angle over radius. 
By default it is 1/pi, which bring angle in the interval [0,1].}

\item{form, x}{Formula and dataset providing the coordinates of the
seasonal space. Must be of the form `radius ~ angle`.}

\item{r, a}{Coordinates in the seasonal space: radius `r` and angle `a`.}
}
\description{
Return a matrix of distances between points in the seasonal space that
characterizes timing and regularity.
It is equivalent to Euclidean distance applied to regularity (radius)
and timing (angle) separately.
}
\examples{

scoord <- data.frame(radius = runif(5), 
                     angle = runif(5,0,2*pi))

DistSeason(scoord)
DistSeason(radius ~ angle , scoord)


}
\references{
Durocher, M., Burn, D. H., & Ashkar, F. (2019). Comparison of estimation
  methods for a nonstationary index-flood model in flood frequency
  analysis using peaks over threshold. https://doi.org/10.31223/osf.io/rnepc
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
