% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{doys}
\alias{doys}
\title{Days of year and water year}
\usage{
doys(Date, mon = 10)
}
\arguments{
\item{Date}{A vector of R dates.}

\item{mon}{The month starting the water year, default is 10 (October).}
}
\value{
Returns a dataframe with differently-formatted dates
\describe{
 \item{Date}{original Date}
 \item{year}{numeric calendar year}
 \item{month}{numeric calendar month}
 \item{doy}{numeric day of year}
 \item{wyear}{numeric water year}
 \item{dwy}{numeric day of water year}
}
}
\description{
Converts a vector of dates into a dataframe with date, doy, dowy, year
}
\examples{
dd <- seq.Date(as.Date("2010-01-01"), as.Date("2018-01-01"), by = 1)
output <- doys(dd)
head(output)

}
\author{
Paul Whitfield <paul.h.whitfield@gmail.com>
}
\keyword{internal}
