% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdcurve.R
\name{fdcurve}
\alias{fdcurve}
\title{Flow Duration Curve based upon Observations}
\usage{
fdcurve(flow, title, normal = FALSE, gust = TRUE)
}
\arguments{
\item{flow}{- array of length n containing the flow data}

\item{title}{: uses format of 05BB001 - Bow River at Banff AB* from get_wscstation}

\item{normal}{: FALSE is default, if normal=TRUE then exceedance probability is normalized}

\item{gust}{: True is default,  adds the curves from Gustard et al 1992; if FALSE they are omitted}
}
\description{
A flow duration curve is a plot of flow magnitude against exceedance probability
}
\examples{

\dontrun{test <-fdcurve(flow, title, normal=FALSE, gust=TRUE)}

}
\references{
Gustard, A., A. Bullock, and J.M. Dixon. 1992. Low flow estimation in the United Kingdom. Institute of Hydrology, 292. Wallingford: Institute of Hydrology.

Vogel, R.M., and N.M. Fennessy. 1994. Flow-duration curves. I: New Interpretation and confidence intervals. Journal of Water Resources Planning and Management ASCE 120:485-504.

Vogel, R.M., and N.M. Fennessy. 1995. Flow duraction curves II: A review of applications in water resources planning. Water Resources Bulletin 31:1030-9.
}
\author{
Paul Whitfield <paul.h.whitfield@gmail.com>
}
