% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_rfa_FindThresh.R
\name{FindThresh}
\alias{FindThresh}
\title{Find automatically a best threshold}
\usage{
FindThresh(
  x,
  method = "sgn",
  tol.sgn = 0.25,
  tol.ppy = 2,
  qua = "q10",
  tol.qua = Inf,
  ppy = c(0, Inf)
)
}
\arguments{
\item{x}{output form \link{SearchThresh}.}

\item{method}{Method to identify the threshold. Must be one of 
\code{sgn}, \code{sgn-max}, \code{sgn-ppy}, \code{max} or \code{ppy}.}

\item{tol.sgn}{Selection criteria associated with a given p-value of 
the Anderson-Darling test.}

\item{tol.ppy}{Selection criteria associated with a given average of 
peaks per year.}

\item{qua}{Flood quantiles. Must be one of \code{'q2'}, \code{'q5'},
\code{'q10'}, \code{'q20'}, \code{'q50'} or \code{'q100'}.}

\item{tol.qua}{Selection criteria associated with a flood quantile.}

\item{ppy}{Range of accepted averages of peaks per year.}
}
\description{
The function \code{Findthresh} can be used to identify the best
threshold according to a specific rules.
}
\details{
The method \code{sgn} find the first threshold that has a p-value over a 
given value (\code{tol.sgn}) of the Anderson-Darling test, while \code{max} 
find the maximum p-value.
The \code{ppy} method find the threshold that has the nearest average of
peaks per year to \code{tol.ppy}.
The method \code{sgn-max} find the thresholds associated with both methods
and return the lowest one. 
The method \code{sgn-ppy} try first to find the \code{sgn} threshold.
If not threshold meets the required p-value, the \code{ppy} threshold is 
returned.

The criteria \code{ppy}, initially filter all the candidates thresholds that 
are not found in a given interval of average peaks per year. 
With the \code{sgn} method an additional condition associated to a flood 
quantile \code{tol.qua} can be specified. 
It imposes that the relative discrepancies between the candidate threshold 
and a reference is respected in addition to the criteria with the p-value.
This reference is the average of the flood quantiles of the 5 lowest 
candidate thresholds in the set of candidates.
Normally, lower thresholds should have reached stability and their choice can
be controlled by the argument \code{ppy}.

The method \code{fdr} is a variation of the \code{sgn} method. 
It verifies that the p-value of a given threshold is above a critical value 
and also  that here after the GPA distribution is a proper distribution for 
all the thresholds using false discovery rate (FDR) at the same significance level.
}
\examples{

## Create list of candidate threshold
lstu <- which.floodPeaks(flow~date, flowStJohn, u =500, r = 14)
lstu <- sort(unique(flowStJohn[lstu,'flow']))
lstu <- lstu[seq(1,length(lstu)-30,2)]

out <- SearchThresh(flow~date, flowStJohn, u = lstu, declust = 'wrc', r = 14)

cid <- c('u','ppy','ad')
FindThresh(out, method = 'sgn', tol.sgn = 0.25)[,cid]
FindThresh(out, method = 'ppy', tol.ppy = 2)[,cid]
FindThresh(out, method = 'max')[,cid]

## Make find the first threshold that have at-least p-value > 0.1
## and relative descripencies between Q100 of less than 20\%.
## Otherwise the ppy threshold is chosen.
## Note here that sgn is chosen because the flood quantile stabilizes 
## quickly. 
FindThresh(out, method = 'sgn-ppy', tol.sgn = 0.25, tol.ppy = 1.5,
           ppy = c(1,3), qua = 'q10', tol.qua = 0.2)[,cid]


}
\references{
Durocher, M., Zadeh, S. M., Burn, D. H., & Ashkar, F. (2018). Comparison of 
  automatic procedures for selecting flood peaks over threshold based on 
  goodness-of-fit tests. Hydrological Processes, 0(0).
  https://doi.org/10.1002/hyp.13223
  
Solari, S., Eguen, M., Polo, M. J., & Losada, M. A. (2017). Peaks Over 
  Threshold (POT): A methodology for automatic threshold estimation using 
  goodness of fit p-value. Water Resources Research, 53(4), 2833-2849. 
  https://doi.org/10.1002/2016WR019426
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
