% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_FitGev.R
\name{FitGev}
\alias{FitGev}
\title{Fit Generalized Extreme Value (GEV) distribution using Generalized
maximum likelihood}
\usage{
FitGev(x, varcov = TRUE, mu = -0.1, sig2 = 0.015,
  method.optim = "BFGS", ...)
}
\arguments{
\item{x}{Data.}

\item{varcov}{Logical. Should the covariance matrix be returned.}

\item{mu, sig2}{Mean and variance of the Beta prior.}

\item{method.optim}{Optimisation method used by \code{\link{optim}}}

\item{...}{Other parameter pass to \code{\link{optim}}}
}
\description{
Fit a GEV distribution on annual maxima using the generalized
maximum likelihood method with Beta prior.
The output is of the class \code{amax}. See \link{FitAmax}.
Asymptotic result are computed like the maximum likelihood approach.
}
\section{References}{


Martins, E.S., Stedinger, J.R., 2000. Generalized maximum-likelihood
generalized extreme-value quantile estimators for hydrologic data.
Water Resour. Res. 36, 737–744. https://doi.org/10.1029/1999WR900330
}

\examples{

x <- ExtractAmax(flow~date, flowStJohn)$flow

## Using the default physiographic prior.
fit <- FitGev(x)
print(fit)
coef(fit)
vcov(fit)
predict(fit, ci = 'delta')

## A uniform prior on interlval -0.5 to 0.5 can be used to approximate
## the maximum likelihood estimate.

AIC(FitAmax(x, distr = 'gev', method ='mle'))
AIC(FitGev(x, mu = 0, sig2 = 1/12))

## A regional study can be performed by using an empirical prior
## Here 20 sites with the same GEV distribution are simulated
## without priors.
## Then a regional estimate is obtained using an empirical prior

xmat <- replicate(20, rAmax(20, c(100,3, -.1), 'gev') )
flist <- apply(xmat,2, FitAmax, distr = 'gev', varcov = FALSE)
pmat <- sapply(flist, getElement,'para')

kap0 <- pmin(.5,pmax(-.5,pmat[3,]))

FitGev(xmat[,1], mu = mean(kap0), sig2 = var(kap0))
FitAmax(xmat[,1], distr = 'gev', method = 'mle')

}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
