% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_rfa_FitNsAmax.R, R/ch_rfa_print_nsamax.R
\name{FitNsAmax}
\alias{FitNsAmax}
\alias{print.nsamax}
\title{Nonstationary model for annual maxima fitted by least-squares and L-moments}
\usage{
FitNsAmax(form, x, distr = "gev", type = "add", k = 2, tol.gev = 0)

\method{print}{nsamax}(x, ...)
}
\arguments{
\item{form}{Formula of the trend.}

\item{x}{Data including covariates.}

\item{distr}{Distribution.}

\item{type}{Type of model. Either additive (\code{"add"}) 
or multiplicative (\code{"mult"}).}

\item{k}{The penalty per parameter to be used in AIC. See \link{AIC}.}

\item{tol.gev}{Accepted difference between the AIC of the GEV and the best 
best distribution. If the difference is inferior to \code{tol.gev}, the 
GEV distribution is prefered.}

\item{...}{Other parameters.}
}
\value{
\item{formula}{Formula of the trend.}
\item{data}{Input data.}
\item{beta}{Parameters of the trend.}
\item{fitted}{Fitted value of the trend.}
\item{type}{Type of model.}
\item{para}{Parameters of the distribution. See \link{lmom2par}}
}
\description{
Return the fitting of a nonstationary model for annual
maxima that fit first a trend using regression and then estimates the
residual distribution with L-moments. 
If a list of distribution is provided, the best distribution is selected by AIC.
}
\examples{

data(flowStJohn)

amax <- ExtractAmax(flow~date, flowStJohn, tol = 365)

## Recenter year to start at 1
amax$year <- as.integer(amax$yy) - 1926

## Add a trend
amax$flow <- amax$flow + amax$year * 10

fit <- FitNsAmax(flow~year, amax, 'gev', type = 'mult')

## Examination of the residuals
print(fit)
plot(fitted(fit), residuals(fit))

## Visualization of the trend
plot(flow~year, amax)
lines(amax$year, fitted(fit), col = 'red')

## Using alternative distribution and descriptors
fit1 <- FitNsAmax(flow~year, amax, 'ln3', type = 'mult')
fit2 <- FitNsAmax(flow~poly(year,2), amax, 'ln3', type = 'mult')

## Evaluation of the AIC for model selection
AIC(fit1)
AIC(fit2)

## Automatic selection of the distribution
fit <- FitNsAmax(flow~year, amax, c('gev','ln3'), type = 'mult')
print(fit)

## Simulate the model
simulate(fit)

}
\seealso{
\link{FitAmax}, \link{predict.nsamax}.
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
