% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_FitNsAmaxMle.R, R/rfa_print_nsamax.mle.R
\name{FitNsAmaxMle}
\alias{FitNsAmaxMle}
\alias{print.nsamax.mle}
\title{Nonstationary model for annual maxima fitted by maximum likelihood}
\usage{
FitNsAmaxMle(form, x, distr, type = "add", varcov = TRUE,
  method = "BFGS", control = NULL, para0 = NULL, quiet = FALSE,
  ...)

\method{print}{nsamax.mle}(x, ...)
}
\arguments{
\item{form}{Formula describing the trend in the location parameter.}

\item{x}{Data.}

\item{distr}{Fitted distribution. Must be one of 
\code{('gev','glo','gno','pe3')}.}

\item{type}{Type of model defining the relation between the location
and the scale parameter. Must be one of \code{('add','mult')}.}

\item{varcov}{Logical. Should the covariance matrix be returned.}

\item{method, control, ...}{Argument pass to \link{optim}}

\item{para0}{Starting parameter estimates.}

\item{quiet}{Logical. Should the message and warnings be displayed.}
}
\value{
\item{formula}{Formula for the trend in the location parameter.}
\item{data}{Input data.}
\item{distr}{Distribution.}
\item{type}{Type of model.}
\item{para}{Model parameters. The last two parameters characterize the
  scale and shape parameters.}
\item{varcov}{Covariance matrix.}
\item{convergence,message}{Output form \link{optim} describing the 
  convergence of the algorithm.}
}
\description{
Return the fitting of a nonstationary model for annual
maxima fitted by maximum likelihood. A trend is set for the location parameter.
The scale parameter can be constant or proportional to the location parameter.
}
\examples{

data(flowStJohn)

amax <- ExtractAmax(flow~date, flowStJohn, tol = 365)

## Recenter year to start at 1
amax$year <- as.integer(amax$yy) - 1926

## Add a trend
amax$flow <- amax$flow + amax$year * 10

fit <- FitNsAmaxMle(flow~year, amax, distr = 'gev', type = 'add')
print(fit)

## Function to extract the parameter for every year.
cc <- coef(fit)

## Covariance matrix
vcov(fit)

## Graphics of the location parameter +/- the scale parameters
plot(flow~year, amax)
lines(cc[,1], col = 'red')
lines(cc[,1] + cc[,2], col = 'red', lty = 2)
lines(cc[,1] - cc[,2], col = 'red', lty = 2)

## Selecting a model based on AIC
fit2 <- FitNsAmaxMle(flow~year, amax, distr = 'gno', type = 'mult')

AIC(fit)
AIC(fit2)

## Simulate the model
simulate(fit2)

}
