% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_rfa_FitPoolMargin.R
\name{FitPoolMargin}
\alias{FitPoolMargin}
\title{Fitting marginal distributions of a pooling group}
\usage{
FitPoolMargin(x, distr, method = "lmom", method.optim = "BFGS", ...)
}
\arguments{
\item{x}{Data with time in rows and sites in columns.}

\item{distr}{Fitting distribution}

\item{method}{Estimation method.}

\item{method.optim, }{... Further arguments pass to \link{optim} when 
maximum likelihood method is used}

\item{...}{Other parameters.}
}
\value{
\item{para}{Matrix of the estimated parameters.}
\item{distr}{Fitting distribution.}
\item{na}{Indices of the missing values.}
\item{dim}{Dimension of the input data.}
}
\description{
Returns an object containing the parameter estimates of each site sites
inside a pooling groups. Estimation can be performed by L-moments or 
maximum likelihood for the five distributions: GEV, GLO, GNO, PE3, GPA.
}
\examples{

x <- replicate(3, rgev(100, 100, 30, 0))

fit <- FitPoolMargin(x, 'gev', method = 'mle')

print(fit)
predict(fit, p = c(.9, .99))

xx <- simulate(fit, nsim = 2, corr = .2)


}
\seealso{
\link{FitPoolMle}, \link{Amax}, \link{GPA}.
}
