% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_FitPoolMle.R
\name{FitPoolMle}
\alias{FitPoolMle}
\title{Fitting a regional model by likelihood techniques.}
\usage{
FitPoolMle(x, distr = "gev", type = "mean", index = NULL,
  para0 = NULL, ...)
}
\arguments{
\item{x}{Data with times in rows and sites in columns.}

\item{distr}{Fitting distribution: \code{('gev','gpa','glo','gno','pe3')}.}

\item{type}{Type of regional models to be fitted. Either ones of 
\code{'mean'}, \code{'cv'}, \code{'shape'}.
See the details section.}

\item{index}{A scale factor to standardize each sites.}

\item{para0}{Initial parameter estimates.}

\item{...}{Other parameters to pass to the \link{optim} function.}
}
\value{
\item{index}{Site-specific parameter.}
\item{para}{Regional parameters.}
\item{llik}{The sum of the independent log-likelihood.}
\item{type}{Type of regional model.}
\item{na}{Indices of the missing values.}
\item{dim}{Dimension of the input data.}
}
\description{
Return the fitting regional of a multiple sites based on likelihood 
techniques.
}
\details{
Different regional models can be selected by the argument \code{type}. 
All models assumed the same distribution for all sites, which must be one of
the followings: GEV, GLO, GNO, PE3 and GPA. 
Please see Hosking and Wallis (1997) for more details on the distributions.
All distributions have a location, scale and shape parameter.
One exception is the GPA distribution for which the location parameter is 
known to be zero.
The regional models combine local and regional parameters to fully 
characterize each at-site distribution.
All sites are estimated jointly by maximizing the sum of the log-likelihood
of the individual sites.

For \code{type = 'mean'} an index-flood model is estimated by assuming that all 
distributions are proportional up to a scale factor. 
By default the scale factor is the at-site mean, but can be replaced 
by other quantities as passed to \code{index}. 
The likelihood techniques is then used to estimate the regional parameters.

The option \code{type = 'shape'} assume that site has a common shape 
parameter and the remaining parameters are site-specific. 
For \code{type = 'cv'} the only the location parameters are site-specific.
Therefore, the scale parameters are set as proportional to 
the location parameters up to a regional parameter 
that has similar a interpretation as a coefficient of variation.
For this regional model, a constant shape parameter is also assumed for all
sites. 

It should be pointed out that the independent log-likelihood can be used to 
select the distributions when it has the same number of parameters. 
Consequently, GPA should not be compared to the other distributions and  the
same type of model should be used.
}
\examples{

xd <- flowAtlantic$ams 
## transpose by year
xd$year <- format(xd$date, '\%Y')
xw <- DataWide(ams ~ id + year, xd)

## Fitting a regional shape parameter
fit.gev <- FitPoolMle(xw, 'gev', type = 'shape')
fit.gno <- FitPoolMle(xw, 'gno', type = 'shape')
print(fit.gev)

## In this case GEV is a better model. 
print(fit.gev$llik)
print(fit.gno$llik)

## Return all the parameters of the marginal distribution.
coef(fit.gev)

## Exemple of index-flood model estimated by MLE
fit <- FitPoolMle(xw, 'gev', type = 'cv')
print(fit)


}
\references{
* Hosking, J. R. M., & Wallis, J. R. (1997). Regional frequency analysis: 
an approach based on L-moments. Cambridge Univ Pr.

* Padoan, S. A., Ribatet, M., & Sisson, S. A. (2010). Likelihood-Based 
Inference for Max-Stable Processes. Journal of the American Statistical 
Association, 105(489), 263-277. https://doi.org/10.1198/jasa.2009.tm08577
}
\seealso{
\link{FitPoolMargin} and \link{predict.poolmle}.
}
