% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_FitPot.R
\name{FitPot}
\alias{FitPot}
\alias{FitPot.data.frame}
\alias{FitPot.matrix}
\alias{FitPot.formula}
\alias{FitPot.numeric}
\alias{coef.fpot}
\alias{vcov.fpot}
\alias{print.fpot}
\alias{logLik.fpot}
\alias{AIC.fpot}
\title{Peak over threshold (POT)}
\usage{
FitPot(x, ...)

\method{FitPot}{data.frame}(x, ...)

\method{FitPot}{matrix}(x, ...)

\method{FitPot}{formula}(form, x, ...)

\method{FitPot}{numeric}(x, dt = NULL, u = 0, method = "mle",
  declust = "none", r = 1, rlow = 0.75, nsim = 1000,
  varcov = TRUE, unit = 365.25, ...)

\method{coef}{fpot}(object, rate = FALSE, ci = FALSE, alpha = 0.05,
  ...)

\method{vcov}{fpot}(object, rate = FALSE, ...)

\method{print}{fpot}(x, ...)

\method{logLik}{fpot}(object, ...)

\method{AIC}{fpot}(object, k = 2, ...)
}
\arguments{
\item{x, form}{Dataset. If a matrix or data.frame is passed as argument the
first column must be the time and the second the values. A formula can be
used to specify which variables of a data.frame to use. In this case, it
must have the form \code{value~time}.}

\item{...}{Other parameters.}

\item{dt}{Date or time of observation.}

\item{u}{Threshold.}

\item{method}{Estimation method. Either \code{'lmom'}, \code{'mle'} or
\code{'mle2'}.}

\item{declust}{If necessary, declustering method.
Either \code{'run'} or \code{'wrc'}.}

\item{r}{Lag parameter for declustering. Either the running length between clusters
or the minimum separating time between two flood Peaks.
The scale must coincide with the observation date \code{dt}.}

\item{rlow}{For WRC, recession level between two flood peaks in percentages.}

\item{nsim}{Number of bootstrap samples.}

\item{varcov}{Logical. Should the covariance matrix of the estimated 
parameters be evaluated.}

\item{unit}{Length of cycle. Data are normaly years and \code{unit} = 365.25.
Can be change to 12 or 52 for daily and weekly data.}

\item{object}{Output of \code{FitPot}.}

\item{rate}{Logical. Should the estimated probability of exeedance must be
included in the covariance matrix.}

\item{ci}{For \code{'coef'} should confidence interval be returned.}

\item{alpha}{Probability outside the confidence interval.}

\item{k}{The penalty per parameter to be used.}
}
\description{
Fit the parameters of a thresholding model using Generalized Pareto
distribution (GPA). Include declustering techniques.
}
\details{
The access functions \code{coef} and \code{vcov} return respectively the
parameters and the variance-covariance matrix of the POT model. For the L-moment
method the covariance matrix is using bootstraps.
The access function \code{predict} evaluates flood quantiles.
If \code{dt} is a Date the return period is computed in years using the range
of observation.
}
\section{References}{


Coles S. (2001) An introduction to statistical modeling of extreme values.
  Springer Verlag.

Davison AC, Smith RL. (1990) Models for Exceedances over High Thresholds.
  Journal of the Royal Statistical Society Series B (Methodological).
  52(3), 393-442.
}

\examples{

xd <- rgpa(100, 1, -.2)
fit <- FitPot(xd, u = 0)

print(fit)
vcov(fit)
predict(fit)
coef(fit, ci = TRUE)

fit <- FitPot(flow~date, flowStJohn, u = 1000,
               declust = 'wrc', r = 14)

print(fit)
plot(flow~date,flowStJohn, type = 'l')
points(fit$time,fit$excess+fit$u, col = 2, pch = 16)
abline(h=1000, col = 3, lwd = 2)

predict(fit, se = TRUE, ci = 'delta')

}
\seealso{
\link{which.floodPeaks}, \link{which.clusters}, \link{PlotMrl}.
}
