% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_FitRegLmom.R
\name{FitRegLmom}
\alias{FitRegLmom}
\alias{coef.reglmom}
\alias{print.reglmom}
\alias{plot.reglmom}
\alias{sitenames}
\title{Fitting an index-flood model}
\usage{
FitRegLmom(x, distr = NULL, type = "amax", diagnostic = FALSE,
  diagnostic.nsim = 1000, nmom = 4)

\method{coef}{reglmom}(obj, distr = NULL)

\method{print}{reglmom}(obj)

\method{plot}{reglmom}(obj)

sitenames(obj)
}
\arguments{
\item{x}{Dataset with sites in columns and time in rows.}

\item{distr}{Regional distribution. See \link{Amax}.
If \code{NULL} an automatic selection is performed.}

\item{type}{Type of input data. Either annual maximums (\code{'amax'}) or
peaks over threshold (\code{'pot'}).}

\item{diagnostic}{Should the homogeneity (H) and goodness of fit
(Z-score) be evaluated.}

\item{diagnostic.nsim}{Number of simulations used to evaluate the
diagnostic statistics.}

\item{nmon}{Number of L-moments to evaluate.}
}
\value{
\describe{
  \item{type}{Type of input data}
  \item{distr}{Regional distribution}
  \item{para}{Parameters of the regional distribution}
  \item{lmom}{At-site L-moments.}
  \item{rlmom}{Regional L-moments.}
  \item{nrec}{Record lengths.}
  \item{stat}{Homogenous criteria and Z-score for goodness-of-fit.}
  \item{discord}{Discordance measures.}
}
}
\description{
Returns the at-site L-moments, regional L-moments and the parameters of
the regional growth curve for an index-flood model. 
If required, the homogeneity and goodness-of-fit measures are included.
The function can be used for analyzing both annual maximums and peaks over
threshold.
}
\details{
The function \code{coef} return the parameter of the at-site distributions
estimated L-moments. The function \code{plot} will present a L-moment ratio
diagram.
}
\section{References}{


Hosking, J. R. M., & Wallis, J. R. (1997). Regional frequency analysis:
  an approach based on L-moments. Cambridge Univ Pr.

Mostofi Zadeh, S., Burn, D.H. (2019). A Super Region Approach to Improve
  Pooled Flood Frequency Analysis. Canadian Water Resources Journal
  0, 1-14. https://doi.org/10.1080/07011784.2018.1548946

Mostofi Zadeh, S., Durocher, M., Burn, D.H., Ashkar, F., 2019.
  Pooled flood frequency analysis: a comparison based on peaks-over-threshold
  and annual maximum series. Hydrological Sciences Journal 0, null.
  https://doi.org/10.1080/02626667.2019.1577556
}

\examples{

data(flowAtlantic)
xd <- cbind(flowAtlantic$ams, 
             year = format(flowAtlantic$ams$date, '\%Y'))
             
xmat <- DataWide(ams ~ id + year, xd)

fit <- FitRegLmom(xmat)
print(fit)
plot(fit)
coef(fit, 'pe3')


}
\seealso{
\link{predict.reglmom}, \link{PoolRemove}.
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
