% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_FitRoi.R, R/rfa_predict_roi.R,
%   R/rfa_print_roi.R, R/rfa_residuals_roi.R
\name{FitRoi}
\alias{FitRoi}
\alias{predict.roi}
\alias{print.roi}
\alias{residuals.roi}
\title{Prediction at ungauged sites using region of influence and kriging}
\usage{
FitRoi(
  x,
  xnew,
  nk,
  phy,
  similarity,
  kriging = NULL,
  model = "Exp",
  ker = TRUE,
  fitted = FALSE
)

\method{predict}{roi}(object, x, fold = 5, ...)

\method{print}{roi}(x, ...)

\method{residuals}{roi}(object, x, fold = 5, ...)
}
\arguments{
\item{x}{Data for training the model.}

\item{xnew}{Data at new locations (Validation set).}

\item{nk}{Number of sites in the neighborhoods. If \code{nk} is a list, the 
first element represents a vector of the neighborhood size for all sites 
used for training. The second element .}

\item{phy}{Formula defining the physical descriptors.}

\item{similarity}{Formula defining the covariates used
to evaluate the similarity between site, i.e. Euclidean distance 
to the target.}

\item{kriging}{Formula defining the spatial covariates.
Necessary for predicting residuals using spatial correlation.}

\item{model}{Variogram model. See \link[gstat]{vgm}}

\item{ker}{Should a (Epanechnikov) kernel be used in to weight
local regression model. Otherwise uniform weights are used.}

\item{fitted}{Logical. Should the fitted value be returned. Useful to
obtain fitted values at gauged sites from a ROI model when kriging is not
used.}

\item{object}{Output from \code{FitRoi}.}

\item{fold}{Number of group or group used to perform cross-validation.}

\item{...}{Other parameters.}
}
\value{
\item{pred}{Prediction at new sites.}
\item{phy}{Part of the prediction attributed to physical descriptor.}
\item{fitted}{Fitted values (training sites).}

\item{vgm}{Sample variogram.}
\item{model}{Fitted variogram model.}
}
\description{
Return the prediction of a local regression model using region of
influence (ROI) where the residuals are further predicted
by kriging.
}
\examples{

data(flowUngauged)

## Using multidimensional scaling for projecting coordinates
coord <- flowUngauged[,c('lon','lat')]
coord <- cmdscale(GeoDist(coord))
colnames(coord) <- c('lon','lat')

## Transform data if necessary
xdf <- with(flowUngauged, 
         data.frame(y      = l1,
                    area   = scale(log(area)),
                    wb     = scale(log(wb)),
                    stream = scale(log(stream)),
                    map    = scale(log(map)),
                    coord))

 ## select a validation and training set
 set.seed(9382)
 vid <- runif(nrow(xdf)) > .8
 tid <- !vid

 # formula of the relationship between flood quantile and descriptors
 fphy <- log(y) ~ area + map + stream + wb + I(wb^2) + I(stream^2)
 fsimilarity <- ~ area + map

 ## Fit a local regression model
 fit <- FitRoi(x = xdf[tid,], xnew = xdf[vid,], nk = 60,
               phy = fphy, similarity = fsimilarity)
 
 print(fit)
 response <- log(xdf[vid,'y']) 
 sd(response - fit$pred)

 ## Refit the model and perform the kriging of the residuals
 fitk <- FitRoi(x = xdf[tid,], 
               xnew = xdf[vid, ], 
               nk = 60,
               phy = fphy, 
               similarity = fsimilarity , 
               model = 'Exp',
               kriging = ~ lon + lat)
               
 print(fitk)
 sd(response - fitk$phy)
 sd(response - fitk$pred)

}
\references{
Martin Durocher, Donald H. Burn, Shabnam Mostofi Zadeh & Fahim Ashkar (2019) 
  Estimating flood quantiles at ungauged sites using nonparametric regression 
  methods with spatial components, Hydrological Sciences Journal, 64:9, 
  1056-1070, https://doi.org/10.1080/02626667.2019.1620952
}
\seealso{
\code{\link{CvRoi}}, \code{\link[gstat]{krige}}, \code{\link{lm}}
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
