% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_rfa_fitted_nsamax.R
\name{fitted.nsamax}
\alias{fitted.nsamax}
\title{Fitted trend for the nonstationary AMAX model}
\usage{
\method{fitted}{nsamax}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{Output from \link{FitNsAmax}.}

\item{newdata}{Covariates at the fitted dates.}

\item{...}{Other parameters.}
}
\description{
Return the fitted trend of the nonstationary AMAX model at given dates.
}
\seealso{
\link{FitNsAmax}, \link{predict.nsamax}.
}
