% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_rfa_fitted_nspot.R
\name{fitted.nspot}
\alias{fitted.nspot}
\alias{residuals.nspot}
\alias{coef.nspot}
\title{Extract information from a nonstationary POT model}
\usage{
\method{fitted}{nspot}(object, newdata = NULL, ...)

\method{residuals}{nspot}(object, type = "thresh", ...)

\method{coef}{nspot}(object, type = "all", ...)
}
\arguments{
\item{object}{Output form \cite{FitNsPot}}

\item{newdata}{Datasets used to evaluate the trend and the treshold a 
given dates.}

\item{...}{Other parameters.}

\item{type}{Type of output.}
}
\description{
Return the thresholds, trends, residuals and parameters of the model.
}
\details{
The function \code{fitted} return a data.frame containing the original 
observation, date, threshold, trend and standardized values. The latter
correspond the observations after substracting the threshold and dividing by
the mean excess (trend).

For \code{residuals}, the argument \code{type} can be \code{'thresh'} for
returning the value above the threshold or \code{'scale'} for the 
standardized value. 

The function \code{coef} returned some parameters of the fitted model. 
Can be one of \code{'kappa'} for the shape parameter, \code{'trend'}, 
\code{'thresh'} or \code{'all'}.
}
\seealso{
\link{FitNsPot}
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
