% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Floodnet_functions.R
\name{Floodnet_functions}
\alias{Floodnet_functions}
\title{Floodnet functions}
\description{
These functions were written by Martin Durocher as part of the 
FloodNet program \url{https://www.nsercfloodnet.ca/}.
The functions perform flood-frequency analyses, so they are all prefixed
with \code{ch_rfa_}, so that they can be identified.  

Currently, only a few of the functions have been added to \pkg{CSHShydRology}. 
We are adding more as time permits.
\describe{
 \item{ch_rfa_extractamax}{Extracts the annual maxima of a daily time series}
 \item{ch_rfa_distseason}{Distances between points in seasonal space}
 \item{ch_rfa_julianplot}{Empty rose plot by day of year}
 \item{ch_rfa_seasonstat}{Seasonal statistics for flood peaks}
}
}
