% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_raster.R
\name{flow_raster}
\alias{flow_raster}
\title{Raster plot of streamflows}
\usage{
flow_raster(dframe, title = "", rastercolours = c("lightblue", "cyan",
  "blue", "slateblue", "orange", "red"))
}
\arguments{
\item{dframe}{A data frame of hydrometric data. Must contain the variables \code{Date} and \code{Flow}.}

\item{title}{The (optional) title for the plot}

\item{rastercolours}{A vector of colours used for the raster plot. The default is \code{c("lightblue", "cyan", "blue", "slateblue", "orange", "red")}.}
}
\value{
No value is returned; a standard \R graphic is created.
}
\description{
Produces a raster plot: years x day of year, showing magnitude of flows.  This produces a plot showing the flow data in colours, showing different context than in a hydrograph. High flows are in warm colours
}
\examples{
flow_raster(W05AA008)

}
\seealso{
\code{\link{flow_raster_trend}} \code{\link{flow_raster_qa}}
}
\author{
Paul Whitfield <paul.h.whitfield@gmail.com>
}
