% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_raster_qa.R
\name{flow_raster_qa}
\alias{flow_raster_qa}
\title{Raster plot of streamflows with WSC quality flags}
\usage{
flow_raster_qa(dframe, title = "")
}
\arguments{
\item{dframe}{A data frame of WSC hydrometric data. Must contain the variables \code{Date}, \code{Flow} and \code{SYM} which is the WSC data flag symbol.}

\item{title}{The (optional) title for the plot}
}
\value{
No value is returned; a standard \R graphic is created.
}
\description{
Produces a raster plot of years x day of year showing the flow data in grayscale
overlain by the Water Survey of Canada quality flags. Colours are consistent with
ECDataExplorer. Raster layout lets the use see the flags in a different context than in
a hydrograph.
The data flags are:
\describe{
\item{A (Partial)}{green}
\item{B (Below Ice)}{blue}
\item{D (Dry)}{yellow}
\item{E (Estimated)}{red}
}
}
\examples{
flow_raster_qa(W05AA008, "Station W05AA008")
}
\seealso{
\code{\link{flow_raster_trend}} \code{\link{flow_raster}}
}
\author{
Paul Whitfield <paul.h.whitfield@gmail.com>
}
