% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_rfa_data.R
\docType{data}
\name{flowAtlantic}
\alias{flowAtlantic}
\title{Annual maximums from sites in the Atlantic region of Canada}
\format{
An object of class \code{list} of length 2.
}
\source{
\url{https://wateroffice.ec.gc.ca/}
}
\usage{
flowAtlantic
}
\description{
Contains the annual maximums of 45 hydrometric stations found in the
region '01' of Water Survery of Canada.
Additionaly to the annual maximums, the output list includes catchment
descriptors (longitude, latitude, drainage area, mean annual precipitation)
and the geographical distance between each station.
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
\keyword{datasets}
