% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_rfa_data.R
\docType{data}
\name{flowUngauged}
\alias{flowUngauged}
\title{L-moments and catchment descriptors of hydrometric stations in Canada}
\format{
An object of class \code{data.frame} with 562 rows and 11 columns.
}
\source{
\url{https://wateroffice.ec.gc.ca/}
}
\usage{
flowUngauged
}
\description{
The L-moments of the annual maximums at 562 stations
were extracted from Water Survery of Canada (HYDAT).
Catchment descriptors are available for each station was
provided by Environment and Climate Change Canada (ECCC).
The best at-site distributions according to the AIC criterion are included.
}
\details{
\describe{
   \item{site}{Identification number of the station in HYDAT.}
   \item{area}{Drainage area (sq km).}
   \item{map}{Mean annual precipitation for the catchment (mm).}
   \item{wb}{Drainage area covered by waterbodies (pct).}
   \item{stream}{Stream density (km of streams per catchment sq km).}
   \item{lon,lat}{Longitude and latitude of the catchment center.}
   \item{l1, lcv, lsk}{Sample mean, L-coefficient of variation and
         L-coefficient of skewness of the annual maximum discharges.}
   \item{dstr}{Distribution selected for the annual maximum discharge based
         on the Akaike Information Criterion (AIC).}
 }
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
\keyword{datasets}
