% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floodfrequencyanalysis.R
\name{FrequencyAnalysis}
\alias{FrequencyAnalysis}
\title{FrequencyAnalysis: Fit a given extreme value distribution to an extreme value series}
\usage{
FrequencyAnalysis(series, distribution, nep = nonexceeds())
}
\arguments{
\item{series}{A vector representing an extreme value series (e.g., annual maximum flood)}

\item{distribution}{A three-character name of the extreme value distribution (see ?dist.list())}

\item{nep}{A vector of non-exceedance probabilities}
}
\value{
A list object containing: (1) distribution information and (2) output
 (quantile estimates at various non-exceedance probabilities)
 
\url{http://headwateranalytics.weebly.com/blog/flood-frequency-analysis-in-r} contains a demo
}
\description{
FrequencyAnalysis: Fit a given extreme value distribution to an extreme value series
}
