% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_rfa_GeoDist.R
\name{GeoDist}
\alias{GeoDist}
\alias{GeoDist.default}
\alias{GeoDist.formula}
\title{Great-circle distance}
\usage{
GeoDist(x, ...)

\method{GeoDist}{default}(x, earth.radius = 6371, ...)

\method{GeoDist}{formula}(form, x, ...)
}
\arguments{
\item{x}{Dateset of coordinates (lon,lat).}

\item{...}{Other parameters.}

\item{earth.radius}{Radius of the earth. Default 6371 km.}

\item{form}{Formula to identify coordinates in a dataset. 
Must be of the form : \code{lon~lat}.}
}
\description{
Returns the distance matrix of the great-circle distance between all pairs
  of sites.
}
\examples{

attach(flowAtlantic)

GeoDist(lon~lat, flowAtlantic$info)[1:5,1:5]


}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
