% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wscstation.R
\name{get_wscstation}
\alias{get_wscstation}
\title{Reads station information from a data file produced by ECDE
Retrieves station information for an individual Water Survey of Canada site,
adds a text string at position 21 that combines key elements for a title.}
\usage{
get_wscstation(stnID, stn)
}
\arguments{
\item{stnID}{A Water Survey of Canada station number}

\item{stn}{a data frame of station information from ECDataExplorer. The data frame \option{HYDAT_list} is supplied with this package.}
}
\value{
Returns a data frame with 21 variables
\itemize{
\item	 {Station}  {StationID}
\item	 {StationName} {Station Name}
\item	 {HYDStatus} {Active or Discontinued}
\item	 {Prov} {Province}
\item	 {Latitude}{}
\item	 {Longitude}{}
\item	 {DrainageArea} {km2}
\item	 {Years} {# of years with data}
\item	 {From} {Start Year}
\item	 {To} {End Year}
\item	 {Reg.} {Regulated }
\item	 {Flow} {if TRUE/Yes}
\item	 {Level} {if TRUE/Yes}
\item	 {Sed} {if TRUE/Yes}
\item	 {OperSched} {Continuous or Seasonal}
\item	 {RealTime} {if TRUE/Yes}
\item	 {RHBN} {if TRUE/Yes is in the reference hydrologic basin network}
\item	 {Region} {if TRUE/Yes is in the reference hydrologic basin network}
\item	 {Datum} {if TRUE/Yes is in the reference hydrologic basin network}
\item	 {Operator} {if TRUE/Yes is in the reference hydrologic basin network}
\item   {Station_lname} {Added field combines ID,Name,Province and if RHBN an * is added}
}
}
\description{
Reads station information from a data file produced by ECDE
Retrieves station information for an individual Water Survey of Canada site,
adds a text string at position 21 that combines key elements for a title.
}
\examples{
df <- HYDAT_list
s_info <- get_wscstation("05BB001", df)
title <- s_info[21]
print(title)

}
\author{
Paul Whitfield <paul.h.whitfield@gmail.com>
}
