% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_rfa_GofTest.R
\name{GofTest}
\alias{GofTest}
\alias{GofTest.amax}
\alias{GofTest.fpot}
\title{Goodness-of-fit test}
\usage{
GofTest(object, ...)

\method{GofTest}{amax}(object, method = "ad", nsim = 1000, ...)

\method{GofTest}{fpot}(object, method = "adtab", nsim = 1000, ...)
}
\arguments{
\item{object}{Output from \code{\link{FitAmax}} or \code{\link{FitPot}}.}

\item{...}{Other parameters.}

\item{method}{Test to be performed. Either Anderson-Darling \code{ad},
or modified Shapiro-Wilk ('shapiro').}

\item{nsim}{Number of simulations.}
}
\description{
Return the statistic and p-value of a goodness-of-fit test on
AMAX and POT models. The null hypothesis (composite)
is that the data were generated by the fitted distribution.
}
\details{
For a POT model, the method \code{adtab} perform the Anderson-Darling 
test and interpolates the p-value from a table.
}
\section{References}{


 Choulakian, V., Stephens, M.A., 2001. Goodness-of-Fit Tests for the
 Generalized Pareto Distribution. Technometrics 43, 478-484.
 https://doi.org/10.2307/1270819

 Heo, J.-H., Shin, H., Nam, W., Om, J., Jeong, C., 2013. Approximation of
 modified Anderson-Darling test statistics for extreme value distributions
 with unknown shape parameter. Journal of Hydrology 499, 41-49.
 https://doi.org/10.1016/j.jhydrol.2013.06.008

 Ba, I., Ashkar, F., 2017. Discrimination between a group of
 three-parameter distributions for hydro-meteorological frequency modeling.
 Can. J. Civ. Eng. 45, 351-365. https://doi.org/10.1139/cjce-2017-0416
 
 Fahim Ashkar & Ba, I. (2017) Selection between the generalized Pareto
 and kappa distributions in peaks-over-threshold hydrological frequency 
 modelling, Hydrological Sciences Journal, 62:7, 1167-1180,
 http://dx.doi.org/10.1080/02626667.2017.1302089
}

\examples{

## The same simulated distribution should be accepted p-value >.05
x <- rAmax(100, c(100,30,-.1), 'gev')

fit <- FitAmax(x, 'gev', varcov = FALSE)
GofTest(fit, nsim = 50) ## nsim is small for speeding-up the demonstration

## The same apply to POT model
fit <- FitPot(flow~date, flowStJohn, u = 1000,
               declust = 'flood', r = 14)

## By default a table is used with GPA using MLE
GofTest(fit)

}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
