% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_gpa.R
\name{GPA}
\alias{GPA}
\alias{pgpa}
\alias{rgpa}
\alias{dgpa}
\alias{qgpa}
\title{Generalized Pareto distribution (GPA)}
\usage{
pgpa(q, alpha = 1, kap = 0, lower.tail = TRUE)

rgpa(n, alpha = 1, kap = 0)

dgpa(x, alpha = 1, kap = 0, log = FALSE)

qgpa(p, alpha = 1, kap = 0, lower.tail = TRUE)
}
\arguments{
\item{q, x}{Vector of quantiles.}

\item{alpha}{Scale parameter of the GPA}

\item{kap}{Shape parameter of the GPA}

\item{lower.tail}{Logical. Should the propability of the lower tail 
be returned}

\item{n}{Number of simulations.}

\item{log}{Should the log-density be returned}

\item{p}{Vector of probabilities.}
}
\description{
Distribution, density, quantile and random function for the Generalized
pareto distriution.
}
\examples{

kap <- -.2
a <- 1
xd1 <- rgpa(1e4, a, kap)
xd2 <- qgpa(runif(1e4), a, kap)

qqplot(xd1, xd2)


tt <- seq(0.001,6, len = 100)

hist(xd1[xd1<6], main = 'GPA distribution',
  freq = FALSE, ylim =c(0,1), xlim = c(0,6))

lines(tt, dgpa(tt,a,kap))
lines(tt,pgpa(sort(tt), a, kap), col = 2, lty = 2)


}
\references{
Davison, A. C., & Smith, R. L. (1990). Models for Exceedances over High 
Thresholds. Journal of the Royal Statistical Society. Series B 
(Methodological), 52(3), 393-442. http://www.jstor.org/stable/2345667
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
