% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_gumbel.R
\name{scale_x_gumbelRP}
\alias{scale_x_gumbelRP}
\alias{scale_y_gumbelRP}
\alias{scale_x_gumbelAEP}
\alias{scale_y_gumbelAEP}
\title{Gumbel-transformed axes}
\usage{
scale_x_gumbelRP(...)

scale_y_gumbelRP(...)

scale_x_gumbelAEP(...)

scale_y_gumbelAEP(...)
}
\arguments{
\item{...}{Arguments to pass to \code{scale_x_continuous} or
\code{scale_y_continuous} from \pkg{ggplot2}.}
}
\value{
The same output as \code{scale_x_continuous} or
\code{scale_y_continuous}, but with the appropriate Gumbel spacing.
}
\description{
Transforms return period or annual exceedance probability to be spaced
according to a reduced Gumbel variate. Can be added as a scale layer to
a ggplot2 graphic.
}
\examples{
library(ggplot2)
df <- data.frame(
  y = sort(stats::rexp(100), decreasing = TRUE),
  aep = 1:100 / 101
)

ggplot(df, aes(aep, y)) +
  geom_point() +
  scale_x_gumbelAEP()

ggplot(df, aes(1 / aep, y)) +
  geom_point() +
  scale_x_gumbelRP("Return Period")
}
