% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_gumbel.R
\docType{data}
\name{gumbelAEP_trans}
\alias{gumbelAEP_trans}
\title{Gumbel transformations used for ggplot2 scales}
\format{
An object of class \code{transform} of length 9.
}
\usage{
gumbelAEP_trans
}
\description{
Gumbel transformations used for ggplot2 scales
}
\keyword{datasets}
