% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HYDAT_list.R, R/get_wscstation.R
\docType{data}
\name{HYDAT_list}
\alias{HYDAT_list}
\alias{HYDAT_list}
\title{Station information for HYDAT as of 2016-10-16 release of ECDataExplorer}
\format{a dateframe with 7791 rows and 20 columns of metadata}
\source{
2016-10-16 release of ECDataExplorer
\itemize{
\item	 {Station}  {StationID}
\item	 {StationName} {Station Name}
\item	 {HYDStatus} {Active or Discontinued}
\item	 {Prov} {Province}
\item	 {Latitude}{}
\item	 {Longitude}{}
\item	 {DrainageArea} {km2}
\item	 {Years} {# of years with data}
\item	 {From} {Start Year}
\item	 {To} {End Year}
\item	 {Reg.} {Regulated?} 
\item	 {Flow} {if TRUE/Yes}
\item	 {Level} {if TRUE/Yes}
\item	 {Sed} {if TRUE/Yes}
\item	 {OperSched} {Continuous or Seasonal}
\item	 {RealTime} {if TRUE/Yes}
\item	 {RHBN} {if TRUE/Yes is in the reference hydrologic basin network}
\item	 {Region} {if TRUE/Yes is in the reference hydrologic basin network}
\item	 {Datum} {if TRUE/Yes is in the reference hydrologic basin network}
\item	 {Operator} {if TRUE/Yes is in the reference hydrologic basin network}
}
}
\usage{
HYDAT_list

HYDAT_list
}
\arguments{
\item{stnID}{- a Water Survey of Canada station number}

\item{stn}{default is HYDAT_list - station information from ECDataExplorer}
}
\value{
a data frame with 21 elements

\itemize{
\item	 {Station}  {StationID}
\item	 {StationName} {Station Name}
\item	 {HYDStatus} {Active or Discontinued}
\item	 {Prov} {Province}
\item	 {Latitude}{}
\item	 {Longitude}{}
\item	 {DrainageArea} {km2}
\item	 {Years} {# of years with data}
\item	 {From} {Start Year}
\item	 {To} {End Year}
\item	 {Reg.} {Regulated }
\item	 {Flow} {if TRUE/Yes}
\item	 {Level} {if TRUE/Yes}
\item	 {Sed} {if TRUE/Yes}
\item	 {OperSched} {Continuous or Seasonal}
\item	 {RealTime} {if TRUE/Yes}
\item	 {RHBN} {if TRUE/Yes is in the reference hydrologic basin network}
\item	 {Region} {if TRUE/Yes is in the reference hydrologic basin network}
\item	 {Datum} {if TRUE/Yes is in the reference hydrologic basin network}
\item	 {Operator} {if TRUE/Yes is in the reference hydrologic basin network}
\item   {Station_lname} {Added field combines ID,Name,Province and if RHBN an * is added}
}
}
\description{
The file contains 20 fields of metadata for each station from HYDAT

Retrieves station information for Water Survey of Canada sites
}
\examples{
\dontrun{s_info <- get_wscstation("05BB001")}



}
\author{
Paul Whitfield <paul.h.whitfield@gmail.com>
}
\keyword{datasets}
