% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrograph.plot.R
\name{hydrograph.plot}
\alias{hydrograph.plot}
\title{Plot Hydrograph}
\usage{
hydrograph.plot(flows = NULL, precip = NULL, prd = NULL,
  flow.labels = NULL, precip.label = NULL, winter.shading = T,
  range.mult.flow = NULL, range.mult.precip = 1.5, ylabel = "Flow [m3/s]",
  leg.pos = NULL, leg.box = NULL, zero.axis = T, plot.mode = "base")
}
\arguments{
\item{flows}{data frame of flows (with Date attribute)}

\item{precip}{data frame with precipitation data (with Date attribute)}

\item{prd}{period to use in plotting ("YYYY-MM-DD/YYYY-MM-DD")}

\item{flow.labels}{alternative labels to use for flow colnames}

\item{precip.label}{alternative label for precip data}

\item{winter.shading}{adds shading for winter months (Dec 1 - Mar 31)}

\item{range.mult.flow}{range multiplier for max value in flows}

\item{range.mult.precip}{range multiplier for max value in precipitation}

\item{ylabel}{text label for y-axis of the plot (default 'Flow [m3/s]')}

\item{leg.pos}{string specifying legend placement on plot}

\item{leg.box}{boolean on whether to put legend in an opaque box}

\item{zero.axis}{fixes the y axis to start exactly at zero (default TRUE)}

\item{plot.mode}{plot mode as "base" or "ggplot"}
}
\value{
\item{TRUE}{return TRUE if the function is executed properly}
}
\description{
hydrograph.plot creates a hydrograph plot for provided flows, including 
precipitation if provided.
}
\details{
flows is a data frame with the $Date attribute as the first in the data frame, 
where $Date is a vector of date class values, and any number of flow data attributes 
that also exist in the data frame to be included in the plot (currently up to 10 flow series 
can be plotted).

precip is a data frame with the $Date attribute as the first in the data frame,
where $Date is a vector of date class values, and a second column of precipitation 
values.

flow.labels overwrites the flows data frame attribute names, which are shown 
in the legend instead.

precip.label overwrites the precip data frame attribute name, and is shown in the 
legend if provided instead of the existing attribute name.

winter.shading argument will add a transparent cyan shading for the
December 1st to March 31st period in each year that is plotted.

range.mult.flow and range.mult.precip arguments will increase the maximum 
value that is plotted in the flows and the precip values. This is useful in 
preventing overlap if precip is also plotted. The range.mult.flow value is 
calculated as 1.05 if precip is NULL or 1.5 if precip is not NULL, unless 
the value is provided.

ylabel is the label on the y axis, defined using y.lab in the plot function.
This defaults to 'Flow [m3/s]' intended for plotting hydrographs.

leg.pos is the position for the legend to be placed, e.g. 'topleft',
'right', etc., and is consistent with the legend function options. If this
is left null, the function will place it either topleft or left, depending
on whether precip is also plotted (i.e. left if precip added, topleft
otherwise).

leg.box is a boolean for whether to put the legend in an opaque, white box
or not. If left as NULL, the function will automatically not use a white box
and leave the background of the legend transparent.

zero.axis can be used to set the min value of the y axis (or axes if precip
also plotted) to zero. Note that by default, R will plot the values with a
slight buffer for presentation. A warning that if this option is set to
TRUE, the minimum value is set to zero without checking if any flow values
are less than zero. This option should not be used for stage plotting, since
most reservoirs are not in the range of 'zero' stage for normal operations,
since stage is reported to elevation and not to stage bottom, typically.

plot.mode is used to indicate the plot type to output. Currently only 'base' 
plot type is available, 'ggplot' mode is under construction.

Note that a plot title is purposely omitted in order to allow the automatic
generation of plot titles.
}
\examples{

# example with randomly sampled data
dd <- seq.Date(as.Date("2010-10-01"),as.Date("2013-09-30"),by=1)
x <- abs(rnorm(length(dd)))
y <- abs(rnorm(length(dd)))*x
df <- data.frame("Date"=dd,x,y)
myprd <- "2011-10-01/2012-09-30"

precip <- data.frame("Date"=dd,"precip"=abs(rnorm(length(dd)))*10)

# basic hydrograph plot
hydrograph.plot(flows=df,winter.shading=F) 

with different labels
hydrograph.plot(flows=df,winter.shading=F,flow.labels=c("simulated","observed")) 

# with a few more options turned on
hydrograph.plot(flows=df,precip=precip)

# increase the plot ranges to separate flows and precip; add a legend box
hydrograph.plot(flows=df,precip=precip,range.mult.flow=1.7,range.mult.precip=2,leg.box=T)

}
\author{
Robert Chlumsky <rchlumsk@gmail.com>
}
\keyword{hydrograph}
\keyword{plot}
