% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydrographPlot.R
\name{HydrographPlot}
\alias{HydrographPlot}
\title{Plot Hydrograph}
\usage{
HydrographPlot(flows = NULL, precip = NULL, prd = NULL,
  winter.shading = F, range.mult.flow = NULL, range.mult.precip = 1.5,
  flow.labels = NULL, ylabel = "Flow [m3/s]",
  precip.label = "Precipitation [mm]", leg.pos = NULL, leg.box = NULL,
  zero.axis = T, plot.mode = "base")
}
\arguments{
\item{flows}{data frame of flows to plot}

\item{precip}{data frame of precipitation values to plot}

\item{prd}{period to use in plotting}

\item{winter.shading}{optionally adds shading for winter months (default
FALSE)}

\item{range.mult.flow}{range multiplier for max value in hydrograph}

\item{range.mult.precip}{range multiplier for max value in precipitation plot (default 1.5)}

\item{flow.labels}{string vector of labels for flow values}

\item{ylabel}{text label for y-axis of the plot (default 'Flow [m3/s]')}

\item{precip.label}{text label for precipitation y-axis (default 'Precipitation [mm]')}

\item{leg.pos}{string specifying legend placement on plot}

\item{leg.box}{boolean on whether to put legend in an opaque box}

\item{zero.axis}{fixes the y axis to start exactly at zero (default TRUE)}

\item{plot.mode}{plot mode as "base" or "ggplot"}
}
\value{
\item{TRUE}{return TRUE if the function is executed properly}
}
\description{
HydrographPlot creates a hydrograph plot for simulated, observed, and inflow 
hydrograph series, including precipitation if provided.
}
\examples{


# example with randomly sampled data
dd <- seq.Date(as.Date("2010-10-01"),as.Date("2013-09-30"),by=1)
x <- abs(rnorm(length(dd)))
y <- abs(rnorm(length(dd)))*x
df <- data.frame("Date"=dd,x,y)
myprd <- "2011-10-01/2012-09-30"

precip <- data.frame("Date"=dd,"precip"=abs(rnorm(length(dd)))*10)

# basic hydrograph plot
HydrographPlot(flows=df,winter.shading=F) 

with different labels
HydrographPlot(flows=df,winter.shading=F,flow.labels=c("simulated","observed")) 

# with a few more options turned on
HydrographPlot(flows=df,precip=precip)

# increase the plot ranges to separate flows and precip; add a legend box
HydrographPlot(flows=df,precip=precip,range.mult.flow=1.7,range.mult.precip=2,leg.box=T)

 
}
\author{
Robert Chlumsky <rchlumsk@gmail.com>

This function creates a hydrograph plot using the supplied time series; any
series not supplied will not be plotted. If the precip time series is
supplied, the secondary y axis will be used to plot the precip time series.

The function assumes that the supplied time series have the same length and
duration in time. If this is not true, then the defined period or period
calculated from the first available flow series will be used to determine
the plotting limits in time. If the data is used directly from Raven output,
this is not a concern. The supplied time series should be in xts format,
which again can be obtained directly by using the hyd.extract function.

The winter.shading argument will add a transparent cyan shading for the
December 1st to March 31st period in each year that is plotted.

The range.mult argument will increase the maximum value that is plotted in
the flows and the precip values. This is useful in preventing overlap if
precip is also plotted (i.e. with precip as well, range.mult=1.5 works
well). This value should not be less than 1.0, otherwise the values will be
cutoff in the plot.

ylabel is the label on the y axis, defined using y.lab in the plot function.
This defaults to 'Flow [m3/s]' intended for plotting hydrographs.

leg.pos is the position for the legend to be placed, e.g. 'topleft',
'right', etc., and is consistent with the legend function options. If this
is left null, the function will place it either topleft or left, depending
on whether precip is also plotted (i.e. left if precip added, topleft
otherwise).

leg.box is a boolean for whether to put the legend in an opaque. white box
or not. If left as NULL, the function will automatically not use a white box
and leave the background of the legend transparent.

zero.axis can be used to set the min value of the y axis (or axes if precip
also plotted) to zero. Note that by default, R will plot the values with a
slight buffer for presentation. A warning that if this option is set to
TRUE, the minimum value is set to zero without checking if any flow values
are less than zero. This option should not be used for stage plotting, since
most reservoirs are not in the range of 'zero' stage for normal operations,
since stage is reported to elevation and not to stage bottom, typically.

plot.mode is used to indicate the plot type to output. Currently only 'base' 
plot type is available, 'ggplot' is under construction.

Note that a plot title is purposely omitted in order to allow the automatic
generation of plot titles.
}
\keyword{hydrograph}
\keyword{plot}
