% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydRology-package.R
\docType{package}
\name{hydRology-package}
\alias{hydRology-package}
\title{Functions for Canadian hydrological analysis}
\references{
To cite \pkg{hydRology} in publications, use the command \code{citation("hydRology")} to get the current version of the citation.\cr
}
