% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_rfa_Intersite.R, R/ch_rfa_print_isite.R
\name{Intersite}
\alias{Intersite}
\alias{print.isite}
\title{Estimating intersite correlation}
\usage{
Intersite(
  x,
  type = "emp",
  method = "spearman",
  distance = NULL,
  nmin = 5,
  na.sub = "avg",
  defpos = TRUE,
  smooth = 1,
  distance.max = Inf,
  distance.bin = 15,
  start = NULL
)

\method{print}{isite}(x, ...)
}
\arguments{
\item{x}{Dataset in the wide format}

\item{type}{Model type. Can be either \code{'emp'} for estimating
the empirical correlation matrix or \code{'exp'} 
for fitting power exponential model.}

\item{method}{Measure of association used to evaluate dependence. 
Either \code{'spearman'} or \code{'kendall'}.}

\item{distance}{Matrix of distances. Necessary when using \code{'exp'}.}

\item{nmin}{Minimal number of pairs necessary to compute pairwise
correlations.}

\item{na.sub}{Value for imputing missing pairwise correlation with
\code{'emp'}.}

\item{defpos}{Logical. Should a correction be applied to the empirical
matrix to ensure positive definiteness.}

\item{smooth}{Smooth parameter of the power exponential model.}

\item{distance.max}{Maximal distance to consider for paired observations
in the fitting of the power exponential model.}

\item{distance.bin}{Number of equally space bins used in the fitting of the
of the power exponential model.}

\item{start}{Initial parameter for optimization with the \code{exp} method.
It must have the form \code{(nugget, range)}.}

\item{...}{Other parameters.}
}
\value{
\describe{
  \item{method}{Estimation method.}
  \item{para}{Parameter of the fitted model. For \code{'emp'} it
              is the average.}

  \item{corr}{Raw estimation of the correlations.}
  \item{model}{Final estimate of the correlations.}
  \item{rmse}{For \code{'exp'}, root mean square errors of the fitted model.}
}
}
\description{
Return a matrix of intersite correlation between paired observations.
By default the empirical matrix is estimated and corrected to be 
positive definite. 
Can also be estimated by a power exponential model where the weighted least 
squares approach is used with weights proportional to record lengths.
}
\section{References}{


Durocher, M., Burn, D.H., Mostofi Zadeh, S., 2018. A nationwide
 regional flood frequency analysis at ungauged sites using ROI/GLS with
 copulas and super regions. Journal of Hydrology 567, 191-202.
 https://doi.org/10.1016/j.jhydrol.2018.10.011
}

\examples{
\dontrun{
data(flowAtlantic)

## Organize annual maximums
ams <- flowAtlantic$ams
ams$year <- format(ams$date, '\%Y')
xmat <- DataWide(ams ~ id + year, ams)

## Compute distance between sites
coord <- flowAtlantic$info[, c('lon','lat')]
rownames(coord) <- flowAtlantic$info$id
h <- GeoDist(coord)

## make sure that columns match
h <- h[colnames(xmat), colnames(xmat)]

## estimate intersite correlation using a model
isite <- Intersite(xmat, distance = h, type = 'exp')
print(isite)

## Evaluate emprirical correlation matrix
isite <- Intersite(xmat[,1:5])
print(isite)
round(isite$model,2)
}

}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
