% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_plot_amax.R
\name{plot.amax}
\alias{plot.amax}
\title{Return level plot}
\usage{
\method{plot}{amax}(obj, main = "Return level plot",
  xlab = "Return period (year)", ylab = "Flood quantiles",
  ci = FALSE, col.ci = "red", lty.ci = 2, lwd.ci = 1, ...)
}
\arguments{
\item{obj}{Output from \link{FitAmax}.}

\item{ci}{Logical. Should confidence intervals be displayed. 
See \link{predict.amax} with argument Delta method.}

\item{col.ci, lty.ci, lwd.ci}{Graphical parameters defining the display
of the confidence interval.}

\item{...}{Other graphical parameters. See \code{\link{par}}.}
}
\description{
Create a plot of the estimated versus sample flood quantiles, where the 
x-axis is expressed in terms of return period.
}
\examples{

data(flowStJohn)

x <- ExtractAmax(flow~date, flowStJohn)$flow

fit <- FitAmax(x, distr = 'gev', method = 'mle')

plot(fit, ci = TRUE)
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
