% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_plot_isite.R
\name{plot.isite}
\alias{plot.isite}
\title{Plot of the pairwise intersite correlation.}
\usage{
\method{plot}{isite}(x, xmat, distance, xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{x}{Output from \code{Intersite}.}

\item{xmat}{Dataset in wide format.}

\item{distance}{Matrix of distances.}

\item{xlab, ylab}{Label of the xy-axis.}

\item{...}{Other parameters.}
}
\description{
Produce a graphic of the pairwise intersite correlation coefficient with 
respect to the distance.
}
\examples{

#' Data(flowAtlantic)

## Organize annual maximums
ams <- flowAtlantic$ams
ams$year <- format(ams$date, '\%Y')
xmat <- DataWide(ams ~ id + year, ams)

## Compute distance between sites
coord <- flowAtlantic$info[, c('lon','lat')]
rownames(coord) <- flowAtlantic$info$id
h <- GeoDist(coord)

## make sure that columns match
h <- h[colnames(xmat), colnames(xmat)]

## estimate intersite correlation using a model
isite <- Intersite(xmat, distance = h, method = 'exp')
print(isite)
plot(isite, xmat, h)

}
\seealso{
\link{Intersite}, \link{DataWide}.
}
