% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_plot_nsamax.R
\name{plot.nsamax}
\alias{plot.nsamax}
\title{PLotting a nonstationary AMAX model}
\usage{
\method{plot}{nsamax}(x, ...)
}
\arguments{
\item{x}{Output from \link{FitNsAmax}.}

\item{...}{Other parameters.}
}
\description{
Create a graph of the fitted vs residuals.
}
\seealso{
\link{FitNsAmax}.
}
