% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_plot_nspot.R
\name{plot.nspot}
\alias{plot.nspot}
\title{Plotting the the result of the nonstationary POT model}
\usage{
\method{plot}{nspot}(
  x,
  type = "l",
  line.col = c("red", "blue"),
  line.lty = c(1, 1),
  line.lwd = c(2, 2),
  peak.col = "black",
  peak.pch = 1,
  peak.cex = 0.7,
  do.legend = TRUE,
  ...
)
}
\arguments{
\item{x}{Output from}

\item{type}{Type of display for the observations. See \link{par}.}

\item{line.col, line.lty, line.lwd}{Graphical parameter for the threshold and 
the trend.}

\item{peak.col, peak.pch, peak.cex}{Graphical parameters for the extrated peaks.}

\item{do.legend}{Logical. Should a legend be displayed.}

\item{...}{Other parameters.}
}
\description{
Produce a graphics of the time series with time-varying threshold and 
mean excess.
}
\examples{

fit <- FitNsPot(flow~date, x = flowStJohn, tau = .95,
                trend = ~ poly(date,3), thresh = ~ date, 
                declust = 'wrc', r = 14)
                
                
plot(fit, do.legend = FALSE)    
                             
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
