% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_PlotThresh.R
\name{PlotThresh}
\alias{PlotThresh}
\title{Visual diagnostic for peaks over threshold}
\usage{
PlotThresh(obj, type = "ad", ppy = NULL, ...)
}
\arguments{
\item{obj}{Output from \link{SearchThresh}.}

\item{type}{Type of plot to display. 
Must be one or more of \code{'ad','mrl','alpha', 'kappa'}.}

\item{ppy}{Range of peaks per year to display.}

\item{...}{Other parameters.}
}
\description{
Create various graphics to assess the selection of a threshold.
Includes a plot with respect oh the threshold for the p-values for the 
Anderson-Darling test, estimated parameter and flood quantiles.
Vertical lines are shown for thresholds associated to 1 and 1.5 peaks per 
years.
}
\examples{

lstu <- seq(500,2000,20)
out <- SearchThresh(flow~date, flowStJohn, u = lstu, declust = 'wrc', r = 14)

PlotThresh(out)

PlotThresh(out, type = 'q50')

PlotThresh(out, type = paste0('q',c(2,5,10,20,50,100)))

par(mfrow = c(2,2))
PlotThresh(out, type = c('ad','kappa', 'mrl','q10'))

}
\seealso{
\link{SearchThresh}, \link{FindThresh}.
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
