% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_PoolRemove.R
\name{PoolRemove}
\alias{PoolRemove}
\title{Remove heterogenous sites from a pooling group}
\usage{
PoolRemove(
  obj,
  method = "H1",
  tol = 2,
  nmin = 15,
  ntot.min = 0,
  nsim = 1000,
  distr.fix = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{obj}{An output from \link{FitRegLmom}.}

\item{method, tol}{Which heterogenity statistics used in the procedure.
The choices are \code{'H1'}, \code{'H2'} and \code{'H3'}. 
The algorithm stop when the heterogenity goes below the value \code{tol}.}

\item{nmin, ntot.min}{Additional stopping criteria. Respectively the
minimal number of sites or station-year to be included in the region.}

\item{nsim}{Number of simulations used to evaluate the heterogenous statistic}

\item{distr.fix}{Logical, should the selection of the distribution be
re-evaluated after removing the site.}

\item{verbose}{Logical. Should a trace of the removed sites be display.}
}
\description{
Return a regional model (\code{'FitRegLmom'}) where heterogenous sites
are removed in a stepwise manner.
At each step, the removed site that best improves the homogeneity statistic
is permanently removed until a stopping criterion.
The first site that was passed in a dataset is considered as the target and
it cannot be removed.
}
\examples{

data(flowAtlantic)

## Organize data
ams <- flowAtlantic$ams
ams$year <- format(ams$date, '\%Y')
xmat <- DataWide(ams ~ id + year, ams)

## Compute distance
coord <- flowAtlantic$info[,c('lon','lat')]
rownames(coord) <- flowAtlantic$info$id
h <- GeoDist(coord)

## Fit a index flood model inside a pooling group
xmat0 <- FindNearest(xmat, h[1, colnames(xmat)], 25)
fit <- FitRegLmom(xmat)

## Remove site until homogeneity is reached
fit.new <- PoolRemove(fit, tol = 2)

## Get the data from the pooling group
xmat1 <- xmat[,sitenames(fit.new)] 


}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
