% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_predict_amax.R
\name{predict.amax}
\alias{predict.amax}
\title{Predict return levels}
\usage{
\method{predict}{amax}(object, p = c(0.5, 0.8, 0.9, 0.95, 0.98, 0.99),
  se = FALSE, ci = "none", alpha = 0.05, nsim = 1000,
  out.matrix = FALSE, ...)
}
\arguments{
\item{object}{Output from  \code{\link{FitAmax}}}

\item{p}{Probabilities associated to the return level. For example,
a 100 years return period is equivalent to \code{p = 0.99}.}

\item{se}{Return the standard deviation of the return level using
the delta method. The fitted model must}

\item{ci}{Method to compute the confident interval. One of \code{'delta'}
for the delta method, \code{'boot'} for parametric boostrap and \code{'norm'}
for Monte-Carlo approximation assuming normality of the parameters.}

\item{alpha}{Probability outside the confident interval.}

\item{nsim}{Number of simulation use for resampling.}

\item{out.matrix}{Logical. Should the resampling be returned. If true,
a list is returned containing the prediction table (\code{pred}),
the parameters (\code{para}) and the return levels (\code{qua}).}

\item{...}{Other parameters.}
}
\description{
Return the flood quantile of annual maximum distribution and
Confident intervals are provided by bootstrap.
}
\examples{

#' ## Extract an time series of annual maxima
x <- ExtractAmax(flow~date, flowStJohn)$flow

## Fitting of GEV distribution using L-moments
fit <- FitAmax(x,'gev', method = 'mle')

## Get the estimated quantile of 10 and 100 years return period
rp <- 1-1/c(10,100)
predict(fit, rp)
predict(fit, se = TRUE, ci = 'delta')

## The bootstrap sample used for CI are returned
fit <- FitAmax(x,'gev', varcov = FALSE)
boot <- predict(fit, rp, se = FALSE, ci = 'boot',
                nsim = 500, out.matrix = TRUE)

}
