% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_predict_fpot.R
\name{predict.fpot}
\alias{predict.fpot}
\title{Prediction of flood quantiles of a given return period.}
\usage{
\method{predict}{fpot}(object, rt = c(2, 5, 10, 20, 50, 100),
  se = FALSE, ci = "none", alpha = 0.05, nsim = 1000,
  out.matrix = FALSE, ...)
}
\arguments{
\item{object}{Output of \link{FitPot}.}

\item{rt}{Return period.}

\item{se}{Logical. Should the standard deviation be returned.}

\item{ci, alpha}{Method for evaluation the confidence intervals with 
probability \code{1-alpha}. Available methods are : 
Delta method (\code{'delta'}), profile likelihood (\code{'profile'}) 
and nonparametric bootstrap (\code{'boot'})}

\item{nsim}{Number of bootstrap sample.}

\item{out.matrix}{Logical. Should the bootstrap sample be returned?}

\item{...}{Other parameters.}
}
\description{
Return a vector or matrix of the flood quantiles, its standard deviation and 
confidence intervals.
}
\examples{

data(flowStJohn)

fit <- FitPot(flow~date, flowStJohn, u = 1000,
               declust = 'wrc', r = 14)
     
predict(fit, se = TRUE, ci = 'delta')

}
\references{
Coles S. (2001) An introduction to statistical modeling of extreme values.
  Springer Verlag.
}
\seealso{
\link{FitPot}
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
