% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_predict_nsamax_mle.R
\name{predict.nsamax.mle}
\alias{predict.nsamax.mle}
\alias{BootNsAmaxMle}
\title{Prediction of flood quantiles for the nonstationary model using MLE}
\usage{
\method{predict}{nsamax.mle}(object, p = c(0.5, 0.8, 0.9, 0.95, 0.98,
  0.99), newdata = NULL, type = "quantile", ...)

BootNsAmaxMle(object, p = c(0.5, 0.8, 0.9, 0.95, 0.98, 0.99),
  newdata = NULL, reliability = FALSE, nsim = 1000, alpha = 0.95,
  verbose = TRUE, ...)
}
\arguments{
\item{object}{Output of \link{FitNsAmaxMle}.}

\item{p}{Exceeding probability associated to return period or reliability level.
The reliability is \code{p^s} where \code{s} is the size of \code{newdata}.}

\item{newdata}{Covariates. If \code{NULL} the input data that served to fit 
the model is used}

\item{type}{Type of output. Must be one of 
\code{'location','quantile', 'reliability'}.}

\item{...}{Other parameters.}

\item{reliability}{Logical. Should the flood quantile associated with 
reliability level be returned. See \link{FitNsAmax}.}

\item{nsim}{Sample size.}

\item{alpha}{Confidence level.}

\item{verbose}{Logical. Should a progress bar be displayed.}
}
\description{
Return the flood quantiles associatied to specifc level of 
exceeding probability or reliability level.
The function \code{BootNsAmaxMle} is used to obtain bootstrap samples of 
the parameters and flood quantiles.
}
\examples{

data(flowStJohn)

amax <- ExtractAmax(flow~date, flowStJohn, tol = 365)

## Recenter year to start at 1
amax$year <- as.integer(amax$yy) - 1926

## Add a trend
amax$flow <- amax$flow + amax$year * 10

fit <- FitNsAmaxMle(flow~year, amax, 'gno', type = 'add')

## Last 30 years
ref.period <- 59:88

## trend ##
loc <- predict(fit, c(.5, .9), type = 'loc')
hat <- predict(fit, c(.5, .9))
rel <- predict(fit, c(.5, .9), newdata = amax[ref.period,], 
               type = 'rel')

## Data cloud
plot(flow~year, amax)

## location parameter
lines(loc, col = 'red')

## Flood quantiles
for(ii in 1:2) lines(hat[,ii], col = 'red', lty = 2)
for(ii in 1:2) lines(cbind(ref.period, rel[ii]), lwd = 3, col = 'blue')

## Using boostrap to evaluate the model uncertainty
bs <- BootNsAmaxMle(fit, c(.5, .9), nsim = 50, reliability = TRUE,
                 method = 'Nelder-Mead', control = list(maxit = 2000))

out <- summary(bs)
print(out)

}
\seealso{
\link{FitNsAmax}, \link{predict.nsamax.mle}.
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
