% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_rfa_predict_nsamax.R
\name{predict.nsamax}
\alias{predict.nsamax}
\title{Prediction of the nonstationary model for AMAX using regression.}
\usage{
\method{predict}{nsamax}(
  object,
  p = c(0.5, 0.8, 0.9, 0.95, 0.98, 0.99),
  newdata = NULL,
  reliability = FALSE,
  ...
)
}
\arguments{
\item{object}{Output form \link{FitNsAmax}.}

\item{p}{Exceeding probability associated to a quantile or reliability level.}

\item{newdata}{Covariates. If \code{NULL} the input data that served to fit 
the model is used}

\item{reliability}{Logical. Should the flood quantile associated with 
reliability level be returned. In that case, \code{newdata} must represent
the period of interest. See Durocher et al. (2019) for more details.}

\item{...}{Other parameters.}
}
\description{
Return the flood quantiles associatied to specifc level of 
exceeding probability or reliability level. 
The function \code{ch_rfa_boot_nsamax} is used to obtain bootstrap samples of 
the parameters and flood quantiles.
}
\details{
The reliability is defined as the probability that there is no event that 
will exceed as given design level for a specific period. 
When \code{reliability = TRUE}, the elements \code{newdata} are used to define
that period. For instance the last 30 years can be obtained by passing the
the last 30 records of the fitted dataset. The reliability level is control 
by \code{p} and is \code{p^s} where \code{s} is the size of \code{newdata}.
}
\examples{

data(flowStJohn)

amax <- ExtractAmax(flow~date, flowStJohn, tol = 365)

## Recenter year to start at 1
amax$year <- as.integer(amax$yy) - 1926

fit <- FitNsAmax(flow~year, amax, 'gev', type = 'mult')

## Last 30 years
ref.period <- 59:88

hat <- predict(fit, c(.5, .9))
rel <- predict(fit, c(.5, .9), newdata = amax[ref.period,], reliability = TRUE)

## Data cloud
plot(flow~year, amax)

## Time-depedent return level
for(ii in 1:2) lines(hat[,ii], col = 'red')

## Reliability levels for last 30 years
for(ii in 1:2) lines(cbind(ref.period, rel[ii]), lwd = 3, col = 'blue')

## Using boostrap to evaluate the model uncertainty
bs <- ch_rfa_boot_nsamax(fit, c(.5, .9), nsim = 50, 
                 newdata = amax[ref.period,], reliability = TRUE)
summary(bs)

## Return a single simulation of the model
simulate(fit)

}
\references{
Durocher, M., Burn, D. H., & Ashkar, F. (2019). Comparison of estimation 
  methods for a nonstationary index-flood model in flood frequency analysis 
  using peaks over threshold [Preprint]. 
  https://doi.org/10.31223/osf.io/rnepc
}
\seealso{
\link{FitNsAmaxMle}, \link{FitNsPot}, \link{predict.nsamax}.
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
