% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_predict_poolmle.R
\name{predict.poolmle}
\alias{predict.poolmle}
\title{Prediction of the flood quantile for regional model}
\usage{
\method{predict}{poolmle}(object, p = c(0.5, 0.8, 0.9, 0.95, 0.98, 0.99), index = NULL, ...)
}
\arguments{
\item{object}{Output from \link{FitPoolMle}.}

\item{p}{Probabilities.}

\item{index}{Site specific parameters.}

\item{...}{Other parameters.}
}
\description{
Return the flood quantiles of given probabilities.
}
\examples{

xd <- flowAtlantic$ams 
## transpose by year
xd$year <- format(xd$date, '\%Y')
xw <- DataWide(ams ~ id + year, xd)

## Fitting a regional shape parameter
fit.gev <- FitPoolMle(xw, 'gev', type = 'mean')

## A quick illustration on how to do bootstrap for inference
boot <- simulate(fit.gev, nsim = 30)
fboot <- lapply(boot, function(z) FitPoolMle(z, 'gev', type = 'mean') ) 
pboot <- sapply(fboot, getElement, 'para')
summary(pboot)

## Inference for flood quantiles at target site
Fz <- function(z) predict(z, c(.9,.99), index = z$index[1])
qboot <- sapply(fboot, Fz)
summary(t(qboot))


}
