% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_predict_reglmom.R
\name{predict.reglmom}
\alias{predict.reglmom}
\title{Flood quantiles estimates}
\usage{
\method{predict}{reglmom}(
  object,
  q = c(0.5, 0.8, 0.9, 0.95, 0.98, 0.99),
  ci = FALSE,
  corr = 0,
  nsim = 1000,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{object}{An output from \link{FitRegLmom}.}

\item{q}{Probability associated to the flood quantiles.}

\item{ci}{Logical. Should the confident intervals and the standard deviation
be evaluated?}

\item{corr}{Intersite correlation. Either a matrix or a constant coefficient
for all pairs}

\item{nsim}{Number of simulations used for approximating
the confident intervals.}

\item{alpha}{Probability outside the confidence intervals.}

\item{...}{Other parameters.}
}
\description{
Predict the flood quantile of index-flood model for a specific scale
factor. By default the flood quantile of the first site (target) is returned.
}
\examples{

data(flowAtlantic)

h <- GeoDist(flowAtlantic$info[,c('lon','lat')])
ams <- flowAtlantic$ams
ams$year <- format(ams$date, '\%Y')
xd <- DataWide(ams ~ id + year, ams)

xd <- FindNearest(xd, h[1,], 25)
h  <- FindNearest(h , h[1,], 25, row = TRUE)

## Fit the regional model
fit <- FitRegLmom(xd)

## estimate flood quantiles
predict(fit, c(.3,.7))

## Evaluate intersite correlation
isite <- Intersite(xd)

predict(fit, ci = TRUE, corr = isite$model)

}
\references{
Hosking, J. R. M., & Wallis, J. R. (1997). Regional frequency analysis:
  an approach based on L-moments. Cambridge Univ Pr.
}
\seealso{
\link{FitRegLmom}, \link{Intersite}.
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
