% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_qa.R
\name{raster_qa}
\alias{raster_qa}
\title{Raster plot with WSC quality flags}
\usage{
raster_qa(dframe)
}
\arguments{
\item{dfile}{- a data frane from WSC. Read in using \code{read_wsc}.}
}
\description{
{
Produces a raster plot: years against day of year, showing all the data flags
A (Partial) green
B (Below Ice)  blue
D (Dry)  yellow
E (Estimated) red
}
}
\examples{
\dontrun{qaplot <-raster_qa(05AA008_Daily_flow_ts)}
}
\author{
Paul Whitfield <paul.h.whitfield@gmail.com>
Produces a gray scale plot of observations with colour symbol showing where data has a quality flag
}
