% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_trend.R
\name{raster_trend}
\alias{raster_trend}
\title{Raster plot of WSC observations}
\usage{
raster_trend(Year, doy, Flow, step, title, missing = FALSE,
  rastercolours = c("lightblue", "cyan", "blue", "slateblue", "darkblue",
  "red"))
}
\arguments{
\item{Year}{- array of length n containing the Year as numeric}

\item{doy}{- array of length n containing the numeric day of year}

\item{Flow}{- array of length n containing the flow data}

\item{step}{- a number indicating the degree of smoothing eg. 1, 5, 11.}

\item{title}{- uses format of 05BB001 - Bow River at Banff AB* from get_WSCstation}

\item{missing}{- If FALSE years with missing data are excluded. If TRUE partial years
are included.}

\item{rastercolours}{Default is c("lightblue","cyan", "blue", "slateblue", "darkblue", "red"). 
Can be changed to another series as desired.}
}
\value{
a list containing:
	\code{stationID}  Station ID eg 05BB001
	\code{missing} How missing values were used FALSE=used ,TRUE=removed
\code{step} # of days in a bin
\code{periods} # of periods in a year
\code{period} period numbers i.e. 1:365/step
\code{bins} values for each period in each year
\code{med_period} median for each period
\code{max_period} maximum for each period
\code{min_period} minimum for each period
\code{tau_period} Kendall's Tau for each period
\code{prob_period} pr0bability of Tau for each period
\code{year} years spanning the data
\code{median_year} median bin for each year
\code{max_year} maximum bin for each year
\code{min_year} minimum bin for each year
\code{tau_median_year} value of tau and probability for median_year
\code{tau_maximum_year} value of tau and probability for max_year
\code{tau_minimum_year} value of tau and probability for min_year
}
\description{
Creates a raster plot plus trend plots for day of year, and over time
which may be binned by a number of days
}
\details{
Plot contains four panels based upon binned data. 
1. the annual maximum,minimum,and median with a trend test #' 
for each period: red arrows indicate decreases, blue arrows indicate increases.
2. scale bar for the colours used in the raster plot, 
3. the rasterplot with a colour for each 
period and each year where data exist, and 
4. a time series plot of the minimum, median, and maximum annual bin values.
If there is no trend the points are black, decreasing is red, 
and increasing in blue. 
p<=0.05.
}
\examples{
	\dontrun{mdata <- read_wsc("05AA008_Daily_Flow_ts.csv")  ## read a water survey data file
station <- as.character(mdata[1,1])		## find the station ID
sname<- get_wscstation(station)		## get the station information
title=sname$Station_lname			## set the title to the long form of the name
date <-as.Date(mdata$Date,"\%Y/\%m/\%d")
Year <- as.numeric(format(date,"\%Y"))
doy <- as.numeric(dayOfYear(as.timeDate(date)))
flow <- mdata$Flow
step=5}

 \dontrun{mplot <- raster_trend(Year, doy, flow, step, title)}

}
\author{
Paul Whitfield <paul.h.whitfield@gmail.com>
}
