% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_AHCCD_daily.R
\name{read_AHCCD_daily}
\alias{read_AHCCD_daily}
\title{Reads AHCCD daily file}
\usage{
read_AHCCD_daily(daily_file)
}
\arguments{
\item{daily_file}{Required. Name of the file to be read.}
}
\value{
If successful, returns the values in a dataframe, consisting of the date, 
the value and the data code. If unsuccessful, returns the value \code{FALSE}.
}
\description{
This program reads an Adjusted and Homogenized Canadian Climate Data (AHCCD) of daily 
precipitation or temperatures. The values are arranged as
month x day, which makes them difficult to read using standard \R functions.
}
\examples{
\dontrun{
stoon_daily_tmax <- read_AHCCD_daily("dx40657120.txt")}
}
\references{
Monthly AHCCD data are available from \url{http://www.ec.gc.ca/dccha-ahccd}. 
Daily values must be requested. Any use of the data must cite 
\cite{Mekis, E and L.A. Vincent, 2011: An overview of the second generation 
adjusted daily precipitation dataset for trend analysis in Canada. 
Atmosphere-Ocean, 49 (2), 163-177.}
}
\seealso{
\code{\link{read_AHCCD_monthly}} \code{\link{get_AHCCD_monthly}}
}
\author{
Kevin Shook
}
