% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_AHCCD_daily_precip.R
\name{read_AHCCD_daily_precip}
\alias{read_AHCCD_daily_precip}
\title{Reads AHCCD daily precipitation file}
\usage{
read_AHCCD_daily_precip(precip_file)
}
\arguments{
\item{precip_file}{Required. Name of the file to be read.}
}
\value{
If successful, returns the values in a data frame, consisting of the \code{date}, 
the precipitation variable and the data \code{code}. The name of the precipitation variable, \code{rain}, 
\code{snow} or \code{total} is inferred from the file name.
If unsuccessful, returns the value \code{FALSE}.
}
\description{
This program reads Adjusted and Homogenized Canadian Climate Data (AHCCD) 
file of rainfall, snowfall or total precipitation. The values are arranged as
month x day, which makes them difficult to read using standard \R functions.
}
\examples{
\dontrun{
stoon_rainfall <- read_AHCCD_daily_precip("dr_4057120.txt")}
}
\references{
Monthly AHCCD data are available from \url{http://www.ec.gc.ca/dccha-ahccd}. 
Daily values must be requested. Any use of the precipitation data must cite 
\cite{Mekis, E and L.A. Vincent, 2011: An overview of the second generation 
adjusted daily precipitation dataset for trend analysis in Canada. 
Atmosphere-Ocean, 49 (2), 163-177.}
}
\seealso{
\code{\link{read_AHCCD_daily_temp}} \code{\link{read_AHCCD_monthly_precip}}
}
\author{
Kevin Shook
}
