% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_AHCCD_daily_temp.R
\name{read_AHCCD_daily_temp}
\alias{read_AHCCD_daily_temp}
\title{Reads AHCCD daily temperature file}
\usage{
read_AHCCD_daily_temp(temp_file)
}
\arguments{
\item{temp_file}{Required. Name of the file to be read.}
}
\value{
If successful, returns the values in a dataframe, consisting of the date, 
the value and the data code. If unsuccessful, returns the value \code{FALSE}.
}
\description{
This program reads Adjusted and Homogenized Canadian Climate Data (AHCCD) of daily 
min, max or mean air temperatures. The values are arranged as
month x day, which makes them difficult to read using standard \R functions.
}
\examples{
\dontrun{
stoon_tmax <- read_AHCCD_daily_temp("dx40657120.txt")}
}
\references{
Monthly AHCCD data are available from \url{http://www.ec.gc.ca/dccha-ahccd}. 
Daily values must be requested. Any use of the temperature data must cite 
\cite{Vincent, L. A., X. L. Wang, E. J. Milewska, H. Wan, F. Yang, and V. Swail, 2012. 
A second generation of homogenized Canadian monthly surface air temperature for climate 
trend analysis, J. Geophys. Res., 117, D18110, doi:10.1029/2012JD017859.}
}
\seealso{
\code{\link{read_AHCCD_daily_temp}} \code{\link{read_AHCCD_monthly_precip}}
}
\author{
Kevin Shook
}
