% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_AHCCD_monthly_precip.R
\name{read_AHCCD_monthly_precip}
\alias{read_AHCCD_monthly_precip}
\title{Reads AHCCD monthly precipitation file}
\usage{
read_AHCCD_monthly_precip(precip_file)
}
\arguments{
\item{precip_file}{Required. Name of the file to be read.}
}
\value{
If successful, returns the values in a dataframe, consisting of the \code{year},
the \code{month}, the value and the data \code{code}. If unsuccessful, returns the value 
\code{FALSE}.
}
\description{
This program reads Adjusted and Homogenized Canadian Climate Data (AHCCD) data of 
rainfall, snowfall or total precipitation. The values are arranged as year x month, 
which makes them difficult to read using standard \R functions.
}
\examples{
\dontrun{
stoon_monthly_total <- read_AHCCD_monthly_precip("mt4057120.txt")}
}
\references{
Monthly AHCCD data are available from \url{http://www.ec.gc.ca/dccha-ahccd}. Any use of 
the precipitation data must cite \cite{Mekis, E and L.A. Vincent, 2011: An overview of 
the second generation adjusted daily precipitation dataset for trend analysis in Canada. 
Atmosphere-Ocean, 49 (2), 163-177.}
}
\seealso{
\code{\link{read_AHCCD_monthly_precip}} \code{\link{read_AHCCD_daily_temp}}
}
\author{
Kevin Shook
}
