% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wsc.R
\name{read_wsc}
\alias{read_wsc}
\title{A helper for wsc data using read.csv returns a datafrrame of the data
with the last two lines removed.  
Those contain the EC disclaimer.
read_wsc}
\usage{
read_wsc(filename)
}
\arguments{
\item{filename}{Datafile retrieved from ECDataExplorer}
}
\value{
a dataframe with the last two rows removed and the Date as Date
}
\description{
A helper for wsc data using read.csv returns a datafrrame of the data
with the last two lines removed.  
Those contain the EC disclaimer.
read_wsc
}
\examples{
\dontrun{mdata <- read_wsc("05AA008_Daily_Flow_ts.csv")}

}
\author{
Paul Whitfield <paul.h.whitfield@gmail.com>
}
