% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_RegSim.R
\name{RegSim}
\alias{RegSim}
\alias{RegSim.matrix}
\alias{RegSim.reglmom}
\title{Simulation of a regional dataset with intersite correlation.}
\usage{
RegSim(x, ...)

\method{RegSim}{matrix}(
  x,
  distr,
  nrec,
  corr = 0,
  corr.sqrt = FALSE,
  lmom = TRUE,
  lscale = FALSE,
  long = FALSE,
  ...
)

\method{RegSim}{reglmom}(x, n = 1, corr = 0, margin = "atsite", ...)
}
\arguments{
\item{x}{Matrix (in rows) of parameters or L-moments for all sites to 
simulate.
Can also be an output form \link{FitRegLmom}.}

\item{...}{Other parameters.}

\item{distr}{Marginal distribution of each site.
If only one value is passed as argument, the same is used for all sites.}

\item{nrec}{Record lengths of the sites.
If only one value is passed in argument, the same is used for all sites.}

\item{corr}{Correlation matrix for the dependence between site.
If only one value is passed, the correlation is assumed
the for every pair of sites.}

\item{corr.sqrt}{Squared correlation matrix. Can be passed to speed up
multiple calls.}

\item{lmom}{Logical. Is the argument `x` a matrix of L-moments or
distribution parameters}

\item{lscale}{Logical. Is the second L-moments the scale (`TRUE`)
or the LCV (`FALSE`).}

\item{long}{Logical. Should the output be returned in a long format.}

\item{n}{Number of simulations.}

\item{margin}{Which marginal distribution should be used. Either
based on the at-site (\code{'atsite'}) or regional (\code{'reg'}) 
distribution.}
}
\description{
Returns a dataset containing multiple time series in the form of
a matrix where the sites are in columns.
Different record lengths can be specified for each site
and missing values are filled accordingly at the beginning.
The rows (time) are independent and the intersite correlation model
is based on a multivariate Normal distribution.
}
\examples{

## Extract data
data(flowUngauged)
lmom <- flowUngauged[1:5, c('l1','lcv','lsk')]

## Simulate data base on at-site L-moments
sim <- RegSim(lmom, distr = 'gev', nrec = 11:15, corr = .4)

head(sim)

sim <- RegSim(lmom,
              distr = c(rep('gev',4),'gno'),
              nrec = 10:15,
              corr = .4, long = TRUE)
head(sim)

 
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
