% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_residuals_nsamax.R
\name{residuals.nsamax}
\alias{residuals.nsamax}
\title{Detrendend of the nonstationary AMAX model}
\usage{
\method{residuals}{nsamax}(object, scale = FALSE, ...)
}
\arguments{
\item{object}{Output from \link{FitNsAmax}.}

\item{scale}{Logical. Should the residuals be transformed to a standard 
normal distribution.}

\item{...}{Other parameters}
}
\description{
Return the detrended values of the nonstationary AMAX model.
}
\seealso{
\link{fitted.nsamax}, \link{FitNsAmax}.
}
