% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_SearchThresh.R
\name{SearchThresh}
\alias{SearchThresh}
\title{Fit POT models for a list of candidate thresholds}
\usage{
SearchThresh(form, x, u, nmin = 20, verbose = TRUE, ...)
}
\arguments{
\item{u}{Vector of candidate thresholds.}

\item{nmin}{Stopping condition verifying that a minimal number of peaks
are extracted.}

\item{verbose}{Logical. Should a progress bar be displayed.}

\item{...}{Other arguments passed to \link{FitPot}.}
}
\description{
The function \code{SearchThresh} returns a data frame containing
key value for the automatic selection of a threshold and where each row
corresponds to a candidate threshold.
The result is extracted from several calls to the function \code{FitPot}.
}
\examples{

# Create a list of candidate threshold
lstu <- which.floodPeaks(flow~date, flowStJohn, u =500, r = 14)
lstu <- sort(unique(flowStJohn[lstu,'flow']))
lstu <- lstu[seq(1,length(lstu)-30,2)]

out <- SearchThresh(flow~date, flowStJohn, u = lstu, declust = 'wrc', r = 14)
head(out)

FindThresh(out, method = 'sgn', tol.sgn = 0.1)
FindThresh(out, method = 'ppy', tol.ppy = 2)
FindThresh(out, method = 'max')

}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
