% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_SeasonStat.R
\name{SeasonStat}
\alias{SeasonStat}
\alias{SeasonStat.data.frame}
\alias{SeasonStat.formula}
\title{Seasonal statistics for flood peaks}
\usage{
SeasonStat(x, ...)

\method{SeasonStat}{data.frame}(x, ...)

\method{SeasonStat}{formula}(form, x, ...)
}
\arguments{
\item{x}{Data. If data.frame with two columns, they must be respectively 
the date and a site variable.}

\item{...}{Other parameters.}

\item{form}{Formula that specifies the date and site variable. Must be of the
form \code{date ~ site}.}
}
\description{
Return the circular or seasonal statistics of flood peaks.
The angle represent the average timing of the floods and the radius its
regularity. 
For instance a radius of one represent perfect regularity. 
Can perform the analysis of multiple sites.
}
\examples{

dt <- ExtractAmax(flow~date, flowStJohn)$date

SeasonStat(dt)

## Illustration of the analysis of multiple sites

F0 <- function(ii) data.frame(site = ii, dt = sample(dt, replace = TRUE))
x <- lapply(1:10, F0)
x <- do.call(rbind, x)
           
st <- SeasonStat(dt ~ site, x)

JulianPlot()
points(y ~ x, st, col = 2, pch = 16)

}
\references{
Burn, D.H. (1997). Catchment similarity for regional flood frequency analysis
  using seasonality measures. Journal of Hydrology 202, 212-230.
  https://doi.org/10.1016/S0022-1694(97)00068-1
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
