% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_simulate_nsamax.R,
%   R/rfa_simulate_nsamax_mle.R
\name{simulate.nsamax}
\alias{simulate.nsamax}
\alias{simulate.nsamax.mle}
\title{Simulation of nonstationary AMAX model.}
\usage{
\method{simulate}{nsamax}(object, nsim = 1, seed = NULL, ...)

\method{simulate}{nsamax.mle}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{Output from \link{FitNsAmax}.}

\item{nsim}{Number of simulations.}

\item{seed}{Seed for random number generator.}

\item{...}{Other parameters.}
}
\description{
Return the simulation of a fitted nonstationary AMAX model.
}
\seealso{
\link{FitNsAmax}.
}
