% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_rfa_simulate_nspot.R
\name{simulate.nspot}
\alias{simulate.nspot}
\title{Simulating a nonstationary POT model.}
\usage{
\method{simulate}{nspot}(object, nsim = 1, seed, newdata = NULL, ...)
}
\arguments{
\item{object}{Output from \link{FitNsPot}.}

\item{nsim}{Number of simulations.}

\item{seed}{Seed for the random number generator.}

\item{newdata}{Dataset of the covariates at the predicted dates.}

\item{...}{Other parameters.}
}
\description{
Return the simulations of a nonstationary POT model.
}
\seealso{
\link{FitNsPot}, \link{predict.nspot}.
}
