% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_simulate_poolmargin.R,
%   R/rfa_simulate_poolmle.R
\name{simulate.poolmargin}
\alias{simulate.poolmargin}
\alias{simulate.poolmle}
\title{Simulation of a pooling group models}
\usage{
\method{simulate}{poolmargin}(object, nsim = 1, seed = NULL,
  corr = 0, ...)

\method{simulate}{poolmle}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{Output from \link{FitPoolMle}.}

\item{nsim}{Number of simulations.}

\item{seed}{Seed for random number generator.}

\item{corr}{Correlation matrix of an underlying multivariate normal distribution.
If a scalar is passed, a constant coefficient is used.}

\item{...}{Other parameters.}
}
\description{
Return the simulation of multiple sites following a regional model.
}
\seealso{
\link{FitPoolMle}.
}
