% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice}
\alias{slice}
\title{Converts data into bins based upon a chosen step size}
\usage{
slice(doy, step)
}
\arguments{
\item{doy}{- an array of the day of calendar year for the dataset}

\item{step}{- width of bin in days}
}
\value{
{An array of bin numbers that can be used as a factor for each day in the dataset
prints a message indicating the handling of partial bins 
}
}
\description{
Converts data into bins based upon a chosen step size
}
\examples{
\dontrun{dice <- slice(doy,5)}
}
\author{
Paul Whitfield <paul.h.whitfield@gmail.com>
}
