% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalHydrology-methods.R
\name{StatisticalHydrology-methods}
\alias{StatisticalHydrology-methods}
\title{Statistical analysis functions}
\description{
These functions perform statistical analyses
\describe{
 \item{binned_MannWhitney}{ Compares two time periods of data using Mann-Whitney test}
 \item{fdcurve}{Finds flow exceedence probabilities}
 \item{get_peaks}{Finds peak flows over a specified threshold}
}
}
