% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_summary_bootns.R
\name{summary.bootns}
\alias{summary.bootns}
\title{Summary bootstrap sample for the nonstationary AMAX model}
\usage{
\method{summary}{bootns}(object, variable = "para", alpha = 0.05, ...)
}
\arguments{
\item{object}{Output from \link{ch_rfa_boot_nsamax} or \link{ch_rfa_boot_nsamaxMle}.}

\item{variable}{Variable to summarize. Must be one of 
\code{('para', 'beta', 'qua')}.}

\item{alpha}{Confidence level.}

\item{...}{Other parameters.}
}
\description{
Return the standard deviation and the confident interval of the parameters
and flood quantiles.
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
