% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/W05AA008.R
\docType{data}
\name{W05AA008}
\alias{W05AA008}
\title{W05AB}
\format{a dateframe with 7791 rows and 20 columns of metadata}
\source{
2016-10-16 release of ECDataExplorer
\itemize{
\item	{ID}  {StationID}
\item	{PARAM} {Parameter 1=Flow, 2=Level}
\item	{Date} {in date format}
\item	{Flow} {Flow in m3/sec}
\item	{SYM} {Water Survey FLags A, B, D, E }
}
}
\usage{
W05AA008
}
\description{
The files contains a dataframe produced from read_wsc containing station data
}
\keyword{datasets}
