% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_day.R
\name{which.day}
\alias{which.day}
\title{Where is a specific day of the year is ?}
\usage{
which.day(x, x0 = "0715")
}
\arguments{
\item{x}{Vector of dates}

\item{x0}{The day to find. For example, '0715' standand for 07/15 July 15th.}
}
\description{
Determines the location of a specific day of the year in a vector of 
\link{Date}.
}
\examples{

x <- seq(as.Date('2000/1/1'),as.Date('2020/12/31'), 'day')
yr <- which.day(x, c('0101','0715'))
head(x[yr])

}
