% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_rfa_which_floodPeaks.R
\name{which.floodPeaks}
\alias{which.floodPeaks}
\alias{which.floodPeaks.numeric}
\alias{which.floodPeaks.formula}
\alias{which.floodPeaks.data.frame}
\alias{which.clusters}
\alias{which.clusters.formula}
\alias{which.clusters.numeric}
\title{Extracting peaks}
\usage{
which.floodPeaks(x, ...)

\method{which.floodPeaks}{numeric}(x, dt = NULL, u, r = 1, rlow = 0.75, ini = "wrc", ...)

\method{which.floodPeaks}{formula}(form, x, u, ...)

\method{which.floodPeaks}{data.frame}(x, u, ...)

which.clusters(x, ...)

\method{which.clusters}{formula}(form, x, u, r = 1, ...)

\method{which.clusters}{numeric}(x, dt = NULL, u, r = 1, ...)
}
\arguments{
\item{x, form}{If numeric, \code{x} is a vector of data. If a formula is passed}

\item{...}{Other parameters.}

\item{dt}{Date or time of observations. If not provided, regular step
are asumed.}

\item{u}{Threshold.}

\item{r, rlow, ini}{Declustering parameters. See details.}
}
\description{
Returns the indices of the peaks above a threshold according to the
declustering method put in place by the Water Resources Council or simple
run declustering.
See Lang et al. (1999) for more details.
}
\details{
Two conditions are required for peaks to not be
rejected. First, two peaks \code{Q1} and \code{Q2} must be separated by a
 period of at least \code{r} days, where one recommendation is
  \deqn{4 days + log(A)}
and \eqn{A} is the drainage area in squared kilometers.
The second conditions is
\deqn{Xmin > rlow * min(Q1,Q2).} where \code{Xmin} is lowest point
between \code{Q1} and \code{Q2}. By defautlt, \code{rlow = 0.75}.
When one of the two conditions is not statisfied the lowest of the two
peaks is discarded.
The 2 conditions are verify sequentially, from an initial set of peaks.
If \code{ini = 'run'}, a run declustering method with one lag is first used to
filter the initial set of peaks to the maximums of the respective clusters.
If \code{ini = 'lmax'} the 2 condition are verified accross all local maximums.
If \code{ini = 'wrc'} (default), the second condition id verified
next the first condition is verify on the previously extracted peaks.
The two version are very similar and differ only on few cases where the
modified version is more conservative and reject peaks that are kept
in the initial version.

The function \code{which.clusters} is returning the indices of the peaks
identified by the run declustering method where clusters are separated by a
period of \code{r} consecutive values under the threshold.
}
\section{References}{


Lang M, Ouarda TBMJ, Bobee B. (1999) Towards operational guidelines for
  over-threshold modeling. Journal of Hydrology. Dec 6;225(3):103-17.
}

\examples{

# Declustering using the flood recommendation.
cid <- which.floodPeaks(flow~date, flowStJohn, u = 1000, r = 14, rlow = .75,
                         ini = 'wrc')

plot(flowStJohn, type = 'l')
points(flowStJohn[cid,], col = 'red', pch = 16)

## Simpler run declustering
cid <- which.clusters(flowStJohn$flow, u = 1000, r = 14)

plot(flowStJohn, type = 'l')
points(flowStJohn[cid,], col = 'red', pch = 16)
}
