% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_rfa_amax.R, R/ch_rfa_gev.R, R/ch_rfa_glo.R,
%   R/ch_rfa_gno.R, R/ch_rfa_pe3.R
\name{Amax}
\alias{Amax}
\alias{dAmax}
\alias{fAmax}
\alias{lAmax}
\alias{npAmax}
\alias{pAmax}
\alias{qAmax}
\alias{rAmax}
\alias{dgev}
\alias{fgev}
\alias{pgev}
\alias{qgev}
\alias{rgev}
\alias{dglo}
\alias{fglo}
\alias{pglo}
\alias{qglo}
\alias{rglo}
\alias{dgno}
\alias{fgno}
\alias{pgno}
\alias{qgno}
\alias{rgno}
\alias{ppe3}
\alias{qpe3}
\alias{dpe3}
\alias{rpe3}
\alias{fpe3}
\title{Basic probability functions for distribution of annual maximums}
\usage{
dAmax(x, para, distr, log = FALSE)

fAmax(x, distr)

lAmax(l, distr, ...)

npAmax(distr)

pAmax(q, para, distr)

qAmax(p, para, distr)

rAmax(n, para, distr)

dgev(x, xi, alf, kap, log = FALSE)

fgev(x, p0 = NULL, ...)

pgev(x, xi, alf, kap)

qgev(p, xi, alf, kap)

rgev(n, xi, alf, kap)

dglo(x, xi, alf, kap, log = FALSE)

fglo(x, p0 = NULL, ...)

pglo(x, xi, alf, kap)

qglo(p, xi, alf, kap)

rglo(n, xi, alf, kap)

dgno(x, xi, alf, kap, log = FALSE)

fgno(x, p0 = NULL, ...)

pgno(x, xi, alf, kap)

qgno(p, xi, alf, kap)

rgno(n, xi, alf, kap)

ppe3(x, mu, sig, g)

qpe3(p, mu, sig, g)

dpe3(x, mu, sig, g, log = FALSE)

rpe3(n, mu, sig, g)

fpe3(x, p0 = NULL, ...)
}
\arguments{
\item{x, q}{Vector of quantiles.}

\item{para}{Vector of parameters for the given distribution.}

\item{distr}{Distribution family. See \link{lmom2par}.}

\item{log}{Logical. If TRUE, probabilities \code{p} are given as
\code{log(p)}.}

\item{l}{L-moments.}

\item{...}{Other parameters passed to \link{optim} or \link{lmom2par}.}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations.}

\item{xi, alf, kap, mu, sig, g}{Location, scale and shape parameters.}

\item{p0}{Starting parameter for estimation using maximum likelihood.}
}
\description{
Density, distribution function, quantile function and random generation
for various distribution used in the modeling annual maximums.
}
\details{
The familly of \code{Amax} functions wraps the routines available in 
\link{lmomco-package} in order to simplify their usage within the family of apply
functions. 

The other functions provide a R-like interface to the Generalized Extreme Value
(GEV), Generalized Logistic (GLO), Generalized normal (GNO) and Pearson type
III (PE3) distributions.  
The f-functions such as \code{fgev} are low-level functions that 
carries out parameters estimation by maximum likelihood. 
The likelihood is optimized by the function \link{optim} 
to control the optimization process.
}
\examples{

## Create a list of 5 uniform sample
u <- replicate(5, runif(50), simplify = FALSE)

x <- lapply(u, qAmax, c(100,3,.001), 'gno')
p <- lapply(x, pAmax, c(100,3,.001), 'gno')
d <- lapply(x, dAmax, c(100,3,.001), 'gno')

## Estimation using L-moments
ll <- lapply(x, lmom::samlmu)
para <- lapply(ll, lAmax, 'gno')
para <- lapply(x, fAmax, 'gno')

## R-like Interface
x <- rgev(1000,100, 30, -.05)
p <- pgev(x, 100, 30, -0.05)
x <- qgev((1:1000)/1001, 100, 30, -0.05)
d <- dgev(x, 100, 30, -0.05)
f <- fgev(x)
f <- fpe3(x, method = 'BFGS', control = list(maxit = 1000))

}
\seealso{
\link{FitAmax}, \link{optim}.
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
