% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_data-manipulations-methods.R
\name{Basic_data_manipulations-methods}
\alias{Basic_data_manipulations-methods}
\title{Basic data manipulation functions}
\description{
These functions read in or convert values among formats
\describe{
 \item{read_ECDE_flows}{Reads a file of WSC daily flows from ECDataExplorer}
 \item{get_wscstation}{Reads station information from a data file produced by ECDE}
}
}
