% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booth_plot.R
\name{booth_plot}
\alias{booth_plot}
\title{Create a Booth plot}
\usage{
booth_plot(events, threshold, title, type = "mag", colour1 = 1,
  colour2 = 1)
}
\arguments{
\item{events}{A data frame of POT events from the function \code{get_peaks}}

\item{threshold}{The threshold used by \code{get_peaks}}

\item{title}{Plot title}

\item{type}{The plot type, either \option{mag} (magnitude, the default) or \option{vol} (volume)}

\item{colour1}{A vector of length 12 with line colours of rings or symbols. Defaults to those used by Booth.}

\item{colour2}{A vector of length 12 with fill colours of rings or symbols. Defaults to those used by Booth.}
}
\description{
A Booth plot is a plot of peaks over threhold flood events with duration on the horizontal and
either magnitude (default) or volume on the vertical axis
}
\examples{
\dontrun{
test <-booth_plot(events, threshold, title, type='mag')

test <-booth_plot(events, threshold, title, type='vol')}

}
\references{
Booth, E.G., Mount, J.F., Viers, J.H. 2006. Hydrologic Variability of the Cosumnes River Floodplain.
San Francisco Estuary & Watershed Science 4:21.

Whitfield, P.H., and J.W. Pomeroy. 2016. Changes to flood peaks of a mountain river: implications
for analysis of the 2013 flood in the Upper Bow River, Canada. Hydrological Processes 30:4657-73. doi:
10.1002/hyp.10957.
}
