% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_ams_timeseries.R
\name{ch_ams_timeseries}
\alias{ch_ams_timeseries}
\title{Produce a Time Series Plot}
\usage{
ch_ams_timeseries(date, flow, data = NULL)
}
\arguments{
\item{date}{Vector of dates or years, or name of data column containing the 
dates (unquoted).}

\item{flow}{Vector of flows, or name of data column containing the flows
(unquoted).}

\item{data}{Optional; data frame containing the specified data. If
specified, this data frame will be searched first when looking for data
vectors.}
}
\value{
A ggplot2 object. This means that you are able to add more layers
downstream.
}
\description{
Uses ggplot2 to produce a time series plot of values (e.g., flow) 
over time, where time is binned by year. Subsequent years are connected with
a line, which is broken when data are missing.
}
\examples{
library(lubridate)
set.seed(42)
dates <- ymd(paste(
  c(1991:1993, 1995, 1997:2012), "-",
  sample(1:12, size = 20, replace = TRUE), "-",
  sample(1:28, size = 20, replace = TRUE)
))
y <- stats::rexp(20)
y[18] <- NA
ch_ams_timeseries(dates, y)
df <- data.frame(year = year(dates), flow = y)
ch_ams_timeseries(year, 35.31467 * flow, data = df) +
  ylab("Peak Instantaneous Flow (cfs)") +
  theme_bw()
}
