% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ch_axis_doy}
\alias{ch_axis_doy}
\title{Generates the x axis for day of year}
\usage{
ch_axis_doy(wyear = 1)
}
\arguments{
\item{wyear}{Month to begin water year. Use \code{wyear = 1} for calendar year,
\code{wyear = 10} for October 1.}
}
\description{
Used by \code{ch_regime_plot}. This code deals only with the axis 
adjustments. Day of water year needs to be done separately
}
\author{
Paul Whitfield
}
\keyword{internal}
