% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_checkcatchment.R
\name{ch_checkcatchment}
\alias{ch_checkcatchment}
\title{Check Catchments}
\usage{
ch_checkcatchment(dem, catchment, outlet, outlet_label = NULL)
}
\arguments{
\item{dem}{Raster. DEM that catchment was generated from}

\item{catchment}{sf. catchment polygon}

\item{outlet}{sf. location of catchment outlet}

\item{outlet_label}{string. Label for outlet}
}
\value{
Generates contour map with catchment polygon and outlet
}
\description{
Generates a map with contours, catchment polygons, and outlet
}
\examples{
/dontrun{map <- ch_checkcatchment(dem, catchment_shape, outlet)}
}
