% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_checkcatchment.R
\name{ch_checkcatchment}
\alias{ch_checkcatchment}
\title{Check Catchments}
\usage{
ch_checkcatchment(dem, catchment, outlet, outlet_label = NULL)
}
\arguments{
\item{dem}{Raster. DEM that catchment was generated from}

\item{catchment}{sf. catchment polygon}

\item{outlet}{sf. location of catchment outlet}

\item{outlet_label}{string. Label for outlet}
}
\value{
\item{area_df}{Generates contour map with catchment polygon and outlet check}
}
\description{
This function generates a simple map to allow a visual assessment of the catchment 
boundaries relative to the elevation contours. It also generates a table summarizing the catchments, 
including the coordinates of the outlet point and the catchment area.
}
\examples{
\dontrun{
ch_checkcatchment()
}

}
\seealso{
\code{\link{ch_saga_fillsinks}} to fill sinks instead of removing
}
\author{
Dan Moore <dan.moore@ubc.ca>
}
