% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_checkchannels.R
\name{ch_checkchannels}
\alias{ch_checkchannels}
\title{Check Channels}
\usage{
ch_checkchannels(dem, channels, outlet)
}
\arguments{
\item{dem}{Raster. DEM catchment was generated from}

\item{outlet}{sf. location of catchment outlet}

\item{catchment}{sf. catchment polygon}
}
\value{
generates map with channel layer
}
\description{
Generates a map of the generated channel layer
}
\examples{
\dontrun{map <-  ch_checkchannels(dem, channels, outlet)}
}
