% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_checkchannels.R
\name{ch_checkchannels}
\alias{ch_checkchannels}
\title{Check Channels}
\usage{
ch_checkchannels(dem, channels, outlet)
}
\arguments{
\item{dem}{Raster. DEM catchment was generated from}

\item{channels}{sf. channel polyline}

\item{outlet}{sf. location of catchment outlet}
}
\value{
\item{check_map}{generates map with channel layer}
}
\description{
Generates a map of the generated channel layer.
}
\details{
This function generates a simple map of the drainage network plotted over the contours to allow a visual assessment.
}
\examples{
\dontrun{
map <-  ch_checkchannels(dem, channels, outlet)
}

}
\seealso{
\code{\link{ch_saga_fillsinks}} to fill sinks instead of removing
}
\author{
Dan Moore <dan.moore@ubc.ca>
}
