% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_clear_wd.R
\name{ch_clear_wd}
\alias{ch_clear_wd}
\title{Clear SAGA Working Directory}
\usage{
ch_clear_wd(wd)
}
\arguments{
\item{wd}{working directory file path
ch_clear_wd()}
}
\description{
This function empties and removes a SAGA working directory
}
