% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_clear_wd.R
\name{ch_clear_wd}
\alias{ch_clear_wd}
\title{Clear SAGA Working Directory}
\usage{
ch_clear_wd(wd)
}
\arguments{
\item{wd}{working directory file path}
}
\value{
\item{result}{returns TRUE upon successful execution}
}
\description{
Empties and removes a SAGA working directory.
}
\details{
The data for raster layers read in as SAGA sdat 
files (e.g. the contributing area, carea) are held on disk rather than in memory; if you clear the working directory you
will get an error if you try to use the raster layer.
}
\examples{
\dontrun{

# not run as clearing all files in a given directory cannot be tested in CRAN

# create a saga working directory
saga_wd <- tempdir()
ch_create_wd(saga_wd) # confirm creation

# clear the saga working directory
ch_clear_wd(saga_wd)
}

}
\seealso{
\code{\link{ch_create_wd}} to create working SAGA directory
}
\author{
Dan Moore
}
