% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_clear_wd.R
\name{ch_clear_wd}
\alias{ch_clear_wd}
\title{Clear SAGA Working Directory}
\usage{
ch_clear_wd(wd)
}
\arguments{
\item{wd}{working directory file path}
}
\value{
\item{result}{returns TRUE upon successful execution}
}
\description{
This function empties and removes a SAGA working directory. Note: the data for raster layers read in as SAGA sdat 
files (e.g. the contributing area, carea) are held on disk rather than in memory; if you clear the working directory you
will get an error if you try to use the raster layer.
}
\examples{
\dontrun{
# clear the current working directory
ch_clear_wd(getwd())
}

}
\seealso{
\code{\link{ch_create_wd}} to create working SAGA directory
}
\author{
Dan Moore <dan.moore@ubc.ca>
}
