% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_color_gradient.R
\name{ch_color_gradient}
\alias{ch_color_gradient}
\title{ch_color_gradient}
\usage{
ch_color_gradient(
  x,
  colors = c("darkred", "red", "white", "green", "darkgreen"),
  colsteps = 100,
  climits = NULL
)
}
\arguments{
\item{x}{array of variable}

\item{colors}{an array of colours to form the desired gradient. Default is 
("darkred", "red", "white", "green", "darkgreen")}

\item{colsteps}{number of steps to be used in gradient, default is 100.}

\item{climits}{provide specific limits for common scaling}
}
\value{
vector of colors
}
\description{
set colour gradient
}
\examples{
cxin <- c(0, 1, 1, 3, 4, 5, 10)
cxout <- ch_color_gradient(cxin)
#[1] "#8B0000" "#B50000" "#B50000" "#FF2B2B" "#FF9292"
#[6] "#FFF9F9" "#006400"
}
\author{
Paul Whitfield
}
