% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_contours.R
\name{ch_contours}
\alias{ch_contours}
\title{Create Contours}
\usage{
ch_contours(dem, zmin = NULL, zmax = NULL, n_levels = 10, z_levels = NULL)
}
\arguments{
\item{dem}{raster object of your dem in the desired projection (note: should have had sinks removed)}

\item{zmin}{minimum elevation value for contours}

\item{zmax}{maximum elevation value for contours}

\item{n_levels}{number of contour lines}

\item{z_levels}{levels at which to plot contours}
}
\value{
\item{contours_sf}{sf object containing contours}
}
\description{
Creates contour lines from a DEM.
}
\details{
Generates contour lines from a DEM, which are returned as an sf object. 
The user can either provide a vector of elevation values by specifying the \code{z_levels} argument, 
or by supplying the minimum and maximum elevations (\code{zmin} and \code{zmax}) 
and the number of contour lines (\code{n_levels}).
}
\examples{
# use volcano DEM
dem <- ch_volcano_raster()
# generate contours
contours <- ch_contours(dem)

# plot contours map
plot(contours)

}
\seealso{
\code{\link{ch_saga_fillsinks}} to fill sinks instead of removing
}
\author{
Dan Moore
}
