% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_contours.R
\name{ch_contours}
\alias{ch_contours}
\title{Create Contours}
\usage{
ch_contours(dem, zmin = NULL, zmax = NULL, n_levels = 10, z_levels = NULL)
}
\arguments{
\item{dem}{Raster object of your dem in the desired projection - should have had sinks removed}

\item{zmin}{minimum elevation value for contours}

\item{zmax}{maximum elevation value for contours}

\item{n_levels}{number of contour lines}

\item{z_levels}{levels at which to plot contours}
}
\value{
\item{contours_sf}{sf object containing contours}
}
\description{
creates contour lines from a DEM.
}
\details{
This function generates contour lines from a DEM, which are returned as an **sf** object. 
The user can either provide a vector of elevation values by specifying the `z_levels` argument, 
or by supplying the minimum and maximum elevations (`zmin` and `zmax`) and the number of contour lines (`n_levels`).
}
\examples{
\dontrun{
ch_contours()
}

}
\seealso{
\code{\link{ch_saga_fillsinks}} to fill sinks instead of removing
}
\author{
Dan Moore <dan.moore@ubc.ca>
}
