% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_create_wd.R
\name{ch_create_wd}
\alias{ch_create_wd}
\title{Create SAGA Working Directory}
\usage{
ch_create_wd(wd)
}
\arguments{
\item{wd}{working directory file path}
}
\value{
\item{\code{TRUE}}{returns \code{TRUE} upon successful execution}
}
\description{
Creates a SAGA working directory.
}
\examples{
ch_create_wd(tempdir())

}
\seealso{
\code{\link{ch_clear_wd}} to clear the working SAGA directory
}
\author{
Dan Moore
}
