% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_create_wd.R
\name{ch_create_wd}
\alias{ch_create_wd}
\title{Create SAGA Working Directory

This function creates a SAGA working directory.}
\usage{
ch_create_wd(wd)
}
\arguments{
\item{wd}{working directory file path}
}
\value{
\item{TRUE}{returns TRUE upon successful execution}
}
\description{
Create SAGA Working Directory

This function creates a SAGA working directory.
}
\examples{
\dontrun{
ch_create_wd()

}

}
\seealso{
\code{\link{ch_clear_wd}} to clear the working SAGA directory
}
\author{
Dan Moore <dan.moore@ubc.ca>
}
