% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_create_wd.R
\name{ch_create_wd}
\alias{ch_create_wd}
\title{Create working directory}
\usage{
ch_create_wd(wd)
}
\arguments{
\item{wd}{name of a directory in which to store files created by WhiteboxTools functions}
}
\value{
\item{\code{TRUE}}{returns \code{TRUE} upon successful execution}
}
\description{
Creates a working directory.
}
\examples{
\donttest{
# not tested automatically as will return a warning
ch_create_wd(tempdir())
}
}
\seealso{
\code{\link{ch_clear_wd}} to clear the working directory
}
\author{
Dan Moore
}
