% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_cut_block.R
\name{ch_cut_block}
\alias{ch_cut_block}
\title{Extracts a specific time period from a longer record}
\usage{
ch_cut_block(DF, st_date, end_date)
}
\arguments{
\item{DF}{A daily streamflow data frame as from \code{ch_read_ECDE_flows}}

\item{st_date}{starting date format is \%Y/\%m/\%d}

\item{end_date}{ending date format is \%Y/\%m/\%d}
}
\value{
Returns a data within the specific time period as a dataframe.
}
\description{
Extracts a specific time period from a longer record.
 Could be used to get the same period of time from several stations for comparison.
}
\examples{
data(W05AA008)
subset <- ch_cut_block(W05AA008,"2000/01/01", "2010/12/31")
}
\author{
Paul Whitfield
}
