% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_cut_block.R
\name{ch_cut_block}
\alias{ch_cut_block}
\title{Extracts a specified time period from a longer record}
\usage{
ch_cut_block(DF, st_date, end_date)
}
\arguments{
\item{DF}{A daily streamflow data frame as from \code{ch_read_ECDE_flows}}

\item{st_date}{starting date format is \%Y/\%m/\%d}

\item{end_date}{ending date format is \%Y/\%m/\%d}
}
\value{
Returns a portion of the original dataframe.
}
\description{
The function could also be used to get the same period of time from several station for comparison.
}
\examples{
data(CAN05AA008)
subset <- ch_cut_block(CAN05AA008,"2000/01/01", "2010/12/31")
}
\author{
Paul Whitfield
}
