% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_decades_plot.R
\name{ch_decades_plot}
\alias{ch_decades_plot}
\title{Plots output from ch_binned_MannWhitney on x/y plot.}
\usage{
ch_decades_plot(mplot)
}
\arguments{
\item{mplot}{output from the function \code{ch_binned_MannWhitney}}
}
\value{
No value is returned; a standard \R graphic is created.
}
\description{
Creates a simple plot comparing two decades from the output of \code{ch_binned_MannWhitney}.
}
\examples{
\dontrun{
# mplot from ch_binned_MannWhitney
ch_decades_plot(mplot)   
}   

}
\author{
Paul Whitfield
}
