% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_decades_plot.R
\name{ch_decades_plot}
\alias{ch_decades_plot}
\title{Plots the output from \code{ch_binned_MannWhitney}}
\usage{
ch_decades_plot(mplot)
}
\arguments{
\item{mplot}{output from \code{ch_binned_MannWitney}}
}
\value{
a plot that compares two decades (time period)
}
\description{
Plots the output from \code{ch_binned_MannWhitney}, which compares two decades (time period) 
from a dataset.
}
\examples{
\dontrun{
# mplot from ch_binned_MannWhitney
ch_decades_plot(mplot)   
}   

}
\author{
Paul Whitfield
}
