% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_doys.R
\name{ch_doys}
\alias{ch_doys}
\title{Days of year and water year.}
\usage{
ch_doys(Date, water_yr = 10)
}
\arguments{
\item{Date}{an array as.Date}

\item{water_yr}{the month starting the water year, default is 10 (October).}
}
\value{
a dataframe with different date information:

\code{Date} - in Date format

\code{year}  - numeric calendar year

\code{month} - number calendar month

\code{doy}   - numeric day of year

\code{wyear} - numeric water year starting on day 1 of selected month

\code{dwy}   - numeric day of water year
}
\description{
Converts an array of dates into a dataframe with date, Year, month, doy,
wyear, dowy. 
Calculates day of year and day of water year since 10.01 of present water year,
 using base::isodate.
}
\details{
Converts a date array into a data frame with years, wateryears, and days of year and of water year.
}
\examples{
dd <- seq.Date(as.Date("2010-01-01"), as.Date("2018-01-01"),by = 1)
output <- ch_doys(dd, water_yr=10)
head(output)

}
\author{
Paul Whitfield
}
