% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_fdcurve.R
\name{ch_fdcurve}
\alias{ch_fdcurve}
\title{Plot Flow Duration Curve}
\usage{
ch_fdcurve(flow, title = "", normal = FALSE, gust = TRUE)
}
\arguments{
\item{flow}{Vector containing daily flows}

\item{title}{The plot title}

\item{normal}{If \code{normal = TRUE} then exceedance probability is normalized. Default is \code{FALSE}}

\item{gust}{If \code{TRUE} (the default), adds the curves from Gustard et al 1992}
}
\value{
Plots the flow durations and returns a data frame containing the exceedance probabilty and flow
}
\description{
A flow duration curve is a plot of flow magnitude against exceedance probability.
The plot may contain the Gustard Curves or they can be omitted. The default is for
curves to be plotted against probability, but an option is to plot against the normalized
exceedance probability. In that case, the x axis represents a normal distribution.
}
\examples{
flow <- W05AA008$Flow
# plot with Gustard 1992 curves
test <- ch_fdcurve(flow, title = "Station", normal = FALSE, gust = TRUE)

# plot with normalized exceedance probability
test <- ch_fdcurve(flow, title = "Station", normal = TRUE, gust = FALSE)
}
\references{
Gustard, A., A. Bullock, and J.M. Dixon. 1992. Low flow estimation in the United Kingdom.
Institute of Hydrology, 292. Wallingford: Institute of Hydrology.

Vogel, R.M., and N.M. Fennessy. 1994. Flow-duration curves. I: New Interpretation and
confidence intervals. Journal of Water Resources Planning and Management ASCE 120:485-504.
}
\author{
Paul Whitfield <paul.h.whitfield@gmail.com>
}
\keyword{plot}
