% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_flow_raster_trend.R
\name{ch_flow_raster_trend}
\alias{ch_flow_raster_trend}
\title{Raster plot and simple trends of observed streamflows}
\usage{
ch_flow_raster_trend(
  date,
  flow,
  step = 5,
  stationID = "",
  title = "",
  missing = FALSE,
  colours = c("lightblue", "cyan", "blue", "slateblue", "darkblue", "red")
)
}
\arguments{
\item{date}{A numeric vector of the date as an \R date. Must be same length as the \code{flow}}

\item{flow}{A numeric vector of daily streamflows}

\item{step}{An integer indicating the degree of smoothing eg. 1, 5, 11.}

\item{stationID}{Station ID number, e.g. "05BB001". This value is optional, but is included in the output to help you identify the results.}

\item{title}{Title of the plot}

\item{missing}{If \code{FALSE} years with missing data are excluded. If \code{TRUE} partial years
are included.}

\item{colours}{A vector of colours used for the raster plot. The default is \code{c("lightblue","cyan", "blue", "slateblue", "darkblue", "red")}.}
}
\value{
a list containing:
\describe{
	\item{stationID}{Station ID eg 05BB001}
	\item{missing}{How missing values were used FALSE=used, TRUE=removed}
	\item{step}{number of days in a bin}
	\item{periods}{number of periods in a year}
	\item{period}{period numbers i.e. 1:365/step}
	\item{bins}{values for each period in each year}
	\item{med_period}{median for each period}
	\item{max_period}{maximum for each period}
	\item{min_period}{minimum for each period}
	\item{tau_period}{Kendall's Tau for each period}
	\item{prob_period}{probability of Tau for each period}
	\item{year}{years spanning the data}
	\item{median_year}{median bin for each year}
	\item{max_year}{maximum bin for each year}
	\item{min_year}{minimum bin for each year}
	\item{tau_median_year}{value of tau and probability for median_year}
	\item{tau_maximum_year}{value of tau and probability for max_year}
	\item{tau_minimum_year}{value of tau and probability for min_year}
	}
}
\description{
Creates a raster plot plus trend plots for day of year, and over time which may be binned by a number of days
}
\details{
The plot contains four panels based upon binned data:
\enumerate{
\item The annual maximum, minimum, and median dlow with a trend test number
for each period: red arrows indicate decreases, blue arrows indicate increases.
\item The scale bar for the colours used in the raster plot,
\item The rasterplot with a colour for each
period and each year where data exist, and
\item A time series plot of the minimum, median, and maximum annual bin values. If there is no trend (p > 0.05) the points are black. Decreasing trend are in red, increasing trends are in blue..
}
}
\examples{
 mdoy <- doys(W05AA008$Date)
 mplot <- ch_flow_raster_trend(W05AA008$Date, W05AA008$Flow, step = 5, station = "05AA008")

}
\seealso{
\code{\link{ch_flow_raster}}
}
\author{
Paul Whitfield <paul.h.whitfield@gmail.com>
}
\keyword{plot}
