% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_regime_plot.R
\name{ch_regime_plot}
\alias{ch_regime_plot}
\title{Plots the regime of daily streamflows}
\usage{
ch_regime_plot(date, flow, title = "", wyear = 1, colour = TRUE, mx = 1)
}
\arguments{
\item{date}{Vector of dates}

\item{flow}{Vector of daily streamflows. Must be the same length as \code{date}}

\item{title}{Text to be used as the graph title}

\item{wyear}{Beginning month of water year. USe \code{wyear = 10} for October water year, \code{wyear = 1} for calendar year}

\item{colour}{Logical. If \code{TRUE} plot is in colour, if \code{FALSE} plot is grayscale}

\item{mx}{The maximum y value; if \code{mx = 1} then maximum value of the flows is used to set the maximum y-axis value. The value of \code{mx} can be specified to produce a series of plots with the same scale.}
}
\value{
No value is returned; a standard \R graphic is created.
}
\description{
Produces a regime hydrograph similar to that in the reference. It shows the flow quantiles for each
day of the year and the maximum and minimum. Parameters can be set to change colours and fix the y scale
to allow plots of same scale to be produced.
}
\examples{
ch_regime_plot(W05AA008$Date, W05AA008$Flow, title = "05AA008", colour = TRUE, wyear = 10)

}
\references{
MacCulloch, G. and P. H. Whitfield (2012). Towards a Stream Classification System
for the Canadian Prairie Provinces. Canadian Water Resources Journal 37: 311-332.
}
\author{
Paul Whitfield
}
