% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_regime_plot2.R
\name{ch_regime_plot2}
\alias{ch_regime_plot2}
\title{Plots the regime of daily streamflows using quantiles}
\usage{
ch_regime_plot2(
  date,
  flow,
  data = NULL,
  id = NULL,
  wyear = 1,
  colour = TRUE,
  metadata = HYDAT_list,
  ylab = expression(paste("Mean Daily Discharge (", m^3, "/sec)")),
  quant = c(0.95, 0.9, 0.75, 0.5, 0.25, 0.1, 0.05),
  ...
)
}
\arguments{
\item{date, flow}{Vectors of dates and flow data, or names of the columns
containing such data in `data`.}

\item{data}{Optional data frame containing the data.}

\item{id}{Gauge code used for creating a smart plot title; could also
be a column name in `data`. If `NULL` (the default), title creation
is bypassed.}

\item{wyear}{set \code{wyear = 10} for October, \code{water year = 1} for calendar year, can be any month}

\item{colour}{if \code{TRUE} plot is in colour, if \code{FALSE} plot is grayscale.}

\item{metadata}{a data frame of metadata to look for gauge code for `id`, 
defaults to `HYDAT_list` and is not used if `id` isn't.}

\item{ylab}{Y axis label.}

\item{quant}{quantiles; default is \code{quant = c(0.95,0.9,0.75,0.5,0.25,0.1,0.05)}. 
Can be changed but the length must be 7 and the 4th value must be 0.5 (median)}

\item{...}{Other arguments to pass to the `plot()` function. Will take
precedence over defaults.}
}
\value{
No value is returned; a standard \R graphic is created.
}
\description{
Produces a regime hydrograph similar to that in the reference. It shows the flow quantiles for each
day of the year and the maximum and minimum. Parameters can be set to change colours and set the y-scale
to allow plots of same scale to be produced.
}
\examples{
# Refer to columns in the `CAN05AA008` dataset:
ch_regime_plot2(Date, Flow, data = CAN05AA008, id = ID)

# Optionally, can insert vectors.
x <- CAN05AA008$Date
y <- CAN05AA008$Flow
ch_regime_plot2(x, y)

# Override `plot()` defaults using ...; for instance, zero-in on the freshet.
ch_regime_plot2(x, y, xlim = c(90, 220))
}
\references{
MacCulloch, G. and P. H. Whitfield (2012). Towards a Stream Classification System 
for the Canadian Prairie Provinces. Canadian Water Resources Journal 37: 311-332.
}
\author{
Paul Whitfield
}
