% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_saga_carea.R
\name{ch_saga_carea}
\alias{ch_saga_carea}
\title{Contributing Area Grid}
\usage{
ch_saga_carea(
  dem,
  saga_wd,
  method = 4,
  linear_threshold = Inf,
  saga_env = RSAGA::rsaga.env()
)
}
\arguments{
\item{dem}{raster object of your raw dem in the desired projection}

\item{saga_wd}{working directory to write and read saga files}

\item{method}{character or numeric: choice of processing algorithm (default \option{mfd}, or \code{4})}

\item{linear_threshold}{numeric (number of grid cells): threshold above which 
linear flow (i.e. the Deterministic 8 algorithm) will be used; linear flow is 
disabled for \code{linear_threshold = Inf} (the default)}

\item{saga_env}{SAGA environment object. Default is to let saga find it on its own.}
}
\value{
\item{carea}{raster of contributing areas for each pixel.}
}
\description{
Generates a grid of contributing area for each grid cell.
}
\details{
Generates a raster of contributing areas based on a DEM, 
which should have had sinks removed in a pre-processing step. For more information, see 
\code{?rsaga.topdown.processing}.
}
\examples{
\dontrun{
# note: example not run in package compilation
# - requires creating and accessing a temporary directory
# - requires downloading spatial data from Zenodo repository
# - requires a potentially lengthy GIS operation

# create saga wd using base::tempdir()
saga_wd <- tempdir()

# download LiDAR DEM for 240 and 241 creek
ff <- "gs_be240.tif"
ra_fn <- file.path(saga_wd, ff)
ra_url <- sprintf("https://zenodo.org/record/4781469/files/\%s",ff)
dem <- ch_get_url_data(ra_url, ra_fn)

# fill sinks
filled_dem <-  ch_saga_fillsinks(dem_raw=dem, saga_wd=saga_wd)

# determine contributing area raster using filled_dem
carea <- ch_saga_carea(filled_dem, saga_wd)

# plot contributing area raster
library(raster)
plot(carea)
}

}
\seealso{
\code{\link{ch_saga_fillsinks}} to fill sinks instead of removing

\code{\link{rsaga.topdown.processing}} for more information
}
\author{
Dan Moore
}
