% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_saga_carea.R
\name{ch_saga_carea}
\alias{ch_saga_carea}
\title{Contributing Area Grid}
\usage{
ch_saga_carea(
  dem,
  saga_wd,
  method = 4,
  linear_threshold = Inf,
  saga_env = RSAGA::rsaga.env()
)
}
\arguments{
\item{dem}{Raster object of your raw dem in the desired projection}

\item{saga_wd}{working directory to write and read saga files}

\item{method}{character or numeric: choice of processing algorithm (default "mfd", or 4)}

\item{linear_threshold}{numeric (number of grid cells): threshold above which linear flow (i.e. the Deterministic 8 algorithm) will be used; linear flow is disabled for linear_threshold=Inf (the default)}

\item{saga_env}{SAGA environment object.  Default is to let saga find it on its own.}
}
\value{
\item{carea}{raster of contributing areas for each pixel.}
}
\description{
Generates a grid of contributing area for each grid cell.
}
\details{
This function generates a raster of contributing areas based on a DEM, 
which should have had sinks removed in a pre-processing step. For more information, see 
?`rsaga.topdown.processing`.
}
\examples{
\dontrun{
ch_saga_carea()
}

}
\seealso{
\code{\link{ch_saga_fillsinks}} to fill sinks instead of removing

\code{\link{rsaga.topdown.processing}} for more information
}
\author{
Dan Moore <dan.moore@ubc.ca>
}
