% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_saga_catchment.R
\name{ch_saga_catchment}
\alias{ch_saga_catchment}
\title{Create Catchment Polygons}
\usage{
ch_saga_catchment(
  dem,
  saga_wd,
  outlet,
  carea = NULL,
  carea_flag = 0,
  outlet_label = NULL,
  buffsize = 100,
  saga.env = RSAGA::rsaga.env()
)
}
\arguments{
\item{dem}{raster object of your dem in the desired projection - should have had sinks removed}

\item{saga_wd}{name of working directory}

\item{outlet}{sf point object with coordinates of catchment outlets in the same projection as the dem}

\item{carea}{raster object containing contributing areas (default none provided)}

\item{carea_flag}{if \code{carea = NULL}, 0 = create carea from dem; 1 = read in carea.sdat}

\item{outlet_label}{character vector of labels; if \option{NULL}, numbers are assigned}

\item{buffsize}{numeric; the buffer radius (m) around catchment outlet to find location on digital stream network.}

\item{saga.env}{SAGA environment object. Default is to let saga find it on its own.}
}
\value{
\item{catchment_sf}{Returns an \pkg{sf} object containing catchment polygons}
}
\description{
Generates catchment boundaries for a set of specified points of interest along 
a stream network. The output is a \pkg{sf} object containing the polygons.
}
\details{
The function has two gridded inputs: a DEM (with sinks removed as appropriate) 
and a grid of contributing areas. Other inputs include \code{outlet}, a 
\pkg{sf} object containing the coordinates of the catchment outlets, and 
\code{outlet_label}, an optional vector of character strings that serve as labels 
or descriptions of the outlet points.

The DEM is input as a raster object via the \code{dem} argument.

There are three ways to provide the contributing area grid, which can be 
accomplished using the \code{carea} and \code{carea_flag} arguments, as summarized below. 

  1. The first option (\code{carea = NULL} and \code{carea_flag = 0}) would be 
  appropriate if you have not generated a contributing area grid prior to 
  application of \code{ch_saga_catchment}. If you use this option, the 
  contributing area grid is generated using the standard SAGA defaults, 
  with no options for the user to change them.  

  2. The second option (\code{carea = NULL} and \code{carea_flag = 1}) would be 
  appropriate if you have run the \code{ch_saga_carea} function and have not 
  cleared the working directory.  

  3. The third option would be appropriate if you have generated a contributing 
  area grid as a raster object using either \code{ch_saga_carea} or another function.  

The second and third options are particularly useful if the user either (a) wants to use a different approach than the SAGA defaults to generate the contributing area grid or (b) wants to use the contributing area grid in other analyses (e.g., computing a topographic wetness index).

Note that only one of the \code{carea} and \code{carea_flag} arguments need be specified.
}
\examples{

\donttest{
# note: example not tested in package compilation
# - requires creating and accessing a temporary directory
# - requires downloading spatial data from Zenodo repository
# - requires multiple potentially lengthy GIS operations

# create saga wd using base::tempdir()
saga_wd <- tempdir()

# download LiDAR DEM for 240 and 241 creek
ff <- "gs_be240.tif"
ra_fn <- file.path(saga_wd, ff)
ra_url <- sprintf("https://zenodo.org/record/4781469/files/\%s",ff)
dem <- ch_get_url_data(ra_url, ra_fn)

# fill sinks
filled_dem <-  ch_saga_fillsinks(dem_raw=dem, saga_wd=saga_wd)

# download station locations for 240, 241 (use as catchment outlets)
ff <- "gs_weirs.GeoJSON"
gs_fn <- file.path(saga_wd, ff)
gs_url <- sprintf("https://zenodo.org/record/4781469/files/\%s",ff)
stns <- ch_get_url_data(gs_url, gs_fn)[1:2,]

# determine contributing area raster using filled_dem
carea <- ch_saga_carea(filled_dem, saga_wd)

# run catchment delineation
catchments <-  ch_saga_catchment(dem=filled_dem, saga_wd=saga_wd, outlet=stns, carea=carea)
}

}
\seealso{
\code{\link{ch_saga_fillsinks}} to fill sinks instead of removing
}
\author{
Dan Moore
}
