% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_saga_channels.R
\name{ch_saga_channels}
\alias{ch_saga_channels}
\title{Create Catchment Channels}
\usage{
ch_saga_channels(
  dem,
  saga_wd,
  carea = NULL,
  carea_flag = 0,
  out_shp = TRUE,
  out_ntwrk = FALSE,
  out_route = FALSE,
  initmethod = 2,
  initvalue = 0,
  divcells = 5,
  minlen = 10,
  saga.env = RSAGA::rsaga.env()
)
}
\arguments{
\item{dem}{Raster object of your dem in the desired projection - should have had sinks removed}

\item{saga_wd}{name of working directory}

\item{carea}{raster object containing contributing areas (default none provided)}

\item{carea_flag}{if carea = NULL, 0 = create carea from dem; 1 = read in carea.sdat}

\item{out_shp}{if TRUE, return channel network as sf object}

\item{out_ntwrk}{if TRUE, return specified object as a raster}

\item{out_route}{if TRUE, return specified object as a raster}

\item{initmethod}{parameters for SAGA function}

\item{initvalue}{parameters for SAGA function}

\item{divcells}{parameters for SAGA function}

\item{minlen}{parameters for SAGA function}

\item{saga.env}{SAGA environment object.  Default is to let saga find it on its own.}
}
\value{
out_list list containing (as specified) channels (shp), ntwrk (raster) and route (raster)
ch_saga_channels()
}
\description{
determines channel layer spatial files
}
