% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_saga_channels.R
\name{ch_saga_channels}
\alias{ch_saga_channels}
\title{Create Catchment Channels}
\usage{
ch_saga_channels(
  dem,
  saga_wd,
  carea = NULL,
  carea_flag = 0,
  out_shp = TRUE,
  out_ntwrk = FALSE,
  out_route = FALSE,
  initmethod = 2,
  initvalue = 0,
  divcells = 5,
  minlen = 10,
  saga.env = RSAGA::rsaga.env()
)
}
\arguments{
\item{dem}{raster object of your dem in the desired projection - should have had sinks removed}

\item{saga_wd}{name of working directory}

\item{carea}{raster object containing contributing areas (default none provided)}

\item{carea_flag}{if \code{carea = NULL}, 0 = create carea from dem; 1 = read in carea.sdat}

\item{out_shp}{if TRUE, return channel network as sf object}

\item{out_ntwrk}{if TRUE, return specified object (channel network) as a raster}

\item{out_route}{if TRUE, return specified object (channel route) as a raster}

\item{initmethod}{parameters for SAGA function}

\item{initvalue}{parameters for SAGA function}

\item{divcells}{parameters for SAGA function}

\item{minlen}{parameters for SAGA function}

\item{saga.env}{SAGA environment object. Default is to let saga find it on its own.}
}
\value{
\item{out_list}{list containing (as specified) channels (shp), ntwrk (raster) and route (raster)}
}
\description{
Determines channel layer spatial files
}
\details{
Generates a drainage network using the SAGA \code{ta_channels} module with option 0.

The function requires a DEM and a contributing area grid. Like \code{ch_saga_catchment}, 
the user can provide the contributing area grid either as a raster object or by reading 
in a SAGA file; alternatively, the function can compute the contributing area grid from the DEM.

The SAGA function generates three outputs: 
 
 1. a shape file containing the drainage network,  
 
 2. a grid containing flow directions, and  
 
 3. a grid in which, if a cell is part of a channel, its value equals the channel order; otherwise the cell is marked as no-data.  

The output is a list, which will have elements that will include the channel network shape as an
 \pkg{sf} object (if \code{out_shp = TRUE}), the flow direction grid as a \pkg{raster} 
 object (if \code{out_route = TRUE}) and the network grid as a \pkg{raster} object 
 (if \code{out_ntwrk = TRUE})
}
\examples{

\donttest{
# note: example not tested in package compilation
# - requires creating and accessing a temporary directory
# - requires downloading spatial data from Zenodo repository
# - requires multiple potentially lengthy GIS operations

# create saga wd using base::tempdir()
saga_wd <- tempdir()

# download 25m DEM
ff <- "gs_dem25.tif"
ra_fn <- file.path(saga_wd, ff)
ra_url <- sprintf("https://zenodo.org/record/4781469/files/\%s",ff)
dem <- ch_get_url_data(ra_url, ra_fn)

# fill sinks
filled_dem <-  ch_saga_fillsinks(dem_raw=dem, saga_wd=saga_wd)

# determine contributing area raster using filled_dem
carea <- ch_saga_carea(filled_dem, saga_wd)

# generate channels sf object
channels <- ch_saga_channels(dem=filled_dem, saga_wd=saga_wd, carea=carea)
}

}
\seealso{
\code{\link{ch_saga_catchment}} for catchment delineation.

For more details, see the following: (http://www.saga-gis.org/saga_tool_doc/2.2.5/ta_channels_0.html)
}
\author{
Dan Moore
}
