% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_saga_fillsinks.R
\name{ch_saga_fillsinks}
\alias{ch_saga_fillsinks}
\title{Fill DEM}
\usage{
ch_saga_fillsinks(
  dem_raw,
  saga_wd,
  sinkmethod = "planchon.darboux.2001",
  minslope = 0.1,
  saga.env = RSAGA::rsaga.env()
)
}
\arguments{
\item{dem_raw}{Raster object of your raw dem in the desired projection}

\item{saga_wd}{working directory to write and read saga files}

\item{sinkmethod}{options are planchon.darboux.2001" (default), "xxl.wang.liu.2006" and "wang.liu.2006"}

\item{minslope}{minimum slope to impose on filled cells to force drainage. Default = 0.1}

\item{saga.env}{SAGA environment object.  Default is to let saga find it on its own.}
}
\value{
\item{dem_ns}{filled dem as a raster object.}
}
\description{
Removes sinks in a DEM by filling. This function as a wrapper to the \code{RSAGA::rsaga.fill.sinks} function.
}
\examples{
\dontrun{
ch_saga_fillsinks()

# consider sample DEM data with this
}

}
\seealso{
\code{\link{ch_saga_removesinks}} to remove sinks instead of filling
}
\author{
Dan Moore <dan.moore@ubc.ca>
}
