% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_saga_removesinks.R
\name{ch_saga_removesinks}
\alias{ch_saga_removesinks}
\title{Deepen Drainage Network}
\usage{
ch_saga_removesinks(dem_raw, saga_wd, saga.env = RSAGA::rsaga.env())
}
\arguments{
\item{dem_raw}{Raster object of your raw dem in the desired projection}

\item{saga_wd}{working directory to write and read saga files}

\item{saga.env}{SAGA environment object.  Default is to let saga find it on its own.}
}
\value{
dem_ns processed dem as a raster object 
ch_saga_removesinks()
}
\description{
removes sinks by deepening drainage network; alternative to ch_saga_fillsinksg
}
