% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_saga_removesinks.R
\name{ch_saga_removesinks}
\alias{ch_saga_removesinks}
\title{Deepen Drainage Network}
\usage{
ch_saga_removesinks(dem_raw, saga_wd, saga_env = RSAGA::rsaga.env())
}
\arguments{
\item{dem_raw}{Raster object of your raw dem in the desired projection}

\item{saga_wd}{working directory to write and read saga files}

\item{saga_env}{SAGA environment object.  Default is to let saga find it on its own.}
}
\value{
{dem_ns}{processed dem as a raster object.}
}
\description{
Removes sinks by deepening drainage network; alternative to ch_saga_fillsinks. This function acts as a wrapper to 
the `rsaga.sink.removal` function.
}
\examples{
\dontrun{
ch_saga_removesinks()

# consider sample DEM data with this
}

}
\seealso{
\code{\link{ch_saga_fillsinks}} to fill sinks instead of removing
}
\author{
Dan Moore <dan.moore@ubc.ca>
}
