% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ch_test_url_file}
\alias{ch_test_url_file}
\title{Tests url to see if it will work}
\usage{
ch_test_url_file(url, quiet = FALSE)
}
\arguments{
\item{url}{Required. URL to be checked}

\item{quiet}{Optional. If \code{FALSE} (the default) messages are printed.}
}
\value{

}
\description{
Tests url to see if it will work
}
\author{
Kevin Shook
}
\keyword{internal}
