% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_tidyhydat_ECDE.R
\name{ch_tidyhydat_ECDE}
\alias{ch_tidyhydat_ECDE}
\title{Converts tidyhydat daily flow tibble data to ECCCDE format}
\usage{
ch_tidyhydat_ECDE(data)
}
\arguments{
\item{data}{Tibble of daily flows retrieved using \pkg{tidyhydat} function \code{hy_daily_flows}.}
}
\value{
A dataframe or a list of flows with formats consistent with datafiles from ECDE.
}
\description{
Converts a tibble obtained from a \pkg{tidyhydat} to a dataframe 
with standard Environment and Climate Change Canada Data Explorer (ECCCDE) headings.
}
\examples{
\dontrun{
library(tidyhydat)
mdata <- hy_daily_flow(station_number=c("05CK004"))
m_data <- ch_tidyhydat_ECDE(mdata)

mdata <- hy_daily_flows(station_number=c("05CK004","08MF005","05BB001"))
mnew <- ch_tidyhydat_ECDE(mdata)
str(mnew[[1]])
str(mnew[[2]])
str(mnew[[3]]) 
#note the order is in increasing alphabetical order}
}
\seealso{
\code{\link{ch_tidyhydat_ECDE_meta}}
}
\author{
Paul Whitfield <paul.h.whitfield@gmail.com>
}
