% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_volcano_pourpoints.R
\name{ch_volcano_pourpoints}
\alias{ch_volcano_pourpoints}
\title{Creates a sample file of pour points}
\usage{
ch_volcano_pourpoints(pp_shp)
}
\arguments{
\item{pp_shp}{Name for shapefile to hold pour points}
}
\value{
Returns \code{NULL}. Writes pour points to specified file.
}
\description{
Creates a file of pour points for the \code{volcano} DEM
}
\seealso{
\code{\link{ch_volcano_raster}} \code{\link{ch_wbt_pourpoints}}
}
\author{
Dan Moore
}
