% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_wbt_catchment.R
\name{ch_wbt_catchment}
\alias{ch_wbt_catchment}
\title{Delineate catchment boundaries}
\usage{
ch_wbt_catchment(
  fn_pp_snap,
  fn_flowdir,
  fn_catchment_ras,
  fn_catchment_vec,
  return_vector = TRUE,
  ...
)
}
\arguments{
\item{fn_pp_snap}{}

\item{fn_flowdir}{}

\item{fn_catchment_ras}{}

\item{fn_catchment_vec}{}

\item{return_vector}{If \code{TRUE} (the default) a vector of the catchment will be returned.}

\item{...}{}
}
\value{

}
\description{
Delineate catchment boundaries
}
\seealso{
\code{\link{ch_wbt_filenames}} \code{\link{ch_wbt_catchment_onestep}}
}
\author{
Dan Moore
}
