% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_wbt_catchment.R
\name{ch_wbt_catchment}
\alias{ch_wbt_catchment}
\title{Delineate catchment boundaries}
\usage{
ch_wbt_catchment(
  fn_pp_snap,
  fn_flowdir,
  fn_catchment_ras,
  fn_catchment_vec,
  return_vector = TRUE
)
}
\arguments{
\item{fn_pp_snap}{Name of file containing snapped pour points}

\item{fn_flowdir}{Name of file containing flow accumulations.}

\item{fn_catchment_ras}{Raster file to contain delineated catchment.}

\item{fn_catchment_vec}{Vector file to contain delineated catchment.}

\item{return_vector}{If \code{TRUE} (the default) a vector of the catchment will be returned.}
}
\value{
If \code{return_vector == TRUE} a vector of the catchment is returned. Otherwise
nothing is returned.
}
\description{
Delineate catchment boundaries
}
\seealso{
\code{\link{ch_wbt_catchment_onestep}}
}
\author{
Dan Moore and Kevin Shook
}
