% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_wbt_channels.R
\name{ch_wbt_channels}
\alias{ch_wbt_channels}
\title{Generate stream network}
\usage{
ch_wbt_channels(
  fn_flowacc,
  fn_flowdir,
  fn_channel_ras,
  fn_channel_vec,
  threshold = NULL,
  ...
)
}
\arguments{
\item{fn_flowacc}{}

\item{fn_flowdir}{}

\item{fn_channel_ras}{}

\item{fn_channel_vec}{}

\item{threshold}{}

\item{...}{}
}
\value{

}
\description{
Generate stream network
}
\author{
Dan Moore
}
