% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_wbt_channels.R
\name{ch_wbt_channels}
\alias{ch_wbt_channels}
\title{Generate stream network}
\usage{
ch_wbt_channels(
  fn_flowacc,
  fn_flowdir,
  fn_channel_ras,
  fn_channel_vec,
  threshold = NULL,
  ...
)
}
\arguments{
\item{fn_flowacc}{File name for flow accumulation grid.}

\item{fn_flowdir}{File name for flow direction grid.}

\item{fn_channel_ras}{File name for raster version of channel network.}

\item{fn_channel_vec}{File name for vector version of channel networks.}

\item{threshold}{Threshold for channel initiation.}

\item{...}{Other parameters for \pkg{whitebox} function \code{wbt_extract_streams}}
}
\value{
Returns a \pkg{sf} vector object of the stream channels.
}
\description{
Generate stream network
}
\author{
Dan Moore
}
