% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_wbt_flow_accumulation.R
\name{ch_wbt_flow_accumulation}
\alias{ch_wbt_flow_accumulation}
\title{Creates flow accumulation grid file}
\usage{
ch_wbt_flow_accumulation(fn_dem_ns, fn_flowacc, return_raster = TRUE)
}
\arguments{
\item{fn_dem_ns}{File name of dem with sinks removed.}

\item{fn_flowacc}{File name for flow accumulation grid to be created.}

\item{return_raster}{If \code{TRUE} (the default), the flow accumulation
grid will be returned as a raster object, in addition to being written to
\option{fn_flowacc}. If \code{FALSE}, the output file will still be created
but a \code{NULL} value is returned.}
}
\value{

}
\description{
Creates flow accumulation grid file
}
\author{
Dan Moore
}
