% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_wbt_flow_direction.R
\name{ch_wbt_flow_direction}
\alias{ch_wbt_flow_direction}
\title{Creates flow direction grid file}
\usage{
ch_wbt_flow_direction(fn_dem_ns, fn_flowdir, return_raster = TRUE)
}
\arguments{
\item{fn_dem_ns}{File name of dem with sinks removed.}

\item{fn_flowdir}{File name for flow direction grid to be created.}

\item{return_raster}{If \code{TRUE} (the default), the flow direction
grid will be returned as a raster object, in addition to being written to
\option{fn_flowdir}. If \code{FALSE}, the output file will still be created
but a \code{NULL} value is returned.}
}
\value{

}
\description{
Creates flow direction grid file
}
\author{
Dan Moore
}
