% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_wbt_pourpoints.R
\name{ch_wbt_pourpoints}
\alias{ch_wbt_pourpoints}
\title{Snap pour points to channels}
\usage{
ch_wbt_pourpoints(
  pp_sf = NULL,
  fn_flowacc,
  fn_pp,
  fn_pp_snap,
  check_crs = TRUE,
  snap_dist = NULL,
  ...
)
}
\arguments{
\item{pp_sf}{}

\item{fn_flowacc}{Name of file containing flow accumulations.}

\item{fn_pp}{}

\item{fn_pp_snap}{}

\item{check_crs}{If \code{TRUE} the projections of the pour points and flow
accumulation files will be checked to ensure they are identical.}

\item{snap_dist}{}

\item{...}{}
}
\value{

}
\description{
Snap pour points to channels
}
\author{
Dan Moore
}
