% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_wbt_pourpoints.R
\name{ch_wbt_pourpoints}
\alias{ch_wbt_pourpoints}
\title{Snap pour points to channels}
\usage{
ch_wbt_pourpoints(
  pp_sf = NULL,
  fn_flowacc,
  fn_pp,
  fn_pp_snap,
  check_crs = TRUE,
  snap_dist = NULL,
  ...
)
}
\arguments{
\item{pp_sf}{\pkg{sf} object containing pour points. These must be supplied by the user. See
the code in \code{\link{ch_volcano_pourpoints}} for an example of creating the object.}

\item{fn_flowacc}{Name of file containing flow accumulations.}

\item{fn_pp}{File name to create un-snapped pour points.}

\item{fn_pp_snap}{File name for snapped pour points.}

\item{check_crs}{If \code{TRUE} the projections of the pour points and flow
accumulation files will be checked to ensure they are identical.}

\item{snap_dist}{Maximum snap distance in map units.}

\item{...}{Additional parameters for \pkg{whitebox} function \code{wbt_snap_pour_points}.}
}
\value{
Returns a \pkg{sf} object of the specified pour points snapped to the
channel network.
}
\description{
Pour points describe the outlets of sub-basins within a DEM. To use
the pour points to delineate catchments, they must align with the drainage
network. This function snaps (forces the locations) of pour points to the
channels.
}
\seealso{
\code{\link{ch_volcano_pourpoints}}
}
\author{
Dan Moore
}
