% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_wtr-yr.R
\name{ch_wtr_yr}
\alias{ch_wtr_yr}
\title{Designation of the water year}
\source{
http://stackoverflow.com/questions/27626533/r-create-function-to-add-water-year-column
}
\usage{
ch_wtr_yr(dates, start_month = 10)
}
\arguments{
\item{dates}{A vector of dates with actual year}

\item{start_month}{Month in which the year starts (defaults to October)}
}
\value{
Year starting in start_month
}
\description{
Display water year
}
\examples{
date <- seq(as.Date("1910/1/1"), as.Date("1912/1/1"), "days")
wtr_yr_date <- ch_wtr_yr(dates=date, start_month=10)
data.frame(wtr_yr_date, date)
}
