% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doys.R
\name{doys}
\alias{doys}
\title{Multiple function Day of Year, Water Year Function}
\usage{
doys(Date)
}
\arguments{
\item{{Date}}{{an array as.Date}}
}
\value{
a dataframe with different format dates

\code{Date}  in Date format

\code{year}  numeric calendar year

\code{doy}   numeric day of year

\code{wyear} numeric day of water year starting on 1 October

\code{dwy}   numeric day of water year

 Conversion of a data array to a dataframe of different forms date,doy, dowy, Year
 Calculates day of year and day of water year since 10.01 of present water year.
 uses base::isodate
}
\description{
Multiple function Day of Year, Water Year Function
}
\examples{
\dontrun{my_dates <- doys(mdata$Date)}



}
\author{
Paul Whitfield <paul.h.whitfield@gmail.com>
}
