% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_fgpa.R
\name{fgpa}
\alias{fgpa}
\alias{fgpa1d}
\alias{fgpa2d}
\alias{fgpaLmom}
\title{Estimation of the Generalized Pareto distribution.}
\usage{
fgpa1d(x, sol = FALSE)

fgpa2d(x, sol = FALSE, par0 = NULL, ...)

fgpaLmom(x)
}
\arguments{
\item{x}{Sample.}

\item{sol}{Does solution from \code{optim} be returned.
In case of \code{fgpa1d}, it returns the variance covariance matrix.}

\item{par0}{Initial parameter.}

\item{...}{aditional arguments to pass to \code{\link{optim}}}
}
\description{
Low level functions for estimating of the generalized Pareto
distribution(GPA) with two parameters. Can use
either maximum likelihood or the method of L-moments.
The algorithm of \code{fgpa2d} is using
\code{optim} to directly optimize the log-likelihood (bivariate), while
the algorithm of \code{fgpa1d} is using a transformation to use
a univariate optimization routine. Additionally, \code{fgpa2d} constraint
the shape parameter between -.5 and 1.
}
\section{Reference}{


Davison AC, Smith RL. (1990) Models for Exceedances over High Thresholds.
  Journal of the Royal Statistical Society Series B (Methodological).
  52(3):393-442.

Hosking JRM (1990). L-Moments: Analysis and Estimation of Distributions Using
  Linear Combinations of Order Statistics. Journal of the Royal Statistical
  Society Series B (Methodological). 52(1):105-24.
}

\examples{

x <- rgpa(1000, 1, -.2)
fgpa1d(x)
fgpa2d(x)
fgpaLmom(x)

}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
