% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_FitNsAmax.R, R/rfa_predict_nsamax.R
\name{FitNsAmax}
\alias{FitNsAmax}
\alias{AIC.nsamax}
\alias{simulate.nsamax}
\alias{predict.nsamax}
\alias{BootNsAmax}
\alias{summary.bootns}
\title{Nonstationary model for annual maxima fitted by least-squares and L-moments}
\usage{
FitNsAmax(form, x, distr = "gev", type = "add", k = 2, tol.gev = 0)

\method{AIC}{nsamax}(object, k = 2)

\method{simulate}{nsamax}(object, u = NULL)

\method{predict}{nsamax}(object, p = c(0.5, 0.8, 0.9, 0.95, 0.98, 0.99),
  newdata = NULL, reliability = FALSE)

BootNsAmax(object, p = c(0.5, 0.8, 0.9, 0.95, 0.98, 0.99),
  newdata = NULL, reliability = FALSE, nsim = 1000, alpha = 0.95,
  verbose = TRUE)

\method{summary}{bootns}(object, variable = "para", alpha = 0.05)
}
\arguments{
\item{form}{Formula of the trend.}

\item{x}{Data.}

\item{distr}{Distribution}

\item{type}{Type of model. Either additive (\code{"add"}) 
or multiplicative (\code{"mult"}).}
}
\description{
Return the estimation of a nonstationary model for the distribution of annual
maximums with a trend in the mean.
}
\examples{

data(flowStJohn)

amax <- ExtractAmax(flow~date, flowStJohn, tol = 365)

fit <- FitNsAmax(flow~date, amax, 'gev', type = 'mult')

print(fit)
plot(fitted(fit), residuals(fit))

plot(flow~date, amax)
lines(amax$date, fitted(fit))

fit.ln3 <- FitNsAmax(flow~date, amax, 'ln3', type = 'mult')

AIC(fit)
AIC(fit.ln3)

FitNsAmax(flow~date, amax, c('gev','ln3'), type = 'mult')


}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
