% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_FitNsPot.R
\name{FitNsPot}
\alias{FitNsPot}
\title{Estimation of a nonstationary index-flood model for peaks over threshold for
a single site.}
\usage{
FitNsPot(form, x, tau, method = "reg-lmom", thresh = ~1, trend = ~1,
  trend.link = "identity", trend.start = NULL, trend.method = "BFGS",
  trend.control = list(), declust = "none", r = 1, rlow = 0.75,
  unit = 365.25, sorted = FALSE, start = NULL, quiet = FALSE, ...)
}
\arguments{
\item{form}{Formula describing the date and the response variable. 
Must be of the form \code{y ~ date}.}

\item{x}{Data.}

\item{tau}{Percentage of daily data above the threshold.}

\item{thresh}{Formula determining the threshold model. 
Must be of the form \code{~ x1 + x2 + ...}.}

\item{trend}{Formula determining the trend of the mean excess. 
Same as \code{thresh}.}

\item{trend.link}{Link function characterizing the mean excess.}

\item{trend.start}{If necessary, starting values for the estimation of 
the mean excess.}

\item{declust}{Type of declustering techniques used to extract peaks over 
threshold. See \link{which.floodpeaks}}

\item{r, rlow}{Parameter of the declustering technique.}
}
\description{
Estimate the parameter of a nonstationary index-flood model by a stepwise
method. First quantile regression is used to estimate 
a time-varying threshold. Second, a trend in the exceedances is determined
using quasi-likelihood and finally, a L-moments are used for estimating 
the shape parameter of the a Generalized Pareto Distribution.
}
\examples{

fit <- FitNsPot(flow~date, x = flowStJohn, tau = .95,
                trend = ~ date, thresh = ~ date, declust = 'wrc', r = 14)

print(fit)
plot(fit)

GofTest(fit)

}
\seealso{
\link{FitPot}, \link{SearchThresh}.
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
