% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_coef_nsamax.R, R/rfa_fitted_nsamax.R
\name{coef.nsamax}
\alias{coef.nsamax}
\alias{fitted.nsamax}
\title{Fitted trend for the nonstationary AMAX model}
\usage{
\method{coef}{nsamax}(object, ...)

\method{fitted}{nsamax}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{Output from \link{FitNsAmax}.}

\item{...}{Other parameters.}

\item{newdata}{Covariates at the fitted dates.}
}
\description{
Return the fitted trend of the nonstationary AMAX model at given dates.
}
\seealso{
\link{FitNsAmax}, \link{predict.nsamax}.
}
