% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_gumbel.R
\name{scale_x_gumbel_rp}
\alias{scale_x_gumbel_rp}
\alias{scale_y_gumbel_rp}
\alias{scale_x_gumbel_aep}
\alias{scale_y_gumbel_aep}
\title{Gumbel-transformed axes}
\usage{
scale_x_gumbel_rp(...)

scale_y_gumbel_rp(...)

scale_x_gumbel_aep(...)

scale_y_gumbel_aep(...)
}
\arguments{
\item{...}{Arguments to pass to \code{scale_x_continuous} or
\code{scale_y_continuous} from \pkg{ggplot2}.}
}
\value{
The same output as \code{scale_x_continuous} or
\code{scale_y_continuous}, but with the appropriate Gumbel spacing.
}
\description{
Transforms return period or annual exceedance probability to be spaced
according to a reduced Gumbel variate. Can be added as a scale layer to
a ggplot2 graphic.
}
\examples{
library(ggplot2)

# A flood frequency plot: annual maximum flows against the probability of
# being exceeded in any given year, on Gumbel spacing.
ams <- ch_rfa_extractamax(Flow ~ Date, CAN05AA008, tol = 350)
ams <- ams[order(ams$Flow), ]
# Weibull plotting positions
ams$aep <- 1 - seq_len(nrow(ams)) / (nrow(ams) + 1)

ggplot(ams, aes(aep, Flow)) +
  geom_point() +
  scale_x_gumbel_aep("Annual exceedance probability") +
  ylab(expression(paste("Annual maximum daily discharge (", m^3, "/s)")))

# The same data, read as return periods instead.
ggplot(ams, aes(1 / aep, Flow)) +
  geom_point() +
  scale_x_gumbel_rp("Return period (years)")

# The y variants behave the same way.
ggplot(ams, aes(Flow, 1 / aep)) +
  geom_point() +
  scale_y_gumbel_rp("Return period (years)")
}
