% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_PadPot.R
\name{PadPot}
\alias{PadPot}
\alias{PadPot.default}
\alias{PadPot.formula}
\alias{PadPot.data.frame}
\title{Return a padded time series for peaks over threshold model.}
\usage{
PadPot(x, ...)

\method{PadPot}{default}(x, ...)

\method{PadPot}{formula}(form, x, ...)

\method{PadPot}{data.frame}(x, z = 0, tol = 346, ...)
}
\arguments{
\item{x}{Data}

\item{...}{Other parameters.}

\item{form}{Formula}

\item{z}{Imputed value.}

\item{tol}{Number of days to consider a year complete.}
}
\description{
Return a padded time series where incomplete years are removed and missing days of complete
years are imputed.
}
\details{
If there is no formula, \code{x} must be a data frame with either 
2 (date, value) or 3 (site, date, value) columns.
Similarly, the formula must be of the form \code{value ~ date} or 
\code{value ~ site + date}.
}
\examples{

## Will remove 1926 that is incomplete
xd <- PadPot(flow ~ date, flowStJohn, tol = 365)
head(xd)

## Will pad 1926 that is "complete"
xd <- PadPot(flow ~ date, flowStJohn, tol = 1)
head(xd)


}
\seealso{
\link{FitPot}.
}
