% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_plot_amax.R
\name{plot.amax}
\alias{plot.amax}
\title{Return level plot for annual maxima}
\usage{
\method{plot}{amax}(
  x,
  main = "Return level plot",
  xlab = "Return period (year)",
  ylab = "Flood quantiles",
  ci = FALSE,
  col.ci = 2,
  lty.ci = 2,
  lwd.ci = 1,
  ...
)
}
\arguments{
\item{x}{Output from \link{FitAmax}.}

\item{main, ylab, xlab}{Graphical parameters. See \code{\link{par}}.}

\item{ci}{Logical. Should confidence intervals be displayed. 
See \link{predict.amax} with argument Delta method.}

\item{col.ci, lty.ci, lwd.ci}{Graphical parameters determining the 
confidence intervals.}

\item{...}{Other graphical parameters. See \code{\link{par}}.}
}
\description{
Create a plot of the estimated versus sample flood quantiles, where the 
x-axis is expressed in terms of return period.
}
\examples{

data(flowStJohn)

x <- ExtractAmax(flow~date, flowStJohn)$flow

fit <- FitAmax(x, distr = 'gev', method = 'mle')

plot(fit, ci = TRUE)
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
