% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_rfa_plot_isite.R
\name{plot.isite}
\alias{plot.isite}
\title{Plot of the pairwise intersite correlation.}
\usage{
\method{plot}{isite}(
  x,
  xmat,
  distance,
  xlab = NULL,
  ylab = NULL,
  pch = "+",
  col = "grey",
  bin.pch = 16,
  bin.cex = 1.5,
  bin.col = "black",
  ...
)
}
\arguments{
\item{x}{Output from \code{Intersite}.}

\item{xmat}{Dataset in wide format.}

\item{distance}{Matrix of distances.}

\item{xlab, ylab}{Label of the xy-axis.}

\item{col, pch}{Graphical parameters for the individual points.}

\item{bin.pch, bin.col, bin.cex}{Graphical parameters for the bins.}

\item{...}{Other graphical parameters passed to \link{plot}.}
}
\description{
Produce a graphic of the pairwise intersite correlation coefficient with 
respect to the distance.
}
\examples{

## Organize annual maximums
ams <- flowAtlantic$ams
ams$year <- format(ams$date, '\%Y')
xmat <- DataWide(ams ~ id + year, ams)

## estimate empirical correlation
isite <- Intersite(xmat, type = 'emp')
print(isite)
plot(isite)

## Compute distance between sites
coord <- flowAtlantic$info[, c('lon','lat')]
rownames(coord) <- flowAtlantic$info$id
h <- GeoDist(coord)

## make sure that columns match
h <- h[colnames(xmat), colnames(xmat)]

## estimate intersite correlation using a model
isite <- Intersite(xmat, distance = h, type = 'exp')
print(isite)
plot(isite, xmat, h)

}
\seealso{
\link{Intersite}, \link{DataWide}.
}
