% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_rfa_PlotMrl.R
\name{PlotMrl}
\alias{PlotMrl}
\title{Mean residual life plot}
\usage{
PlotMrl(
  form,
  x,
  u,
  declust = NULL,
  r = 1,
  rlow = 0.75,
  alpha = 0.05,
  ylab = "Mean Residual Life",
  xlab = "Threshold",
  col = "black",
  lty = 1,
  lwd = 1,
  col.ci = "black",
  lty.ci = 3,
  lwd.ci = 1,
  ylim = NULL,
  display = TRUE,
  ...
)
}
\arguments{
\item{form, x}{Formula and sample. The must be of the form.}

\item{u}{Series of candidate thresholds.}

\item{declust, r, rlow}{Parameter for declustering. See \link{FitPot}.}

\item{alpha}{Confidence interal with level \code{1-alpha/2}.}

\item{col.ci, lty.ci, lwd.ci}{Parameter for controling the confidence intervals.}

\item{ylim, ylab, xlab, col, lty, lwd}{Parameters for controling the graphic.}

\item{display}{Logical. Should the graph be display.}

\item{...}{Other arameters for controling the graphic.}
}
\description{
Produce a mean residual Life plot to help selecting to help
selecting a proper threshold.
}
\examples{

## Find list of candidate thresholds
lstu <- seq(500,2500, len = 50)

PlotMrl(flow~date, flowStJohn, u = lstu, declust = 'wrc', r = 14)

x <- PlotMrl(flow~date, flowStJohn, u = lstu, declust = 'wrc', r = 14)

head(x)

}
\seealso{
\link{FitPot}, \link{which.floodPeaks}.
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
