% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_predict_nspot.R
\name{predict.nspot}
\alias{predict.nspot}
\alias{BootNsPot}
\title{Flood quantiles for the nonstationary index-flood model at one site}
\usage{
\method{predict}{nspot}(object, rt = c(2, 5, 10, 20, 50, 100),
  newdata = NULL, reliability = FALSE, reg.kappa = NULL)

BootNsPot(object, rt, newdata = NULL, reliability = FALSE,
  nsim = 1000, verbose = TRUE, by.year = FALSE)
}
\arguments{
\item{object}{Output form \cite{FitNsPot}}

\item{rt}{Vector of return period to estimate.}

\item{newdata}{Data when to estimate flood quantiles.}

\item{reliability}{Logical. Should the reliability be returned instead of the
flood quantiles.}

\item{nsim}{Size of the bootstrap sample.}
}
\description{
Predict the return levels or design level of
}
\examples{

data(flowStJohn)
 
## Fit the data using the ns-index-flood model 
fit <- FitNsPot(flow~date, x = flowStJohn, tau = .95,
                trend = ~ date, thresh = ~ date, declust = 'wrc', r = 14)

plot(fit)

## Add the return levels to the graphics 
qhat <- predict(fit, rt = c(10, 100))
for(ii in 1:6) lines(flowStJohn$date, qhat[,ii], col = 'magenta')

## Add the design level of the last 30 years

xref <- flowStJohn[flowStJohn$date > as.Date('1980-01-01'), ]
qhat <- predict(fit, rt = c(10, 50), newdata = xref, reliability = TRUE)
for(ii in 1:6) 
  arrows(min(xref$date), qhat[ii], max(xref$date), lwd = 2, col = 'cyan',
         code = 3, angle = 90, length = .05)

## Assuming that a regional estimate of of the shape was available by other 
## methods.        
predict(fit, rt = c(10, 100), reg.kappa = .3)

}
\references{
Durocher, M., Burn, D. H., & Ashkar, F. (2019). Comparison of estimation 
  methods for a nonstationary index-flood model in flood frequency analysis 
  using peaks over threshold [Preprint]. 
  https://doi.org/10.31223/osf.io/rnepc
}
\seealso{
\link{predict.nspot}, \link{FitNsAmax}.
}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
