% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa_simulate_poolmle.R
\name{simulate.poolmle}
\alias{simulate.poolmle}
\title{Simulation of a pooling group models}
\usage{
\method{simulate}{poolmle}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{Output from \link{FitPoolMle}.}

\item{nsim}{Number of simulations.}

\item{seed}{Seed for random number generator.}

\item{...}{Other parameters.}
}
\description{
Return the simulation of multiple sites following a regional model.
}
\seealso{
\link{FitPoolMle}.
}
