% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistical_hydrology_functions.R
\name{StatisticalHydrology-functions}
\alias{StatisticalHydrology-functions}
\title{Statistical analysis functions}
\description{
These functions perform statistical analyses
\describe{
 \item{ch_binned_MannWhitney}{ Compares two time periods of data using Mann-Whitney test}
 \item{ch_fdcurve}{Finds flow exceedence probabilities}
 \item{ch_get_peaks}{Finds peak flows over a specified threshold}
}
}
