% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/W05AA008.R
\docType{data}
\name{W05AA008}
\alias{W05AA008}
\title{W05AA008}
\format{
A dateframe with 25252 rows and 5 columns spanning the period 1910-2013.
}
\source{
Water Survey of Canada

Variables: 
\itemize{
\item{ID} {StationID}
\item{PARAM}{Parameter 1=Flow, 2=Level}
\item{Date}{\R date}
\item{Flow}{Daily flow in  m\eqn{^3}{^3}/s}
\item{SYM}{Water Survey FLags A, B, D, E }
}
}
\usage{
W05AA008
}
\description{
A dataframe of Water Survey of Canada (WSC) daily flows for station W05AA008, CROWSNEST RIVER AT FRANK Alberta.
}
\keyword{datasets}
